Get the metadata and tracklist for an album.

Get the metadata and tracklist for an album on Last.fm using the album name or a musicbrainz id. Implementation of last.fm's album.getInfo API method (http://www.last.fm/api/show/album.getInfo)

album_getInfo(artist, album, mbid = NA, autocorrect = NA, username = NA,
  lang = NA)

Arguments

artist
The artist name.

album
The album name.

mbid
The musicbrainz id for the album.

autocorrect
Transform misspelled artist names into correct artist names, returning the correct version instead. The corrected artist name will be returned in the response. [0|1]

username
The username for the context of the request. If supplied, the user's playcount for this album is included in the response.

lang
The language to return the biography in, expressed as an ISO 639 alpha-2 code.

Value

A list of the metadata and tracklist for an album.

Examples

album_getInfo("Father John Misty", "Fear Fun")

#>       name            artist                                 mbid
#> 1 Fear Fun Father John Misty da07a173-bf3d-4e7e-995d-ce473ce83030
#>                                                    url
#> 1 https://www.last.fm/music/Father+John+Misty/Fear+Fun
#>                                                                     image.extralarge
#> 1 https://lastfm-img2.akamaized.net/i/u/300x300/24960601f04b4317b4d2c5432e77e1c9.png
#>                                                                       image.large
#> 1 https://lastfm-img2.akamaized.net/i/u/174s/24960601f04b4317b4d2c5432e77e1c9.png
#>                                                                     image.medium
#> 1 https://lastfm-img2.akamaized.net/i/u/64s/24960601f04b4317b4d2c5432e77e1c9.png
#>                                                                      image.small
#> 1 https://lastfm-img2.akamaized.net/i/u/34s/24960601f04b4317b4d2c5432e77e1c9.png
#>   listeners playcount
#> 1    249377   4985944
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        tracks
#> 1 Funtimes in Babylon, Nancy From Now On, Hollywood Forever Cemetery Sings, I'm Writing a Novel, O I Long to Feel Your Arms Around Me, Misty's Nightmares 1 & 2, Only Son of the Ladiesman, This Is Sally Hatchet, Well, You Can Do It Without Me, Now I'm Learning to Love the War, Tee Pees 1-12, Everyman Needs a Companion, https://www.last.fm/music/Father+John+Misty/_/Funtimes+in+Babylon, https://www.last.fm/music/Father+John+Misty/_/Nancy+From+Now+On, https://www.last.fm/music/Father+John+Misty/_/Hollywood+Forever+Cemetery+Sings, https://www.last.fm/music/Father+John+Misty/_/I%27m+Writing+a+Novel, https://www.last.fm/music/Father+John+Misty/_/O+I+Long+to+Feel+Your+Arms+Around+Me, https://www.last.fm/music/Father+John+Misty/_/Misty%27s+Nightmares+1+&+2, https://www.last.fm/music/Father+John+Misty/_/Only+Son+of+the+Ladiesman, https://www.last.fm/music/Father+John+Misty/_/This+Is+Sally+Hatchet, https://www.last.fm/music/Father+John+Misty/_/Well,+You+Can+Do+It+Without+Me, https://www.last.fm/music/Father+John+Misty/_/Now+I%27m+Learning+to+Love+the+War, https://www.last.fm/music/Father+John+Misty/_/Tee+Pees+1-12, https://www.last.fm/music/Father+John+Misty/_/Everyman+Needs+a+Companion, 220, 234, 191, 215, 143, 193, 249, 238, 163, 262, 196, 320, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, Father John Misty, Father John Misty, Father John Misty, Father John Misty, Father John Misty, Father John Misty, Father John Misty, Father John Misty, Father John Misty, Father John Misty, Father John Misty, Father John Misty, 172053e5-a7fb-4548-a64d-c53858b5e775, 172053e5-a7fb-4548-a64d-c53858b5e775, 172053e5-a7fb-4548-a64d-c53858b5e775, 172053e5-a7fb-4548-a64d-c53858b5e775, 172053e5-a7fb-4548-a64d-c53858b5e775, 172053e5-a7fb-4548-a64d-c53858b5e775, 172053e5-a7fb-4548-a64d-c53858b5e775, 172053e5-a7fb-4548-a64d-c53858b5e775, 172053e5-a7fb-4548-a64d-c53858b5e775, 172053e5-a7fb-4548-a64d-c53858b5e775, 172053e5-a7fb-4548-a64d-c53858b5e775, 172053e5-a7fb-4548-a64d-c53858b5e775, https://www.last.fm/music/Father+John+Misty, https://www.last.fm/music/Father+John+Misty, https://www.last.fm/music/Father+John+Misty, https://www.last.fm/music/Father+John+Misty, https://www.last.fm/music/Father+John+Misty, https://www.last.fm/music/Father+John+Misty, https://www.last.fm/music/Father+John+Misty, https://www.last.fm/music/Father+John+Misty, https://www.last.fm/music/Father+John+Misty, https://www.last.fm/music/Father+John+Misty, https://www.last.fm/music/Father+John+Misty, https://www.last.fm/music/Father+John+Misty
#>                                                                                                                                                                                                                                         tags
#> 1 2012, best of 2012, availableonemusic, rock, alternative, https://www.last.fm/tag/2012, https://www.last.fm/tag/best+of+2012, https://www.last.fm/tag/availableonemusic, https://www.last.fm/tag/rock, https://www.last.fm/tag/alternative