Get the top tags for an album, ordered by popularity.

Get the top tags for an album on Last.fm, ordered by popularity. Implementation of last.fm's album.getTopTags API method (http://www.last.fm/api/show/album.getTopTags)

album_getTopTags(artist, album, mbid = NA, autocorrect = NA,
  api_key = lastkey)

Arguments

artist
The artist name.

album
The album name.

mbid
The musicbrainz id for the album.

autocorrect
Transform misspelled artist names into correct artist names, returning the correct version instead. The corrected artist name will be returned in the response. [0|1]

api_key
A Last.fm API key.

Value

A list of the top tags for an album.

Examples

album_getTopTags("Miles Davis", "Kind of Blue")

#> $tag
#>    count                                     name
#> 1    100                                     jazz
#> 2     31                                cool jazz
#> 3     24                             albums I own
#> 4     19                          favorite albums
#> 5     18                               modal jazz
#> 6     15                              miles davis
#> 7     12                                  trumpet
#> 8     11                                     1959
#> 9      9                                      50s
#> 10     6                             instrumental
#> 11     5 1001 Albums You Must Hear Before You Die
#> 12     5                                    bebop
#> 13     5                             kind of blue
#> 14     4                         favourite albums
#> 15     4                           Perfect Albums
#> 16     3                                 chillout
#> 17     3                                    blues
#> 18     3                                     cool
#> 19     3                                saxophone
#> 20     3                            john coltrane
#> 21     3                          Officially Shit
#> 22     3                             great albums
#> 23     2                                    chill
#> 24     2                                 Post-Bop
#> 25     2                                    Miles
#> 26     2                                   oldies
#> 27     2                      cannonball adderley
#> 28     2                             classic jazz
#> 29     2                               Bill Evans
#> 30     2                                     bass
#> 31     2                                    Modal
#> 32     2                              Masterpiece
#> 33     2                            paul chambers
#> 34     2                               fav albums
#> 35     2                              jazz albums
#> 36     2                             own on vinyl
#> 37     1                                  classic
#> 38     1                                   Fusion
#> 39     1                                 Hard Bop
#> 40     1                              jazz fusion
#> 41     1                                   genius
#> 42     1                                 relaxing
#> 43     1                                    1950s
#> 44     1                           classic albums
#> 45     1                                     blue
#> 46     1                               jimmy cobb
#> 47     1                             wynton kelly
#>                                                                 url
#> 1                                      https://www.last.fm/tag/jazz
#> 2                                 https://www.last.fm/tag/cool+jazz
#> 3                              https://www.last.fm/tag/albums+I+own
#> 4                           https://www.last.fm/tag/favorite+albums
#> 5                                https://www.last.fm/tag/modal+jazz
#> 6                               https://www.last.fm/tag/miles+davis
#> 7                                   https://www.last.fm/tag/trumpet
#> 8                                      https://www.last.fm/tag/1959
#> 9                                       https://www.last.fm/tag/50s
#> 10                             https://www.last.fm/tag/instrumental
#> 11 https://www.last.fm/tag/1001+Albums+You+Must+Hear+Before+You+Die
#> 12                                    https://www.last.fm/tag/bebop
#> 13                             https://www.last.fm/tag/kind+of+blue
#> 14                         https://www.last.fm/tag/favourite+albums
#> 15                           https://www.last.fm/tag/Perfect+Albums
#> 16                                 https://www.last.fm/tag/chillout
#> 17                                    https://www.last.fm/tag/blues
#> 18                                     https://www.last.fm/tag/cool
#> 19                                https://www.last.fm/tag/saxophone
#> 20                            https://www.last.fm/tag/john+coltrane
#> 21                          https://www.last.fm/tag/Officially+Shit
#> 22                             https://www.last.fm/tag/great+albums
#> 23                                    https://www.last.fm/tag/chill
#> 24                                 https://www.last.fm/tag/Post-Bop
#> 25                                    https://www.last.fm/tag/Miles
#> 26                                   https://www.last.fm/tag/oldies
#> 27                      https://www.last.fm/tag/cannonball+adderley
#> 28                             https://www.last.fm/tag/classic+jazz
#> 29                               https://www.last.fm/tag/Bill+Evans
#> 30                                     https://www.last.fm/tag/bass
#> 31                                    https://www.last.fm/tag/Modal
#> 32                              https://www.last.fm/tag/Masterpiece
#> 33                            https://www.last.fm/tag/paul+chambers
#> 34                               https://www.last.fm/tag/fav+albums
#> 35                              https://www.last.fm/tag/jazz+albums
#> 36                             https://www.last.fm/tag/own+on+vinyl
#> 37                                  https://www.last.fm/tag/classic
#> 38                                   https://www.last.fm/tag/Fusion
#> 39                                 https://www.last.fm/tag/Hard+Bop
#> 40                              https://www.last.fm/tag/jazz+fusion
#> 41                                   https://www.last.fm/tag/genius
#> 42                                 https://www.last.fm/tag/relaxing
#> 43                                    https://www.last.fm/tag/1950s
#> 44                           https://www.last.fm/tag/classic+albums
#> 45                                     https://www.last.fm/tag/blue
#> 46                               https://www.last.fm/tag/jimmy+cobb
#> 47                             https://www.last.fm/tag/wynton+kelly
#> 
#> $`@attr`
#> $`@attr`$artist
#> [1] "Miles Davis"
#> 
#> $`@attr`$album
#> [1] "Kind of Blue"
#> 
#>