Get info of an artist.

Get the metadata for an artist. Includes biography, truncated at 300 characters. Implementation of last.fm's artist.getInfo API method (http://www.last.fm/api/show/artist.getInfo)

artist_getInfo(artist, mbid = NA, lang = NA, autocorrect = NA,
  username = NA)

Arguments

artist
The artist name.

mbid
The musicbrainz id for the artist

lang
The language to return the biography in, expressed as an ISO 639 alpha-2 code.

autocorrect
Transform misspelled artist names into correct artist names, returning the correct version instead. The corrected artist name will be returned in the response. [0|1]

username
The username for the context of the request. If supplied, the user's playcount for this artist is included in the response.

Value

A list of artist info.

Examples

artist_getInfo("Wye Oak")

#>      name                                 mbid
#> 1 Wye Oak 3c8ff4c9-b5b9-4e78-ad9d-5157c323de24
#>                                 url
#> 1 https://www.last.fm/music/Wye+Oak
#>                                                                     image.extralarge
#> 1 https://lastfm-img2.akamaized.net/i/u/300x300/16c41ef2e1e32be54c3b6480f19effa3.png
#>                                                                       image.large
#> 1 https://lastfm-img2.akamaized.net/i/u/174s/16c41ef2e1e32be54c3b6480f19effa3.png
#>                                                                     image.medium
#> 1 https://lastfm-img2.akamaized.net/i/u/64s/16c41ef2e1e32be54c3b6480f19effa3.png
#>                                                                      image.small
#> 1 https://lastfm-img2.akamaized.net/i/u/34s/16c41ef2e1e32be54c3b6480f19effa3.png
#>   streamable ontour stats.listeners stats.playcount
#> 1          0      1          308999         6755745
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               similar
#> 1 Flock of Dimes, Lucy Dacus, Hop Along, Lower Dens, Eleanor Friedberger, https://www.last.fm/music/Flock+of+Dimes, https://www.last.fm/music/Lucy+Dacus, https://www.last.fm/music/Hop+Along, https://www.last.fm/music/Lower+Dens, https://www.last.fm/music/Eleanor+Friedberger, https://lastfm-img2.akamaized.net/i/u/34s/a93f9cf0dba3e8eaf6d4afa5e5e7ac72.png, https://lastfm-img2.akamaized.net/i/u/64s/a93f9cf0dba3e8eaf6d4afa5e5e7ac72.png, https://lastfm-img2.akamaized.net/i/u/174s/a93f9cf0dba3e8eaf6d4afa5e5e7ac72.png, https://lastfm-img2.akamaized.net/i/u/300x300/a93f9cf0dba3e8eaf6d4afa5e5e7ac72.png, https://lastfm-img2.akamaized.net/i/u/300x300/a93f9cf0dba3e8eaf6d4afa5e5e7ac72.png, https://lastfm-img2.akamaized.net/i/u/300x300/a93f9cf0dba3e8eaf6d4afa5e5e7ac72.png, small, medium, large, extralarge, mega, , https://lastfm-img2.akamaized.net/i/u/34s/9345b20751e642b183429f7b8509ffbc.png, https://lastfm-img2.akamaized.net/i/u/64s/9345b20751e642b183429f7b8509ffbc.png, https://lastfm-img2.akamaized.net/i/u/174s/9345b20751e642b183429f7b8509ffbc.png, https://lastfm-img2.akamaized.net/i/u/300x300/9345b20751e642b183429f7b8509ffbc.png, https://lastfm-img2.akamaized.net/i/u/300x300/9345b20751e642b183429f7b8509ffbc.png, https://lastfm-img2.akamaized.net/i/u/300x300/9345b20751e642b183429f7b8509ffbc.png, small, medium, large, extralarge, mega, , https://lastfm-img2.akamaized.net/i/u/34s/eb660e7014ef2b9a7a0725a6465cd593.png, https://lastfm-img2.akamaized.net/i/u/64s/eb660e7014ef2b9a7a0725a6465cd593.png, https://lastfm-img2.akamaized.net/i/u/174s/eb660e7014ef2b9a7a0725a6465cd593.png, https://lastfm-img2.akamaized.net/i/u/300x300/eb660e7014ef2b9a7a0725a6465cd593.png, https://lastfm-img2.akamaized.net/i/u/300x300/eb660e7014ef2b9a7a0725a6465cd593.png, https://lastfm-img2.akamaized.net/i/u/300x300/eb660e7014ef2b9a7a0725a6465cd593.png, small, medium, large, extralarge, mega, , https://lastfm-img2.akamaized.net/i/u/34s/fc3b90f6ffee4d43a199b68ea937f2f1.png, https://lastfm-img2.akamaized.net/i/u/64s/fc3b90f6ffee4d43a199b68ea937f2f1.png, https://lastfm-img2.akamaized.net/i/u/174s/fc3b90f6ffee4d43a199b68ea937f2f1.png, https://lastfm-img2.akamaized.net/i/u/300x300/fc3b90f6ffee4d43a199b68ea937f2f1.png, https://lastfm-img2.akamaized.net/i/u/300x300/fc3b90f6ffee4d43a199b68ea937f2f1.png, https://lastfm-img2.akamaized.net/i/u/300x300/fc3b90f6ffee4d43a199b68ea937f2f1.png, small, medium, large, extralarge, mega, , https://lastfm-img2.akamaized.net/i/u/34s/eafcdc28ae6b46b8981bf671d2db1e86.png, https://lastfm-img2.akamaized.net/i/u/64s/eafcdc28ae6b46b8981bf671d2db1e86.png, https://lastfm-img2.akamaized.net/i/u/174s/eafcdc28ae6b46b8981bf671d2db1e86.png, https://lastfm-img2.akamaized.net/i/u/300x300/eafcdc28ae6b46b8981bf671d2db1e86.png, https://lastfm-img2.akamaized.net/i/u/300x300/eafcdc28ae6b46b8981bf671d2db1e86.png, https://lastfm-img2.akamaized.net/i/u/300x300/eafcdc28ae6b46b8981bf671d2db1e86.png, small, medium, large, extralarge, mega, 
#>                                                                                                                                                                                                                             tags
#> 1 shoegaze, indie, indie rock, seen live, indie folk, https://www.last.fm/tag/shoegaze, https://www.last.fm/tag/indie, https://www.last.fm/tag/indie+rock, https://www.last.fm/tag/seen+live, https://www.last.fm/tag/indie+folk
#>   bio.links.link..text bio.links.link.rel                 bio.links.link.href
#> 1                                original https://last.fm/music/Wye+Oak/+wiki
#>        bio.published
#> 1 10 Feb 2008, 21:50
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    bio.summary
#> 1 Wye Oak (nee Monarch) is a duo of Jenn Wasner and Andy Stack, formed in mid-2006.\n\nWasner and Stack have been playing music together and apart in Baltimore, MD, and chose their name from the honorary state tree of Maryland. If Children is the culmination of many months of recording, polishing and perfecting to achieve what Any Given Tuesday best described as, “. .. . mingling in a room full of shoegaze, alt-country, folk, and noise, shaking hands, kissing cheeks, and making friends with everyone. <a href="https://www.last.fm/music/Wye+Oak">Read more on Last.fm</a>
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       bio.content
#> 1 Wye Oak (nee Monarch) is a duo of Jenn Wasner and Andy Stack, formed in mid-2006.\n\nWasner and Stack have been playing music together and apart in Baltimore, MD, and chose their name from the honorary state tree of Maryland. If Children is the culmination of many months of recording, polishing and perfecting to achieve what Any Given Tuesday best described as, “. .. . mingling in a room full of shoegaze, alt-country, folk, and noise, shaking hands, kissing cheeks, and making friends with everyone.” If Children is an exuberant debut – noisy and restrained in turn, Wye Oak describe their philosophy: "We'd like to think that if a song is good enough, it'll stand on its own regardless of how much noise you pile on top. On the other hand, we do like noise." Listen, and we think you will have to agree!\n\nThe droll tone and fleeting sweetness of a Yo La Tengo are cut into the more deeply textural concerns of shoegaze in the Baltimore group's work. It's an interesting alignment, given the contrast in senses of humor. But Wye Oak successfully shed tonnage from the latter without losing potency. <a href="https://www.last.fm/music/Wye+Oak">Read more on Last.fm</a>. User-contributed text is available under the Creative Commons By-SA License; additional terms may apply.