Get the top tags for an artist.

Get the top tags for an artist on Last.fm, ordered by popularity. Implementation of last.fm's artist.getTopTags API method (http://www.last.fm/api/show/artist.getTopTags)

artist_getTopTags(artist, mbid = NA, autocorrect = NA)

Arguments

artist
The artist name.

mbid
The musicbrainz id for the artist.

autocorrect
Transform misspelled artist names into correct artist names, returning the correct version instead. The corrected artist name will be returned in the response. [0|1]

Value

A list of tags.

Examples

artist_getTopTags("SZA")

#>    count                                name
#> 1    100                                soul
#> 2     88                          electronic
#> 3     79                                 rnb
#> 4     69                     alternative rnb
#> 5     58                    female vocalists
#> 6     43                         alternative
#> 7     39                            Neo-Soul
#> 8     24                            american
#> 9     18                           seen live
#> 10    15                             Hip-Hop
#> 11    15                               indie
#> 12    13                                 r&b
#> 13    13                            Neo Soul
#> 14    13                                 tde
#> 15    13                        glitter trap
#> 16     9                                trap
#> 17     9                               wonky
#> 18     7                         electronica
#> 19     7                                 pop
#> 20     7                           dream pop
#> 21     7                            synthpop
#> 22     7                     female vocalist
#> 23     7                      rhythm & blues
#> 24     7                          future rnb
#> 25     5                            trip-hop
#> 26     5                           downtempo
#> 27     5                   singer-songwriter
#> 28     5                                 10s
#> 29     5                          new jersey
#> 30     5                            Missouri
#> 31     5                            st louis
#> 32     5                               r'n'b
#> 33     5                    contemporary rnb
#> 34     5                    contemporary r&b
#> 35     5                   perfect vocalists
#> 36     3                                 rap
#> 37     3                             ambient
#> 38     3                        experimental
#> 39     3                                 USA
#> 40     3                               urban
#> 41     3                    Psychedelic Soul
#> 42     3                                funk
#> 43     3                 underground hip-hop
#> 44     3                              Dreamy
#> 45     3                            ethereal
#> 46     3                                   f
#> 47     3                            wishlist
#> 48     3                               jazzy
#> 49     3                    rhythm and blues
#> 50     3                           chillwave
#> 51     3                             pop rap
#> 52     3                               2010s
#> 53     3                         songwriting
#> 54     3                         witch house
#> 55     3                             america
#> 56     3                       disappointing
#> 57     3                           cloud rap
#> 58     3                                list
#> 59     3                      kendrick lamar
#> 60     3                             scorpio
#> 61     3                              to try
#> 62     3                        electro soul
#> 63     3                           Maplewood
#> 64     3                            newbreed
#> 65     3                          futuregaze
#> 66     3                          indie soul
#> 67     3                             Popstep
#> 68     3                      rhythm'n'blues
#> 69     3                               again
#> 70     3                            jay rock
#> 71     3                             neo-sul
#> 72     3                         schoolboy q
#> 73     3                             Ab-Soul
#> 74     3                         soultronica
#> 75     3                       isaiah rashad
#> 76     3 musical journey to the ends of time
#> 77     3                                 sza
#> 78     3                            hameryka
#> 79     3                    underground soul
#> 80     3                robertitus coleccion
#> 81     3                               glyph
#> 82     3                          best curls
#> 83     3        alternative rhythm and blues
#> 84     3       eats tylenol like a muthufuka
#> 85     3                         scott moyer
#> 86     3                 buckets and baskets
#> 87     3                          9876543210
#> 88     3                 3 and 5 and 7 and 9
#> 89     3                                crnb
#> 90     3                                 <><
#>                                                            url
#> 1                                 https://www.last.fm/tag/soul
#> 2                           https://www.last.fm/tag/electronic
#> 3                                  https://www.last.fm/tag/rnb
#> 4                      https://www.last.fm/tag/alternative+rnb
#> 5                     https://www.last.fm/tag/female+vocalists
#> 6                          https://www.last.fm/tag/alternative
#> 7                             https://www.last.fm/tag/Neo-Soul
#> 8                             https://www.last.fm/tag/american
#> 9                            https://www.last.fm/tag/seen+live
#> 10                             https://www.last.fm/tag/Hip-Hop
#> 11                               https://www.last.fm/tag/indie
#> 12                                 https://www.last.fm/tag/r&b
#> 13                            https://www.last.fm/tag/Neo+Soul
#> 14                                 https://www.last.fm/tag/tde
#> 15                        https://www.last.fm/tag/glitter+trap
#> 16                                https://www.last.fm/tag/trap
#> 17                               https://www.last.fm/tag/wonky
#> 18                         https://www.last.fm/tag/electronica
#> 19                                 https://www.last.fm/tag/pop
#> 20                           https://www.last.fm/tag/dream+pop
#> 21                            https://www.last.fm/tag/synthpop
#> 22                     https://www.last.fm/tag/female+vocalist
#> 23                      https://www.last.fm/tag/rhythm+&+blues
#> 24                          https://www.last.fm/tag/future+rnb
#> 25                            https://www.last.fm/tag/trip-hop
#> 26                           https://www.last.fm/tag/downtempo
#> 27                   https://www.last.fm/tag/singer-songwriter
#> 28                                 https://www.last.fm/tag/10s
#> 29                          https://www.last.fm/tag/new+jersey
#> 30                            https://www.last.fm/tag/Missouri
#> 31                            https://www.last.fm/tag/st+louis
#> 32                           https://www.last.fm/tag/r%27n%27b
#> 33                    https://www.last.fm/tag/contemporary+rnb
#> 34                    https://www.last.fm/tag/contemporary+r&b
#> 35                   https://www.last.fm/tag/perfect+vocalists
#> 36                                 https://www.last.fm/tag/rap
#> 37                             https://www.last.fm/tag/ambient
#> 38                        https://www.last.fm/tag/experimental
#> 39                                 https://www.last.fm/tag/USA
#> 40                               https://www.last.fm/tag/urban
#> 41                    https://www.last.fm/tag/Psychedelic+Soul
#> 42                                https://www.last.fm/tag/funk
#> 43                 https://www.last.fm/tag/underground+hip-hop
#> 44                              https://www.last.fm/tag/Dreamy
#> 45                            https://www.last.fm/tag/ethereal
#> 46                                   https://www.last.fm/tag/f
#> 47                            https://www.last.fm/tag/wishlist
#> 48                               https://www.last.fm/tag/jazzy
#> 49                    https://www.last.fm/tag/rhythm+and+blues
#> 50                           https://www.last.fm/tag/chillwave
#> 51                             https://www.last.fm/tag/pop+rap
#> 52                               https://www.last.fm/tag/2010s
#> 53                         https://www.last.fm/tag/songwriting
#> 54                         https://www.last.fm/tag/witch+house
#> 55                             https://www.last.fm/tag/america
#> 56                       https://www.last.fm/tag/disappointing
#> 57                           https://www.last.fm/tag/cloud+rap
#> 58                                https://www.last.fm/tag/list
#> 59                      https://www.last.fm/tag/kendrick+lamar
#> 60                             https://www.last.fm/tag/scorpio
#> 61                              https://www.last.fm/tag/to+try
#> 62                        https://www.last.fm/tag/electro+soul
#> 63                           https://www.last.fm/tag/Maplewood
#> 64                            https://www.last.fm/tag/newbreed
#> 65                          https://www.last.fm/tag/futuregaze
#> 66                          https://www.last.fm/tag/indie+soul
#> 67                             https://www.last.fm/tag/Popstep
#> 68                  https://www.last.fm/tag/rhythm%27n%27blues
#> 69                               https://www.last.fm/tag/again
#> 70                            https://www.last.fm/tag/jay+rock
#> 71                             https://www.last.fm/tag/neo-sul
#> 72                         https://www.last.fm/tag/schoolboy+q
#> 73                             https://www.last.fm/tag/Ab-Soul
#> 74                         https://www.last.fm/tag/soultronica
#> 75                       https://www.last.fm/tag/isaiah+rashad
#> 76 https://www.last.fm/tag/musical+journey+to+the+ends+of+time
#> 77                                 https://www.last.fm/tag/sza
#> 78                            https://www.last.fm/tag/hameryka
#> 79                    https://www.last.fm/tag/underground+soul
#> 80                https://www.last.fm/tag/robertitus+coleccion
#> 81                               https://www.last.fm/tag/glyph
#> 82                          https://www.last.fm/tag/best+curls
#> 83        https://www.last.fm/tag/alternative+rhythm+and+blues
#> 84       https://www.last.fm/tag/eats+tylenol+like+a+muthufuka
#> 85                         https://www.last.fm/tag/scott+moyer
#> 86                 https://www.last.fm/tag/buckets+and+baskets
#> 87                          https://www.last.fm/tag/9876543210
#> 88                 https://www.last.fm/tag/3+and+5+and+7+and+9
#> 89                                https://www.last.fm/tag/crnb
#> 90                           https://www.last.fm/tag/%3C%3E%3C