Get corrections of a track title or artist.

Use the last.fm corrections data to check whether the supplied track has a correction to a canonical track Implementation of last.fm's track.getCorrection API method (http://www.last.fm/api/show/track.getCorrection)

track_getCorrection(track, artist, api_key = lastkey)

Arguments

track
The track name to correct.

artist
The artist name to correct.

api_key
A Last.fm API key.

Value

A list of corrections.

Examples

track_getCorrection("King Kunta", "Kendrick Lamar")

#> $correction
#> $correction$track
#> $correction$track$name
#> [1] "King Kunta"
#> 
#> $correction$track$url
#> [1] "https://www.last.fm/music/Kendrick+Lamar/_/King+Kunta"
#> 
#> $correction$track$artist
#> $correction$track$artist$name
#> [1] "Kendrick Lamar"
#> 
#> $correction$track$artist$mbid
#> [1] "381086ea-f511-4aba-bdf9-71c753dc5077"
#> 
#> $correction$track$artist$url
#> [1] "https://www.last.fm/music/Kendrick+Lamar"
#> 
#> 
#> 
#> $correction$`@attr`
#> $correction$`@attr`$index
#> [1] "0"
#> 
#> $correction$`@attr`$artistcorrected
#> [1] "0"
#> 
#> $correction$`@attr`$trackcorrected
#> [1] "0"
#> 
#> 
#>