Get the similar tracks for this track.

Get the similar tracks for this track on Last.fm, based on listening data. Implementation of last.fm's track.getSimilar API method (http://www.last.fm/api/show/track.getSimilar)

track_getSimilar(track, artist, mbid = NA, autocorrect = NA,
  limit = NA)

Arguments

track
The track name.

artist
The artist name.

mbid
The musicbrainz id for the track

autocorrect
Transform misspelled artist and track names into correct artist and track names, returning the correct version instead. The corrected artist and track name will be returned in the response. [0|1]

limit
Maximum number of similar tracks to return.

Value

A list of the similar tracks.

Examples

# NOT RUN {
track_getSimilar("Hard Times", "Thundercat")
# }