Get the top tags for this track.

Get the top tags for this track on Last.fm, ordered by tag count. Supply either track & artist name or mbid. Implementation of last.fm's track.getTopTags API method (http://www.last.fm/api/show/track.getTopTags)

track_getTopTags(track, artist, mbid = NA, autocorrect = NA,
  api_key = lastkey)

Arguments

track
The track name.

artist
The artist name.

mbid
The musicbrainz id for the track

autocorrect
Transform misspelled artist and track names into correct artist and track names, returning the correct version instead. The corrected artist and track name will be returned in the response. [0|1]

api_key
A Last.fm API key.

Value

A list of the top tags for a track.

Examples

track_getTopTags("Tommy Chong", "The Blue Scholars")

#> $tag
#>   count              name                                       url
#> 1   100           Hip-Hop           https://www.last.fm/tag/Hip-Hop
#> 2   100           hip hop           https://www.last.fm/tag/hip+hop
#> 3   100               rap               https://www.last.fm/tag/rap
#> 4   100              2011              https://www.last.fm/tag/2011
#> 5   100 blaze to this one https://www.last.fm/tag/blaze+to+this+one
#> 
#> $`@attr`
#> $`@attr`$artist
#> [1] "Blue Scholars"
#> 
#> $`@attr`$track
#> [1] "Tommy Chong"
#> 
#>