Get a list of tracks by a given artist scrobbled by this user.

Get a list of tracks by a given artist scrobbled by this user, including scrobble time. Can be limited to specific timeranges, defaults to all time. Implementation of last.fm's user.getArtistTracks API method (http://www.last.fm/api/show/user.getArtistTracks).

user_getArtistTracks(user, artist, startTimestamp = NA, endTimestamp = NA,
  page = NA, api_key = lastkey)

Arguments

user
The last.fm username to fetch the recent tracks of.

artist
The artist name you are interested in.

startTimestamp
An unix timestamp to start at.

endTimestamp
An unix timestamp to end at.

page
The page number to fetch. Defaults to first page.

api_key
A Last.fm API key.

Value

A list of tracks by a given artist scrobbled by this user.

Examples

user_getArtistTracks("platyjus", "Leon Bridges")

#> $track
#>    artist.#text artist.mbid                                         name
#> 1  Leon Bridges                                           There She Goes
#> 2  Leon Bridges                                                    River
#> 3  Leon Bridges                                      Twistin' & Groovin'
#> 4  Leon Bridges                                                Pull Away
#> 5  Leon Bridges                                                  Flowers
#> 6  Leon Bridges                                                    Shine
#> 7  Leon Bridges                                           Smooth Sailin'
#> 8  Leon Bridges                                          Brown Skin Girl
#> 9  Leon Bridges                                               Better Man
#> 10 Leon Bridges                                              Coming Home
#> 11 Leon Bridges                                          Brown Skin Girl
#> 12 Leon Bridges                                               Better Man
#> 13 Leon Bridges                                              Coming Home
#> 14 Leon Bridges                                                Pull Away
#> 15 Leon Bridges                                                    Shine
#> 16 Leon Bridges             So Long - From The Motion Picture Concussion
#> 17 Leon Bridges                                              Lisa Sawyer
#> 18 Leon Bridges                                          Here In My Arms
#> 19 Leon Bridges                                          Brown Skin Girl
#> 20 Leon Bridges                                               Better Man
#> 21 Leon Bridges                                           Smooth Sailin'
#> 22 Leon Bridges                                                    River
#> 23 Leon Bridges                                              Coming Home
#> 24 Leon Bridges                                          Brown Skin Girl
#> 25 Leon Bridges                                               Better Man
#> 26 Leon Bridges                                                    River
#> 27 Leon Bridges                                              Coming Home
#> 28 Leon Bridges                                               Outta Line
#> 29 Leon Bridges                                          Here In My Arms
#> 30 Leon Bridges                                       Mississippi Kisses
#> 31 Leon Bridges                                                Daisy Mae
#> 32 Leon Bridges                                           There She Goes
#> 33 Leon Bridges                                                    River
#> 34 Leon Bridges                                      Twistin' & Groovin'
#> 35 Leon Bridges                                                Pull Away
#> 36 Leon Bridges                                                  Flowers
#> 37 Leon Bridges                                              Lisa Sawyer
#> 38 Leon Bridges                                                    Shine
#> 39 Leon Bridges                                           Smooth Sailin'
#> 40 Leon Bridges                                          Brown Skin Girl
#> 41 Leon Bridges                                              Coming Home
#> 42 Leon Bridges                                              Lisa Sawyer
#> 43 Leon Bridges                                               Better Man
#> 44 Leon Bridges                                              Coming Home
#> 45 Leon Bridges                                        Ball Of Confusion
#> 46 Leon Bridges                                                    Shine
#> 47 Leon Bridges                                                Pull Away
#> 48 Leon Bridges                                          Here In My Arms
#> 49 Leon Bridges             So Long - From The Motion Picture Concussion
#> 50 Leon Bridges                                              Lisa Sawyer
#>    streamable mbid
#> 1           0     
#> 2           0     
#> 3           0     
#> 4           0     
#> 5           0     
#> 6           0     
#> 7           0     
#> 8           0     
#> 9           0     
#> 10          0     
#> 11          0     
#> 12          0     
#> 13          0     
#> 14          0     
#> 15          0     
#> 16          0     
#> 17          0     
#> 18          0     
#> 19          0     
#> 20          0     
#> 21          0     
#> 22          0     
#> 23          0     
#> 24          0     
#> 25          0     
#> 26          0     
#> 27          0     
#> 28          0     
#> 29          0     
#> 30          0     
#> 31          0     
#> 32          0     
#> 33          0     
#> 34          0     
#> 35          0     
#> 36          0     
#> 37          0     
#> 38          0     
#> 39          0     
#> 40          0     
#> 41          0     
#> 42          0     
#> 43          0     
#> 44          0     
#> 45          0     
#> 46          0     
#> 47          0     
#> 48          0     
#> 49          0     
#> 50          0     
#>                                                                            album.#text
#> 1                                                                 Coming Home (Deluxe)
#> 2                                                                 Coming Home (Deluxe)
#> 3                                                                 Coming Home (Deluxe)
#> 4                                                                 Coming Home (Deluxe)
#> 5                                                                 Coming Home (Deluxe)
#> 6                                                                 Coming Home (Deluxe)
#> 7                                                                 Coming Home (Deluxe)
#> 8                                                                 Coming Home (Deluxe)
#> 9                                                                 Coming Home (Deluxe)
#> 10                                                                Coming Home (Deluxe)
#> 11                                                                Coming Home (Deluxe)
#> 12                                                                Coming Home (Deluxe)
#> 13                                                                Coming Home (Deluxe)
#> 14                                                                         Coming Home
#> 15                                                                         Coming Home
#> 16                                        So Long (From The Motion Picture Concussion)
#> 17                                                                         Coming Home
#> 18                                                                Coming Home (Deluxe)
#> 19                                                                         Coming Home
#> 20                                                                         Coming Home
#> 21                                                                         Coming Home
#> 22                                                                         Coming Home
#> 23                                                                         Coming Home
#> 24                                                                         Coming Home
#> 25                                                                         Coming Home
#> 26                                                                         Coming Home
#> 27                                                                         Coming Home
#> 28                                                                Coming Home (Deluxe)
#> 29                                                                Coming Home (Deluxe)
#> 30                                                                Coming Home (Deluxe)
#> 31                                                                Coming Home (Deluxe)
#> 32                                                                Coming Home (Deluxe)
#> 33                                                                         Coming Home
#> 34                                                                         Coming Home
#> 35                                                                         Coming Home
#> 36                                                                         Coming Home
#> 37                                                                         Coming Home
#> 38                                                                         Coming Home
#> 39                                                                         Coming Home
#> 40                                                                         Coming Home
#> 41                                                                         Coming Home
#> 42                                                                         Coming Home
#> 43                                                                         Coming Home
#> 44                                                                         Coming Home
#> 45 The Get Down: Original Soundtrack From The Netflix Original Series (Deluxe Version)
#> 46                                                                         Coming Home
#> 47                                                                         Coming Home
#> 48                                                                Coming Home (Deluxe)
#> 49                                        So Long (From The Motion Picture Concussion)
#> 50                                                                         Coming Home
#>    album.mbid
#> 1            
#> 2            
#> 3            
#> 4            
#> 5            
#> 6            
#> 7            
#> 8            
#> 9            
#> 10           
#> 11           
#> 12           
#> 13           
#> 14           
#> 15           
#> 16           
#> 17           
#> 18           
#> 19           
#> 20           
#> 21           
#> 22           
#> 23           
#> 24           
#> 25           
#> 26           
#> 27           
#> 28           
#> 29           
#> 30           
#> 31           
#> 32           
#> 33           
#> 34           
#> 35           
#> 36           
#> 37           
#> 38           
#> 39           
#> 40           
#> 41           
#> 42           
#> 43           
#> 44           
#> 45           
#> 46           
#> 47           
#> 48           
#> 49           
#> 50           
#>                                                                                      url
#> 1                                https://www.last.fm/music/Leon+Bridges/_/There+She+Goes
#> 2                                         https://www.last.fm/music/Leon+Bridges/_/River
#> 3                       https://www.last.fm/music/Leon+Bridges/_/Twistin%27+&+Groovin%27
#> 4                                     https://www.last.fm/music/Leon+Bridges/_/Pull+Away
#> 5                                       https://www.last.fm/music/Leon+Bridges/_/Flowers
#> 6                                         https://www.last.fm/music/Leon+Bridges/_/Shine
#> 7                              https://www.last.fm/music/Leon+Bridges/_/Smooth+Sailin%27
#> 8                               https://www.last.fm/music/Leon+Bridges/_/Brown+Skin+Girl
#> 9                                    https://www.last.fm/music/Leon+Bridges/_/Better+Man
#> 10                                  https://www.last.fm/music/Leon+Bridges/_/Coming+Home
#> 11                              https://www.last.fm/music/Leon+Bridges/_/Brown+Skin+Girl
#> 12                                   https://www.last.fm/music/Leon+Bridges/_/Better+Man
#> 13                                  https://www.last.fm/music/Leon+Bridges/_/Coming+Home
#> 14                                    https://www.last.fm/music/Leon+Bridges/_/Pull+Away
#> 15                                        https://www.last.fm/music/Leon+Bridges/_/Shine
#> 16 https://www.last.fm/music/Leon+Bridges/_/So+Long+-+From+The+Motion+Picture+Concussion
#> 17                                  https://www.last.fm/music/Leon+Bridges/_/Lisa+Sawyer
#> 18                              https://www.last.fm/music/Leon+Bridges/_/Here+In+My+Arms
#> 19                              https://www.last.fm/music/Leon+Bridges/_/Brown+Skin+Girl
#> 20                                   https://www.last.fm/music/Leon+Bridges/_/Better+Man
#> 21                             https://www.last.fm/music/Leon+Bridges/_/Smooth+Sailin%27
#> 22                                        https://www.last.fm/music/Leon+Bridges/_/River
#> 23                                  https://www.last.fm/music/Leon+Bridges/_/Coming+Home
#> 24                              https://www.last.fm/music/Leon+Bridges/_/Brown+Skin+Girl
#> 25                                   https://www.last.fm/music/Leon+Bridges/_/Better+Man
#> 26                                        https://www.last.fm/music/Leon+Bridges/_/River
#> 27                                  https://www.last.fm/music/Leon+Bridges/_/Coming+Home
#> 28                                   https://www.last.fm/music/Leon+Bridges/_/Outta+Line
#> 29                              https://www.last.fm/music/Leon+Bridges/_/Here+In+My+Arms
#> 30                           https://www.last.fm/music/Leon+Bridges/_/Mississippi+Kisses
#> 31                                    https://www.last.fm/music/Leon+Bridges/_/Daisy+Mae
#> 32                               https://www.last.fm/music/Leon+Bridges/_/There+She+Goes
#> 33                                        https://www.last.fm/music/Leon+Bridges/_/River
#> 34                      https://www.last.fm/music/Leon+Bridges/_/Twistin%27+&+Groovin%27
#> 35                                    https://www.last.fm/music/Leon+Bridges/_/Pull+Away
#> 36                                      https://www.last.fm/music/Leon+Bridges/_/Flowers
#> 37                                  https://www.last.fm/music/Leon+Bridges/_/Lisa+Sawyer
#> 38                                        https://www.last.fm/music/Leon+Bridges/_/Shine
#> 39                             https://www.last.fm/music/Leon+Bridges/_/Smooth+Sailin%27
#> 40                              https://www.last.fm/music/Leon+Bridges/_/Brown+Skin+Girl
#> 41                                  https://www.last.fm/music/Leon+Bridges/_/Coming+Home
#> 42                                  https://www.last.fm/music/Leon+Bridges/_/Lisa+Sawyer
#> 43                                   https://www.last.fm/music/Leon+Bridges/_/Better+Man
#> 44                                  https://www.last.fm/music/Leon+Bridges/_/Coming+Home
#> 45                            https://www.last.fm/music/Leon+Bridges/_/Ball+Of+Confusion
#> 46                                        https://www.last.fm/music/Leon+Bridges/_/Shine
#> 47                                    https://www.last.fm/music/Leon+Bridges/_/Pull+Away
#> 48                              https://www.last.fm/music/Leon+Bridges/_/Here+In+My+Arms
#> 49 https://www.last.fm/music/Leon+Bridges/_/So+Long+-+From+The+Motion+Picture+Concussion
#> 50                                  https://www.last.fm/music/Leon+Bridges/_/Lisa+Sawyer
#>                                                                                                                                                                                                                                                                                                                                                                    image
#> 1  https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 2  https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 3  https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 4  https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 5  https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 6  https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 7  https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 8  https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 9  https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 10 https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 11 https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 12 https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 13 https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 14 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 15 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 16 https://lastfm-img2.akamaized.net/i/u/34s/e87809160cb0f03196aa1b9168402fca.png, https://lastfm-img2.akamaized.net/i/u/64s/e87809160cb0f03196aa1b9168402fca.png, https://lastfm-img2.akamaized.net/i/u/174s/e87809160cb0f03196aa1b9168402fca.png, https://lastfm-img2.akamaized.net/i/u/300x300/e87809160cb0f03196aa1b9168402fca.png, small, medium, large, extralarge
#> 17 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 18 https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 19 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 20 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 21 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 22 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 23 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 24 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 25 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 26 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 27 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 28 https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 29 https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 30 https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 31 https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 32 https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 33 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 34 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 35 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 36 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 37 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 38 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 39 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 40 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 41 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 42 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 43 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 44 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 45 https://lastfm-img2.akamaized.net/i/u/34s/e76ff93b6a2ee8acb2872a362b1e26ee.png, https://lastfm-img2.akamaized.net/i/u/64s/e76ff93b6a2ee8acb2872a362b1e26ee.png, https://lastfm-img2.akamaized.net/i/u/174s/e76ff93b6a2ee8acb2872a362b1e26ee.png, https://lastfm-img2.akamaized.net/i/u/300x300/e76ff93b6a2ee8acb2872a362b1e26ee.png, small, medium, large, extralarge
#> 46 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 47 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 48 https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 49 https://lastfm-img2.akamaized.net/i/u/34s/e87809160cb0f03196aa1b9168402fca.png, https://lastfm-img2.akamaized.net/i/u/64s/e87809160cb0f03196aa1b9168402fca.png, https://lastfm-img2.akamaized.net/i/u/174s/e87809160cb0f03196aa1b9168402fca.png, https://lastfm-img2.akamaized.net/i/u/300x300/e87809160cb0f03196aa1b9168402fca.png, small, medium, large, extralarge
#> 50 https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#>      date.uts         date.#text
#> 1  1490198192 22 Mar 2017, 15:56
#> 2  1490197906 22 Mar 2017, 15:51
#> 3  1490112537 21 Mar 2017, 16:08
#> 4  1490112358 21 Mar 2017, 16:05
#> 5  1490112183 21 Mar 2017, 16:03
#> 6  1490024322 20 Mar 2017, 15:38
#> 7  1490024139 20 Mar 2017, 15:35
#> 8  1490023931 20 Mar 2017, 15:32
#> 9  1490023790 20 Mar 2017, 15:29
#> 10 1490023564 20 Mar 2017, 15:26
#> 11 1486786363 11 Feb 2017, 04:12
#> 12 1486786225 11 Feb 2017, 04:10
#> 13 1486786019 11 Feb 2017, 04:06
#> 14 1485293689 24 Jan 2017, 21:34
#> 15 1485293471 24 Jan 2017, 21:31
#> 16 1485293240 24 Jan 2017, 21:27
#> 17 1485292995 24 Jan 2017, 21:23
#> 18 1485292868 24 Jan 2017, 21:21
#> 19 1485292661 24 Jan 2017, 21:17
#> 20 1485292521 24 Jan 2017, 21:15
#> 21 1485292337 24 Jan 2017, 21:12
#> 22 1485292098 24 Jan 2017, 21:08
#> 23 1485291893 24 Jan 2017, 21:04
#> 24 1483852701 08 Jan 2017, 05:18
#> 25 1483852562 08 Jan 2017, 05:16
#> 26 1483852361 08 Jan 2017, 05:12
#> 27 1483852156 08 Jan 2017, 05:09
#> 28 1483064395 30 Dec 2016, 02:19
#> 29 1483064268 30 Dec 2016, 02:17
#> 30 1483064123 30 Dec 2016, 02:15
#> 31 1483063947 30 Dec 2016, 02:12
#> 32 1483063790 30 Dec 2016, 02:09
#> 33 1483063551 30 Dec 2016, 02:05
#> 34 1483063299 30 Dec 2016, 02:01
#> 35 1483063119 30 Dec 2016, 01:58
#> 36 1483062942 30 Dec 2016, 01:55
#> 37 1483062698 30 Dec 2016, 01:51
#> 38 1483062480 30 Dec 2016, 01:48
#> 39 1483062297 30 Dec 2016, 01:44
#> 40 1483062090 30 Dec 2016, 01:41
#> 41 1483061755 30 Dec 2016, 01:35
#> 42 1480697623 02 Dec 2016, 16:53
#> 43 1480697472 02 Dec 2016, 16:51
#> 44 1480697266 02 Dec 2016, 16:47
#> 45 1480260870 27 Nov 2016, 15:34
#> 46 1478371218 05 Nov 2016, 18:40
#> 47 1478371037 05 Nov 2016, 18:37
#> 48 1478370909 05 Nov 2016, 18:35
#> 49 1478370678 05 Nov 2016, 18:31
#> 50 1478370433 05 Nov 2016, 18:27
#> 
#> $`@attr`
#> $`@attr`$user
#> [1] "platyjus"
#> 
#> $`@attr`$artist
#> [1] "Leon Bridges"
#> 
#> $`@attr`$page
#> [1] "1"
#> 
#> $`@attr`$perPage
#> [1] "50"
#> 
#> $`@attr`$totalPages
#> [1] "0"
#> 
#> $`@attr`$total
#> [1] "0"
#> 
#>