Get a list of the recent tracks listened to by user.

Get a list of the recent tracks listened to by this user. Also includes the currently playing track with the nowplaying="true" attribute if the user is currently listening. Implementation of last.fm's user.getRecentTracks API method (http://www.last.fm/api/show/user.getRecentTracks).

user_getRecentTracks(user, limit = NA, page = 1, from = NA, to = NA,
  extended = 0, api_key = lastkey)

Arguments

user
The last.fm username to fetch the recent tracks of.

limit
The number of results to fetch per page. Defaults to 50. Maximum is 200.

page
The page number you wish to scan to.

from
Beginning timestamp of a range - only display scrobbles after this time, in UNIX timestamp format (integer number of seconds since 00:00:00, January 1st 1970 UTC). This must be in the UTC time zone.

to
End timestamp of a range - only display scrobbles before this time, in UNIX timestamp format (integer number of seconds since 00:00:00, January 1st 1970 UTC). This must be in the UTC time zone.

extended
Includes extended data in each artist, and whether or not the user has loved each track

api_key
A Last.fm API key.

Value

A list of the recent tracks.

Examples

user_getRecentTracks("platyjus")

#> $track
#>        artist.#text                          artist.mbid
#> 1           Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 2           Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 3           Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 4           Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 5      Joan Shelley                                     
#> 6      Joan Shelley                                     
#> 7      Joan Shelley                                     
#> 8      Joan Shelley                                     
#> 9      Joan Shelley                                     
#> 10     Joan Shelley                                     
#> 11     Joan Shelley                                     
#> 12     Joan Shelley                                     
#> 13     Joan Shelley                                     
#> 14        Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993
#> 15        Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993
#> 16        Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993
#> 17        Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993
#> 18           Khalid c6bfb05d-f570-46c8-98e1-e25441189770
#> 19           Khalid c6bfb05d-f570-46c8-98e1-e25441189770
#> 20   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 21   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 22   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 23   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 24   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 25   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 26 Childish Gambino 7fb57fba-a6ef-44c2-abab-2fa3bdee607e
#> 27 Childish Gambino 7fb57fba-a6ef-44c2-abab-2fa3bdee607e
#> 28 Childish Gambino 7fb57fba-a6ef-44c2-abab-2fa3bdee607e
#> 29 Childish Gambino 7fb57fba-a6ef-44c2-abab-2fa3bdee607e
#> 30   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 31   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 32   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 33   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 34   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 35   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 36   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 37   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 38   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 39   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 40   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 41   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 42   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 43          Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 44          Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 45      Joey Bada$$ 9757ae7d-44b1-4d63-81ce-b463fe27b9ae
#> 46      Joey Bada$$ 9757ae7d-44b1-4d63-81ce-b463fe27b9ae
#> 47      Joey Bada$$ 9757ae7d-44b1-4d63-81ce-b463fe27b9ae
#> 48          Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 49          Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 50          Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#>                                      name streamable
#> 1                            Hold It Back          0
#> 2                       Built by Pictures          0
#> 3                                  Things          0
#> 4                            Digging Deep          0
#> 5                        Not Over By Half          0
#> 6                           Over and Even          0
#> 7                        Stay On My Shore          0
#> 8                 Brighter Than The Blues          0
#> 9                       The Push and Pull          0
#> 10                      Wild Indifference          0
#> 11                    Where I’ll Find You          0
#> 12               If the Storms Never Came          0
#> 13                           We’d Be Home          0
#> 14                                  Cupid          0
#> 15                Chain Gang - Remastered          0
#> 16                            You Send Me          0
#> 17                 A Change Is Gonna Come          0
#> 18                     Young Dumb & Broke          0
#> 19                               LOCATION          0
#> 20                          Not An Option          0
#> 21                               Patience          0
#> 22                             Fix You Up          0
#> 23 Everything Looks Better (In Hindsight)          0
#> 24                          Fall To Sleep          0
#> 25                             Only Songs          0
#> 26                             Stand Tall          0
#> 27         The Night Me And Your Mama Met          0
#> 28                               Baby Boy          0
#> 29                              terrified          0
#> 30                               Patience          0
#> 31                             Fix You Up          0
#> 32 Everything Looks Better (In Hindsight)          0
#> 33                          Fall To Sleep          0
#> 34                Only Songs (Radio Edit)          0
#> 35                               Fruition          0
#> 36                     The World We Built          0
#> 37                      Catch And Release          0
#> 38                          Back to Earth          0
#> 39                                Capable          0
#> 40                          Not An Option          0
#> 41                               Patience          0
#> 42                             Fix You Up          0
#> 43                            That's Love          0
#> 44                      Built by Pictures          0
#> 45      ROCKABYE BABY (feat. ScHoolboy Q)          0
#> 46                          FOR MY PEOPLE          0
#> 47                 GOOD MORNING AMERIKKKA          0
#> 48                             Want to Be          0
#> 49                            Like Really          0
#> 50                                   NNGE          0
#>                                    mbid
#> 1                                      
#> 2                                      
#> 3                                      
#> 4                                      
#> 5                                      
#> 6                                      
#> 7                                      
#> 8                                      
#> 9                                      
#> 10                                     
#> 11                                     
#> 12                                     
#> 13                                     
#> 14 8086b120-296a-4a08-91eb-0cac99dcfc38
#> 15                                     
#> 16 6a7a997c-bb78-408c-8b06-7a39243f416e
#> 17 2f82aad5-728a-4094-b266-a201e065390d
#> 18                                     
#> 19                                     
#> 20                                     
#> 21                                     
#> 22                                     
#> 23                                     
#> 24                                     
#> 25                                     
#> 26                                     
#> 27                                     
#> 28                                     
#> 29                                     
#> 30                                     
#> 31                                     
#> 32                                     
#> 33                                     
#> 34                                     
#> 35                                     
#> 36                                     
#> 37                                     
#> 38                                     
#> 39                                     
#> 40                                     
#> 41                                     
#> 42                                     
#> 43                                     
#> 44                                     
#> 45                                     
#> 46                                     
#> 47                                     
#> 48                                     
#> 49                                     
#> 50                                     
#>                                         album.#text
#> 1                                       The Iceberg
#> 2                                       The Iceberg
#> 3                                       The Iceberg
#> 4                                       The Iceberg
#> 5                                 Over and Even + 3
#> 6                                 Over and Even + 3
#> 7                                 Over and Even + 3
#> 8                                 Over and Even + 3
#> 9                                      Joan Shelley
#> 10                                     Joan Shelley
#> 11                                     Joan Shelley
#> 12                                     Joan Shelley
#> 13                                     Joan Shelley
#> 14 30 Greatest Hits: Portrait of a Legend 1951-1964
#> 15                                        Swing Low
#> 16 30 Greatest Hits: Portrait of a Legend 1951-1964
#> 17                Ain't That Good News (Remastered)
#> 18                                    American Teen
#> 19                                    American Teen
#> 20                               The World We Built
#> 21                               The World We Built
#> 22                               The World We Built
#> 23                               The World We Built
#> 24                               The World We Built
#> 25                               The World We Built
#> 26                               "Awaken, My Love!"
#> 27                               "Awaken, My Love!"
#> 28                               "Awaken, My Love!"
#> 29                               "Awaken, My Love!"
#> 30                               The World We Built
#> 31                               The World We Built
#> 32                               The World We Built
#> 33                               The World We Built
#> 34                               The World We Built
#> 35                               The World We Built
#> 36                               The World We Built
#> 37                               The World We Built
#> 38                               The World We Built
#> 39                               The World We Built
#> 40                               The World We Built
#> 41                               The World We Built
#> 42                               The World We Built
#> 43                                   The Good Fight
#> 44                                      The Iceberg
#> 45                            ALL-AMERIKKKAN BADA$$
#> 46                            ALL-AMERIKKKAN BADA$$
#> 47                            ALL-AMERIKKKAN BADA$$
#> 48                                      The Iceberg
#> 49                                      The Iceberg
#> 50                                      The Iceberg
#>                              album.mbid
#> 1                                      
#> 2                                      
#> 3                                      
#> 4                                      
#> 5                                      
#> 6                                      
#> 7                                      
#> 8                                      
#> 9                                      
#> 10                                     
#> 11                                     
#> 12                                     
#> 13                                     
#> 14                                     
#> 15 8ac8fd98-f1eb-4049-8cea-9be6e2268272
#> 16                                     
#> 17                                     
#> 18                                     
#> 19                                     
#> 20                                     
#> 21                                     
#> 22                                     
#> 23                                     
#> 24                                     
#> 25                                     
#> 26                                     
#> 27                                     
#> 28                                     
#> 29                                     
#> 30                                     
#> 31                                     
#> 32                                     
#> 33                                     
#> 34                                     
#> 35                                     
#> 36                                     
#> 37                                     
#> 38                                     
#> 39                                     
#> 40                                     
#> 41                                     
#> 42                                     
#> 43                                     
#> 44                                     
#> 45                                     
#> 46                                     
#> 47                                     
#> 48                                     
#> 49                                     
#> 50                                     
#>                                                                                  url
#> 1                                   https://www.last.fm/music/Oddisee/_/Hold+It+Back
#> 2                              https://www.last.fm/music/Oddisee/_/Built+by+Pictures
#> 3                                         https://www.last.fm/music/Oddisee/_/Things
#> 4                                   https://www.last.fm/music/Oddisee/_/Digging+Deep
#> 5                          https://www.last.fm/music/Joan+Shelley/_/Not+Over+By+Half
#> 6                             https://www.last.fm/music/Joan+Shelley/_/Over+and+Even
#> 7                          https://www.last.fm/music/Joan+Shelley/_/Stay+On+My+Shore
#> 8                   https://www.last.fm/music/Joan+Shelley/_/Brighter+Than+The+Blues
#> 9                         https://www.last.fm/music/Joan+Shelley/_/The+Push+and+Pull
#> 10                        https://www.last.fm/music/Joan+Shelley/_/Wild+Indifference
#> 11              https://www.last.fm/music/Joan+Shelley/_/Where+I%E2%80%99ll+Find+You
#> 12                 https://www.last.fm/music/Joan+Shelley/_/If+the+Storms+Never+Came
#> 13                     https://www.last.fm/music/Joan+Shelley/_/We%E2%80%99d+Be+Home
#> 14                                       https://www.last.fm/music/Sam+Cooke/_/Cupid
#> 15                     https://www.last.fm/music/Sam+Cooke/_/Chain+Gang+-+Remastered
#> 16                                 https://www.last.fm/music/Sam+Cooke/_/You+Send+Me
#> 17                      https://www.last.fm/music/Sam+Cooke/_/A+Change+Is+Gonna+Come
#> 18                             https://www.last.fm/music/Khalid/_/Young+Dumb+&+Broke
#> 19                                       https://www.last.fm/music/Khalid/_/LOCATION
#> 20                          https://www.last.fm/music/The+Wild+Reeds/_/Not+An+Option
#> 21                               https://www.last.fm/music/The+Wild+Reeds/_/Patience
#> 22                             https://www.last.fm/music/The+Wild+Reeds/_/Fix+You+Up
#> 23 https://www.last.fm/music/The+Wild+Reeds/_/Everything+Looks+Better+(In+Hindsight)
#> 24                          https://www.last.fm/music/The+Wild+Reeds/_/Fall+To+Sleep
#> 25                             https://www.last.fm/music/The+Wild+Reeds/_/Only+Songs
#> 26                           https://www.last.fm/music/Childish+Gambino/_/Stand+Tall
#> 27       https://www.last.fm/music/Childish+Gambino/_/The+Night+Me+And+Your+Mama+Met
#> 28                             https://www.last.fm/music/Childish+Gambino/_/Baby+Boy
#> 29                            https://www.last.fm/music/Childish+Gambino/_/terrified
#> 30                               https://www.last.fm/music/The+Wild+Reeds/_/Patience
#> 31                             https://www.last.fm/music/The+Wild+Reeds/_/Fix+You+Up
#> 32 https://www.last.fm/music/The+Wild+Reeds/_/Everything+Looks+Better+(In+Hindsight)
#> 33                          https://www.last.fm/music/The+Wild+Reeds/_/Fall+To+Sleep
#> 34                https://www.last.fm/music/The+Wild+Reeds/_/Only+Songs+(Radio+Edit)
#> 35                               https://www.last.fm/music/The+Wild+Reeds/_/Fruition
#> 36                     https://www.last.fm/music/The+Wild+Reeds/_/The+World+We+Built
#> 37                      https://www.last.fm/music/The+Wild+Reeds/_/Catch+And+Release
#> 38                          https://www.last.fm/music/The+Wild+Reeds/_/Back+to+Earth
#> 39                                https://www.last.fm/music/The+Wild+Reeds/_/Capable
#> 40                          https://www.last.fm/music/The+Wild+Reeds/_/Not+An+Option
#> 41                               https://www.last.fm/music/The+Wild+Reeds/_/Patience
#> 42                             https://www.last.fm/music/The+Wild+Reeds/_/Fix+You+Up
#> 43                                 https://www.last.fm/music/Oddisee/_/That%27s+Love
#> 44                             https://www.last.fm/music/Oddisee/_/Built+by+Pictures
#> 45         https://www.last.fm/music/Joey+Bada$$/_/ROCKABYE+BABY+(feat.+ScHoolboy+Q)
#> 46                             https://www.last.fm/music/Joey+Bada$$/_/FOR+MY+PEOPLE
#> 47                    https://www.last.fm/music/Joey+Bada$$/_/GOOD+MORNING+AMERIKKKA
#> 48                                    https://www.last.fm/music/Oddisee/_/Want+to+Be
#> 49                                   https://www.last.fm/music/Oddisee/_/Like+Really
#> 50                                          https://www.last.fm/music/Oddisee/_/NNGE
#>                                                                                                                                                                                                                                                                                                                                                                    image
#> 1  https://lastfm-img2.akamaized.net/i/u/34s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/64s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/174s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/300x300/86ca4b8daad31dbdbb93ef988d8b1d27.png, small, medium, large, extralarge
#> 2  https://lastfm-img2.akamaized.net/i/u/34s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/64s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/174s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/300x300/86ca4b8daad31dbdbb93ef988d8b1d27.png, small, medium, large, extralarge
#> 3  https://lastfm-img2.akamaized.net/i/u/34s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/64s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/174s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/300x300/86ca4b8daad31dbdbb93ef988d8b1d27.png, small, medium, large, extralarge
#> 4  https://lastfm-img2.akamaized.net/i/u/34s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/64s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/174s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/300x300/86ca4b8daad31dbdbb93ef988d8b1d27.png, small, medium, large, extralarge
#> 5  https://lastfm-img2.akamaized.net/i/u/34s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/64s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/174s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/300x300/1337b5bc360ae094c09a91c93839c6c1.png, small, medium, large, extralarge
#> 6  https://lastfm-img2.akamaized.net/i/u/34s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/64s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/174s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/300x300/1337b5bc360ae094c09a91c93839c6c1.png, small, medium, large, extralarge
#> 7  https://lastfm-img2.akamaized.net/i/u/34s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/64s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/174s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/300x300/1337b5bc360ae094c09a91c93839c6c1.png, small, medium, large, extralarge
#> 8  https://lastfm-img2.akamaized.net/i/u/34s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/64s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/174s/1337b5bc360ae094c09a91c93839c6c1.png, https://lastfm-img2.akamaized.net/i/u/300x300/1337b5bc360ae094c09a91c93839c6c1.png, small, medium, large, extralarge
#> 9  https://lastfm-img2.akamaized.net/i/u/34s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/64s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/174s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/300x300/d17d2400932bddb0a5fed85ab4d8d6e7.png, small, medium, large, extralarge
#> 10 https://lastfm-img2.akamaized.net/i/u/34s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/64s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/174s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/300x300/d17d2400932bddb0a5fed85ab4d8d6e7.png, small, medium, large, extralarge
#> 11 https://lastfm-img2.akamaized.net/i/u/34s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/64s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/174s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/300x300/d17d2400932bddb0a5fed85ab4d8d6e7.png, small, medium, large, extralarge
#> 12 https://lastfm-img2.akamaized.net/i/u/34s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/64s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/174s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/300x300/d17d2400932bddb0a5fed85ab4d8d6e7.png, small, medium, large, extralarge
#> 13 https://lastfm-img2.akamaized.net/i/u/34s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/64s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/174s/d17d2400932bddb0a5fed85ab4d8d6e7.png, https://lastfm-img2.akamaized.net/i/u/300x300/d17d2400932bddb0a5fed85ab4d8d6e7.png, small, medium, large, extralarge
#> 14 https://lastfm-img2.akamaized.net/i/u/34s/f0bfbdd6011ed7863d01ddb923c32930.png, https://lastfm-img2.akamaized.net/i/u/64s/f0bfbdd6011ed7863d01ddb923c32930.png, https://lastfm-img2.akamaized.net/i/u/174s/f0bfbdd6011ed7863d01ddb923c32930.png, https://lastfm-img2.akamaized.net/i/u/300x300/f0bfbdd6011ed7863d01ddb923c32930.png, small, medium, large, extralarge
#> 15 https://lastfm-img2.akamaized.net/i/u/34s/c6a5633c314748bfc849e3263640c284.png, https://lastfm-img2.akamaized.net/i/u/64s/c6a5633c314748bfc849e3263640c284.png, https://lastfm-img2.akamaized.net/i/u/174s/c6a5633c314748bfc849e3263640c284.png, https://lastfm-img2.akamaized.net/i/u/300x300/c6a5633c314748bfc849e3263640c284.png, small, medium, large, extralarge
#> 16 https://lastfm-img2.akamaized.net/i/u/34s/f0bfbdd6011ed7863d01ddb923c32930.png, https://lastfm-img2.akamaized.net/i/u/64s/f0bfbdd6011ed7863d01ddb923c32930.png, https://lastfm-img2.akamaized.net/i/u/174s/f0bfbdd6011ed7863d01ddb923c32930.png, https://lastfm-img2.akamaized.net/i/u/300x300/f0bfbdd6011ed7863d01ddb923c32930.png, small, medium, large, extralarge
#> 17 https://lastfm-img2.akamaized.net/i/u/34s/5cc29a39e9394eb8ce94bbc54dd45997.png, https://lastfm-img2.akamaized.net/i/u/64s/5cc29a39e9394eb8ce94bbc54dd45997.png, https://lastfm-img2.akamaized.net/i/u/174s/5cc29a39e9394eb8ce94bbc54dd45997.png, https://lastfm-img2.akamaized.net/i/u/300x300/5cc29a39e9394eb8ce94bbc54dd45997.png, small, medium, large, extralarge
#> 18 https://lastfm-img2.akamaized.net/i/u/34s/625d8735f4f0bf1cbf3c779fd4746bdc.png, https://lastfm-img2.akamaized.net/i/u/64s/625d8735f4f0bf1cbf3c779fd4746bdc.png, https://lastfm-img2.akamaized.net/i/u/174s/625d8735f4f0bf1cbf3c779fd4746bdc.png, https://lastfm-img2.akamaized.net/i/u/300x300/625d8735f4f0bf1cbf3c779fd4746bdc.png, small, medium, large, extralarge
#> 19 https://lastfm-img2.akamaized.net/i/u/34s/625d8735f4f0bf1cbf3c779fd4746bdc.png, https://lastfm-img2.akamaized.net/i/u/64s/625d8735f4f0bf1cbf3c779fd4746bdc.png, https://lastfm-img2.akamaized.net/i/u/174s/625d8735f4f0bf1cbf3c779fd4746bdc.png, https://lastfm-img2.akamaized.net/i/u/300x300/625d8735f4f0bf1cbf3c779fd4746bdc.png, small, medium, large, extralarge
#> 20 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 21 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 22 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 23 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 24 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 25 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 26 https://lastfm-img2.akamaized.net/i/u/34s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/64s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/174s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/300x300/d4f6f6f537a28b6b15f793515b95452b.png, small, medium, large, extralarge
#> 27 https://lastfm-img2.akamaized.net/i/u/34s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/64s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/174s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/300x300/d4f6f6f537a28b6b15f793515b95452b.png, small, medium, large, extralarge
#> 28 https://lastfm-img2.akamaized.net/i/u/34s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/64s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/174s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/300x300/d4f6f6f537a28b6b15f793515b95452b.png, small, medium, large, extralarge
#> 29 https://lastfm-img2.akamaized.net/i/u/34s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/64s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/174s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/300x300/d4f6f6f537a28b6b15f793515b95452b.png, small, medium, large, extralarge
#> 30 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 31 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 32 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 33 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 34 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 35 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 36 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 37 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 38 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 39 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 40 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 41 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 42 https://lastfm-img2.akamaized.net/i/u/34s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/64s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/174s/ba812293aae91094276ecf05ebd75313.png, https://lastfm-img2.akamaized.net/i/u/300x300/ba812293aae91094276ecf05ebd75313.png, small, medium, large, extralarge
#> 43 https://lastfm-img2.akamaized.net/i/u/34s/c4708870fe25430fc3dbf679787192cf.png, https://lastfm-img2.akamaized.net/i/u/64s/c4708870fe25430fc3dbf679787192cf.png, https://lastfm-img2.akamaized.net/i/u/174s/c4708870fe25430fc3dbf679787192cf.png, https://lastfm-img2.akamaized.net/i/u/300x300/c4708870fe25430fc3dbf679787192cf.png, small, medium, large, extralarge
#> 44 https://lastfm-img2.akamaized.net/i/u/34s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/64s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/174s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/300x300/86ca4b8daad31dbdbb93ef988d8b1d27.png, small, medium, large, extralarge
#> 45 https://lastfm-img2.akamaized.net/i/u/34s/71a28691e3ede34259b7ada3819aa7e7.png, https://lastfm-img2.akamaized.net/i/u/64s/71a28691e3ede34259b7ada3819aa7e7.png, https://lastfm-img2.akamaized.net/i/u/174s/71a28691e3ede34259b7ada3819aa7e7.png, https://lastfm-img2.akamaized.net/i/u/300x300/71a28691e3ede34259b7ada3819aa7e7.png, small, medium, large, extralarge
#> 46 https://lastfm-img2.akamaized.net/i/u/34s/71a28691e3ede34259b7ada3819aa7e7.png, https://lastfm-img2.akamaized.net/i/u/64s/71a28691e3ede34259b7ada3819aa7e7.png, https://lastfm-img2.akamaized.net/i/u/174s/71a28691e3ede34259b7ada3819aa7e7.png, https://lastfm-img2.akamaized.net/i/u/300x300/71a28691e3ede34259b7ada3819aa7e7.png, small, medium, large, extralarge
#> 47 https://lastfm-img2.akamaized.net/i/u/34s/71a28691e3ede34259b7ada3819aa7e7.png, https://lastfm-img2.akamaized.net/i/u/64s/71a28691e3ede34259b7ada3819aa7e7.png, https://lastfm-img2.akamaized.net/i/u/174s/71a28691e3ede34259b7ada3819aa7e7.png, https://lastfm-img2.akamaized.net/i/u/300x300/71a28691e3ede34259b7ada3819aa7e7.png, small, medium, large, extralarge
#> 48 https://lastfm-img2.akamaized.net/i/u/34s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/64s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/174s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/300x300/86ca4b8daad31dbdbb93ef988d8b1d27.png, small, medium, large, extralarge
#> 49 https://lastfm-img2.akamaized.net/i/u/34s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/64s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/174s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/300x300/86ca4b8daad31dbdbb93ef988d8b1d27.png, small, medium, large, extralarge
#> 50 https://lastfm-img2.akamaized.net/i/u/34s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/64s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/174s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/300x300/86ca4b8daad31dbdbb93ef988d8b1d27.png, small, medium, large, extralarge
#>      date.uts         date.#text
#> 1  1495063889 17 May 2017, 23:31
#> 2  1495063653 17 May 2017, 23:27
#> 3  1495063427 17 May 2017, 23:23
#> 4  1495063224 17 May 2017, 23:20
#> 5  1495062841 17 May 2017, 23:14
#> 6  1495062595 17 May 2017, 23:09
#> 7  1495062417 17 May 2017, 23:06
#> 8  1495062221 17 May 2017, 23:03
#> 9  1495062037 17 May 2017, 23:00
#> 10 1495061855 17 May 2017, 22:57
#> 11 1495061661 17 May 2017, 22:54
#> 12 1495061482 17 May 2017, 22:51
#> 13 1495061298 17 May 2017, 22:48
#> 14 1495061079 17 May 2017, 22:44
#> 15 1495060923 17 May 2017, 22:42
#> 16 1495060757 17 May 2017, 22:39
#> 17 1495060568 17 May 2017, 22:36
#> 18 1495060311 17 May 2017, 22:31
#> 19 1495060094 17 May 2017, 22:28
#> 20 1495059832 17 May 2017, 22:23
#> 21 1495059575 17 May 2017, 22:19
#> 22 1495059289 17 May 2017, 22:14
#> 23 1495058961 17 May 2017, 22:09
#> 24 1495058690 17 May 2017, 22:04
#> 25 1495058415 17 May 2017, 22:00
#> 26 1494969108 16 May 2017, 21:11
#> 27 1494968893 16 May 2017, 21:08
#> 28 1494968511 16 May 2017, 21:01
#> 29 1494968258 16 May 2017, 20:57
#> 30 1494885587 15 May 2017, 21:59
#> 31 1494885301 15 May 2017, 21:55
#> 32 1494884973 15 May 2017, 21:49
#> 33 1494884702 15 May 2017, 21:45
#> 34 1494873091 15 May 2017, 18:31
#> 35 1494872669 15 May 2017, 18:24
#> 36 1494872463 15 May 2017, 18:21
#> 37 1494872200 15 May 2017, 18:16
#> 38 1494871962 15 May 2017, 18:12
#> 39 1494871705 15 May 2017, 18:08
#> 40 1494871446 15 May 2017, 18:04
#> 41 1494871189 15 May 2017, 17:59
#> 42 1494870905 15 May 2017, 17:55
#> 43 1494436153 10 May 2017, 17:09
#> 44 1494435919 10 May 2017, 17:05
#> 45 1494435702 10 May 2017, 17:01
#> 46 1494435578 10 May 2017, 16:59
#> 47 1494435481 10 May 2017, 16:58
#> 48 1494264755 08 May 2017, 17:32
#> 49 1494264531 08 May 2017, 17:28
#> 50 1494264300 08 May 2017, 17:25
#> 
#> $`@attr`
#> $`@attr`$user
#> [1] "platyjus"
#> 
#> $`@attr`$page
#> [1] "1"
#> 
#> $`@attr`$perPage
#> [1] "50"
#> 
#> $`@attr`$totalPages
#> [1] "121"
#> 
#> $`@attr`$total
#> [1] "6010"
#> 
#>