Get the top albums listened to by a user.

Get the top albums listened to by a user. You can stipulate a time period. Sends the overall chart by default. Implementation of last.fm's user.getTopAlbums API method (http://www.last.fm/api/show/user.getTopAlbums).

user_getTopAlbums(user, period = NA, limit = NA, page = NA,
  api_key = lastkey)

Arguments

user
The user name to fetch top albums for.

period
overall | 7day | 1month | 3month | 6month | 12month - The time period over which to retrieve top albums for.

limit
The number of results to fetch per page. Defaults to 50.

page
The page number to fetch. Defaults to first page.

api_key
A Last.fm API key.

Value

A list of the top albums listened to by a user.

Examples

user_getTopAlbums("platyjus")

#> $album
#>                                                                     name
#> 1                                                                 Malibu
#> 2                                                            Coming Home
#> 3                                                     The Social Network
#> 4                                                          Coloring Book
#> 5                                                                At Swim
#> 6                                       Without My Enemy What Would I Do
#> 7                                                   Coming Home (Deluxe)
#> 8                                                            The Suffers
#> 9                                                        Made In Heights
#> 10                                                      The Sun's Tirade
#> 11                                                    Cleopatra (Deluxe)
#> 12                                                              Telefone
#> 13                                                        Hey Marseilles
#> 14                                                      Run the Jewels 2
#> 15 Sometimes I Sit and Think, And Sometimes I Just Sit (Special Edition)
#> 16                                                       Are You Serious
#> 17                                                        Summertime '06
#> 18                                                                  Takk
#> 19                                                        The Good Fight
#> 20                                                              B4.DA.$$
#> 21                                                 3001: A Laced Odyssey
#> 22                                       Déjenme Llorar (Deluxe Edition)
#> 23                                       MADE IN HEIGHTS (Instrumentals)
#> 24                                                                Malibu
#> 25                                                        Signs of Light
#> 26                                                              No No No
#> 27                                                            KONNICHIWA
#> 28                                                The Headspace Traveler
#> 29                        Gone Girl (Soundtrack from the Motion Picture)
#> 30                Game of Thrones: Season 6 (Music from the HBO® Series)
#> 31                   Sometimes I Sit and Think, and Sometimes I Just Sit
#> 32                                                        Absolute Loser
#> 33                                                       Blind and Brave
#> 34                                                     The Life Of Pablo
#> 35                                                    "Awaken, My Love!"
#> 36                                                             Cleopatra
#> 37                                                         Blank Face LP
#> 38                                                   To Pimp a Butterfly
#> 39                                                                Blonde
#> 40                                                                  Surf
#> 41                                                           The Iceberg
#> 42                                               Stillness In Wonderland
#> 43                                                 Echolocations: Canyon
#> 44                                                        Let's Be Still
#> 45                                                The Colour In Anything
#> 46                                                               Process
#> 47                                                       Bobby Tarantino
#> 48                                                            Section.80
#> 49                                                               Valtari
#> 50                                                 I Love You, Honeybear
#>    playcount                                 mbid
#> 1        142                                     
#> 2        133                                     
#> 3        114 998e28f9-ed94-4de1-af8e-8dc544c1ab31
#> 4        113                                     
#> 5        107                                     
#> 6        103                                     
#> 7         99                                     
#> 8         98                                     
#> 9         97 5f109edf-fc79-4547-919d-f9e8d4c73b1c
#> 10        95                                     
#> 11        83                                     
#> 12        76                                     
#> 13        75                                     
#> 14        60 8d5673a1-f435-4a02-9e32-8741397969c8
#> 15        60                                     
#> 16        58                                     
#> 17        53                                     
#> 18        52                                     
#> 19        52                                     
#> 20        49 8e5b5df1-0c5e-4b15-872d-d466a86c2709
#> 21        46                                     
#> 22        46                                     
#> 23        46                                     
#> 24        46                                     
#> 25        45                                     
#> 26        44                                     
#> 27        43                                     
#> 28        43                                     
#> 29        42                                     
#> 30        41                                     
#> 31        41                                     
#> 32        38                                     
#> 33        38                                     
#> 34        38                                     
#> 35        37                                     
#> 36        37                                     
#> 37        36                                     
#> 38        35                                     
#> 39        34                                     
#> 40        34                                     
#> 41        32                                     
#> 42        31                                     
#> 43        30                                     
#> 44        30 0a625224-3978-47fc-a22d-91936994e891
#> 45        30                                     
#> 46        29                                     
#> 47        28                                     
#> 48        28 d0b24c41-8562-47fb-bfe7-5f03397c41c7
#> 49        28 173c790a-264c-4134-9ffb-9b7aa78da6f5
#> 50        27 5bad490b-2939-4955-955b-9280cf616473
#>                                                                                                                 url
#> 1                                                                    https://www.last.fm/music/Anderson+Paak/Malibu
#> 2                                                                https://www.last.fm/music/Leon+Bridges/Coming+Home
#> 3                                        https://www.last.fm/music/Trent+Reznor+and+Atticus+Ross/The+Social+Network
#> 4                                                         https://www.last.fm/music/Chance+the+Rapper/Coloring+Book
#> 5                                                                   https://www.last.fm/music/Lisa+Hannigan/At+Swim
#> 6                          https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/Without+My+Enemy+What+Would+I+Do
#> 7                                                       https://www.last.fm/music/Leon+Bridges/Coming+Home+(Deluxe)
#> 8                                                                 https://www.last.fm/music/The+Suffers/The+Suffers
#> 9                                           https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/Made+In+Heights
#> 10                                                       https://www.last.fm/music/Isaiah+Rashad/The+Sun%27s+Tirade
#> 11                                                       https://www.last.fm/music/The+Lumineers/Cleopatra+(Deluxe)
#> 12                                                                        https://www.last.fm/music/NoName/Telefone
#> 13                                                          https://www.last.fm/music/Hey+Marseilles/Hey+Marseilles
#> 14                                                        https://www.last.fm/music/Run+the+Jewels/Run+the+Jewels+2
#> 15 https://www.last.fm/music/Courtney+Barnett/Sometimes+I+Sit+and+Think,+And+Sometimes+I+Just+Sit+(Special+Edition)
#> 16                                                            https://www.last.fm/music/Andrew+Bird/Are+You+Serious
#> 17                                                         https://www.last.fm/music/Vince+Staples/Summertime+%2706
#> 18                                                                    https://www.last.fm/music/Sigur+R%C3%B3s/Takk
#> 19                                                                 https://www.last.fm/music/Oddisee/The+Good+Fight
#> 20                                                                   https://www.last.fm/music/Joey+Bada$$/B4.DA.$$
#> 21                                                 https://www.last.fm/music/Flatbush+ZOMBiES/3001:+A+Laced+Odyssey
#> 22                                    https://www.last.fm/music/Carla+Morrison/D%C3%A9jenme+Llorar+(Deluxe+Edition)
#> 23                          https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/MADE+IN+HEIGHTS+(Instrumentals)
#> 24                                                                  https://www.last.fm/music/Anderson+.Paak/Malibu
#> 25                                                  https://www.last.fm/music/The+Head+and+the+Heart/Signs+of+Light
#> 26                                                                        https://www.last.fm/music/Beirut/No+No+No
#> 27                                                                      https://www.last.fm/music/Skepta/KONNICHIWA
#> 28                                                             https://www.last.fm/music/Sol/The+Headspace+Traveler
#> 29           https://www.last.fm/music/Trent+Reznor+and+Atticus+Ross/Gone+Girl+(Soundtrack+from+the+Motion+Picture)
#> 30              https://www.last.fm/music/Ramin+Djawadi/Game+of+Thrones:+Season+6+(Music+from+the+HBO%C2%AE+Series)
#> 31                   https://www.last.fm/music/Courtney+Barnett/Sometimes+I+Sit+and+Think,+and+Sometimes+I+Just+Sit
#> 32                                                              https://www.last.fm/music/Fruit+Bats/Absolute+Loser
#> 33                                                         https://www.last.fm/music/The+Wild+Reeds/Blind+and+Brave
#> 34                                                           https://www.last.fm/music/Kanye+West/The+Life+Of+Pablo
#> 35                                              https://www.last.fm/music/Childish+Gambino/%22Awaken,+My+Love%21%22
#> 36                                                                https://www.last.fm/music/The+Lumineers/Cleopatra
#> 37                                                              https://www.last.fm/music/ScHoolboy+Q/Blank+Face+LP
#> 38                                                     https://www.last.fm/music/Kendrick+Lamar/To+Pimp+a+Butterfly
#> 39                                                                     https://www.last.fm/music/Frank+Ocean/Blonde
#> 40                                            https://www.last.fm/music/Donnie+Trumpet+&+The+Social+Experiment/Surf
#> 41                                                                    https://www.last.fm/music/Oddisee/The+Iceberg
#> 42                                                    https://www.last.fm/music/Little+Simz/Stillness+In+Wonderland
#> 43                                                      https://www.last.fm/music/Andrew+Bird/Echolocations:+Canyon
#> 44                                                https://www.last.fm/music/The+Head+and+the+Heart/Let%27s+Be+Still
#> 45                                                     https://www.last.fm/music/James+Blake/The+Colour+In+Anything
#> 46                                                                         https://www.last.fm/music/Sampha/Process
#> 47                                                                  https://www.last.fm/music/Logic/Bobby+Tarantino
#> 48                                                              https://www.last.fm/music/Kendrick+Lamar/Section.80
#> 49                                                                 https://www.last.fm/music/Sigur+R%C3%B3s/Valtari
#> 50                                                https://www.last.fm/music/Father+John+Misty/I+Love+You,+Honeybear
#>                               artist.name                          artist.mbid
#> 1                           Anderson Paak d02dd67e-f655-4600-bc47-f789f59e7367
#> 2                            Leon Bridges                                     
#> 3           Trent Reznor and Atticus Ross bc880664-6751-4998-984d-ebdbff15d229
#> 4                       Chance the Rapper 373a4c98-a46b-48e4-86ec-f6ca65b4f438
#> 5                           Lisa Hannigan 8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#> 6                   ::M∆DE::IN::HEIGHTS:: 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 7                            Leon Bridges                                     
#> 8                             The Suffers bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 9                   ::M∆DE::IN::HEIGHTS:: 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 10                          Isaiah Rashad 37547347-ea75-45fa-9766-ff36b44d8147
#> 11                          The Lumineers bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 12                                 NoName 9393cc98-b014-493c-bdf1-a643a088f662
#> 13                         Hey Marseilles 4f0ca99b-0ab2-4861-b8ac-e6dadabff23e
#> 14                         Run the Jewels a80c5dc5-b12e-4667-9f5a-b568961f3839
#> 15                       Courtney Barnett 55111838-f001-494a-a1b5-9d818db85810
#> 16                            Andrew Bird 081b133e-ce74-42ba-92c1-c18234acb532
#> 17                          Vince Staples 78e854b8-9713-4ff2-9218-6b3784893bff
#> 18                              Sigur Rós f6f2326f-6b25-4170-b89d-e235b25508e8
#> 19                                Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 20                            Joey Bada$$ 9757ae7d-44b1-4d63-81ce-b463fe27b9ae
#> 21                       Flatbush ZOMBiES f59a5dac-939a-4f12-a35c-9a49676f6898
#> 22                         Carla Morrison 50ee3204-13fb-4ae3-b152-82e3fc845110
#> 23                  ::M∆DE::IN::HEIGHTS:: 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 24                         Anderson .Paak                                     
#> 25                 The Head and the Heart 98cb416e-5271-484c-ab3b-6bed6b4aae12
#> 26                                 Beirut 989276c4-d2e8-4599-a2a0-c9ebeba73104
#> 27                                 Skepta 38474d18-9204-4f60-a137-1fcc6750b869
#> 28                                    Sol f7ef4d71-9f9e-4b53-9f3a-0fa557167c94
#> 29          Trent Reznor and Atticus Ross bc880664-6751-4998-984d-ebdbff15d229
#> 30                          Ramin Djawadi f089f636-6b85-4e0e-a192-0d30ed2b44d1
#> 31                       Courtney Barnett 55111838-f001-494a-a1b5-9d818db85810
#> 32                             Fruit Bats e3f139c7-11bc-4803-82be-2bac85ca3c61
#> 33                         The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 34                             Kanye West 164f0d73-1234-4e2c-8743-d77bf2191051
#> 35                       Childish Gambino 7fb57fba-a6ef-44c2-abab-2fa3bdee607e
#> 36                          The Lumineers bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 37                            ScHoolboy Q bce6d667-cde8-485e-b078-c0a05adea36d
#> 38                         Kendrick Lamar 381086ea-f511-4aba-bdf9-71c753dc5077
#> 39                            Frank Ocean e520459c-dff4-491d-a6e4-c97be35e0044
#> 40 Donnie Trumpet & The Social Experiment                                     
#> 41                                Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 42                            Little Simz 3cdb40fe-a63e-4bb9-b40d-17cda5f50979
#> 43                            Andrew Bird 081b133e-ce74-42ba-92c1-c18234acb532
#> 44                 The Head and the Heart 98cb416e-5271-484c-ab3b-6bed6b4aae12
#> 45                            James Blake 8dc08b1f-e393-4f85-a5dd-300f7693a8b8
#> 46                                 Sampha 130c2184-14f2-4180-a216-04b5a333be57
#> 47                                  Logic b2f45ca6-2c92-438c-b037-ad4d8b92e90e
#> 48                         Kendrick Lamar 381086ea-f511-4aba-bdf9-71c753dc5077
#> 49                              Sigur Rós f6f2326f-6b25-4170-b89d-e235b25508e8
#> 50                      Father John Misty 172053e5-a7fb-4548-a64d-c53858b5e775
#>                                                          artist.url
#> 1                           https://www.last.fm/music/Anderson+Paak
#> 2                            https://www.last.fm/music/Leon+Bridges
#> 3           https://www.last.fm/music/Trent+Reznor+and+Atticus+Ross
#> 4                       https://www.last.fm/music/Chance+the+Rapper
#> 5                           https://www.last.fm/music/Lisa+Hannigan
#> 6           https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 7                            https://www.last.fm/music/Leon+Bridges
#> 8                             https://www.last.fm/music/The+Suffers
#> 9           https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 10                          https://www.last.fm/music/Isaiah+Rashad
#> 11                          https://www.last.fm/music/The+Lumineers
#> 12                                 https://www.last.fm/music/NoName
#> 13                         https://www.last.fm/music/Hey+Marseilles
#> 14                         https://www.last.fm/music/Run+the+Jewels
#> 15                       https://www.last.fm/music/Courtney+Barnett
#> 16                            https://www.last.fm/music/Andrew+Bird
#> 17                          https://www.last.fm/music/Vince+Staples
#> 18                         https://www.last.fm/music/Sigur+R%C3%B3s
#> 19                                https://www.last.fm/music/Oddisee
#> 20                            https://www.last.fm/music/Joey+Bada$$
#> 21                       https://www.last.fm/music/Flatbush+ZOMBiES
#> 22                         https://www.last.fm/music/Carla+Morrison
#> 23          https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 24                         https://www.last.fm/music/Anderson+.Paak
#> 25                 https://www.last.fm/music/The+Head+and+the+Heart
#> 26                                 https://www.last.fm/music/Beirut
#> 27                                 https://www.last.fm/music/Skepta
#> 28                                    https://www.last.fm/music/Sol
#> 29          https://www.last.fm/music/Trent+Reznor+and+Atticus+Ross
#> 30                          https://www.last.fm/music/Ramin+Djawadi
#> 31                       https://www.last.fm/music/Courtney+Barnett
#> 32                             https://www.last.fm/music/Fruit+Bats
#> 33                         https://www.last.fm/music/The+Wild+Reeds
#> 34                             https://www.last.fm/music/Kanye+West
#> 35                       https://www.last.fm/music/Childish+Gambino
#> 36                          https://www.last.fm/music/The+Lumineers
#> 37                            https://www.last.fm/music/ScHoolboy+Q
#> 38                         https://www.last.fm/music/Kendrick+Lamar
#> 39                            https://www.last.fm/music/Frank+Ocean
#> 40 https://www.last.fm/music/Donnie+Trumpet+&+The+Social+Experiment
#> 41                                https://www.last.fm/music/Oddisee
#> 42                            https://www.last.fm/music/Little+Simz
#> 43                            https://www.last.fm/music/Andrew+Bird
#> 44                 https://www.last.fm/music/The+Head+and+the+Heart
#> 45                            https://www.last.fm/music/James+Blake
#> 46                                 https://www.last.fm/music/Sampha
#> 47                                  https://www.last.fm/music/Logic
#> 48                         https://www.last.fm/music/Kendrick+Lamar
#> 49                         https://www.last.fm/music/Sigur+R%C3%B3s
#> 50                      https://www.last.fm/music/Father+John+Misty
#>                                                                                                                                                                                                                                                                                                                                                                    image
#> 1  https://lastfm-img2.akamaized.net/i/u/34s/0b614d44a076dd436f9031b635cd0e88.png, https://lastfm-img2.akamaized.net/i/u/64s/0b614d44a076dd436f9031b635cd0e88.png, https://lastfm-img2.akamaized.net/i/u/174s/0b614d44a076dd436f9031b635cd0e88.png, https://lastfm-img2.akamaized.net/i/u/300x300/0b614d44a076dd436f9031b635cd0e88.png, small, medium, large, extralarge
#> 2  https://lastfm-img2.akamaized.net/i/u/34s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/64s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/174s/48f4e73dbe0f9960572256ff0e5ea377.png, https://lastfm-img2.akamaized.net/i/u/300x300/48f4e73dbe0f9960572256ff0e5ea377.png, small, medium, large, extralarge
#> 3  https://lastfm-img2.akamaized.net/i/u/34s/44d470d009d24d658d0d2b2bcaa2bcd2.png, https://lastfm-img2.akamaized.net/i/u/64s/44d470d009d24d658d0d2b2bcaa2bcd2.png, https://lastfm-img2.akamaized.net/i/u/174s/44d470d009d24d658d0d2b2bcaa2bcd2.png, https://lastfm-img2.akamaized.net/i/u/300x300/44d470d009d24d658d0d2b2bcaa2bcd2.png, small, medium, large, extralarge
#> 4  https://lastfm-img2.akamaized.net/i/u/34s/15b37232ffad3490370868c68f76c730.png, https://lastfm-img2.akamaized.net/i/u/64s/15b37232ffad3490370868c68f76c730.png, https://lastfm-img2.akamaized.net/i/u/174s/15b37232ffad3490370868c68f76c730.png, https://lastfm-img2.akamaized.net/i/u/300x300/15b37232ffad3490370868c68f76c730.png, small, medium, large, extralarge
#> 5  https://lastfm-img2.akamaized.net/i/u/34s/c7f2bfa15c6cc28fce93bbdb64241d28.png, https://lastfm-img2.akamaized.net/i/u/64s/c7f2bfa15c6cc28fce93bbdb64241d28.png, https://lastfm-img2.akamaized.net/i/u/174s/c7f2bfa15c6cc28fce93bbdb64241d28.png, https://lastfm-img2.akamaized.net/i/u/300x300/c7f2bfa15c6cc28fce93bbdb64241d28.png, small, medium, large, extralarge
#> 6  https://lastfm-img2.akamaized.net/i/u/34s/80ddc5637b3dac96fae8756856002f8a.png, https://lastfm-img2.akamaized.net/i/u/64s/80ddc5637b3dac96fae8756856002f8a.png, https://lastfm-img2.akamaized.net/i/u/174s/80ddc5637b3dac96fae8756856002f8a.png, https://lastfm-img2.akamaized.net/i/u/300x300/80ddc5637b3dac96fae8756856002f8a.png, small, medium, large, extralarge
#> 7  https://lastfm-img2.akamaized.net/i/u/34s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/64s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/174s/f9c52cca9148a38dc15900c755b2b9f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/f9c52cca9148a38dc15900c755b2b9f9.png, small, medium, large, extralarge
#> 8  https://lastfm-img2.akamaized.net/i/u/34s/c2ef31178cb1a3f654fde5e7a4f51b19.png, https://lastfm-img2.akamaized.net/i/u/64s/c2ef31178cb1a3f654fde5e7a4f51b19.png, https://lastfm-img2.akamaized.net/i/u/174s/c2ef31178cb1a3f654fde5e7a4f51b19.png, https://lastfm-img2.akamaized.net/i/u/300x300/c2ef31178cb1a3f654fde5e7a4f51b19.png, small, medium, large, extralarge
#> 9  https://lastfm-img2.akamaized.net/i/u/34s/7a2566fdd71a439dc3476b172398f8d6.png, https://lastfm-img2.akamaized.net/i/u/64s/7a2566fdd71a439dc3476b172398f8d6.png, https://lastfm-img2.akamaized.net/i/u/174s/7a2566fdd71a439dc3476b172398f8d6.png, https://lastfm-img2.akamaized.net/i/u/300x300/7a2566fdd71a439dc3476b172398f8d6.png, small, medium, large, extralarge
#> 10 https://lastfm-img2.akamaized.net/i/u/34s/56f99650b0478d05f79ac8958bcf37fa.png, https://lastfm-img2.akamaized.net/i/u/64s/56f99650b0478d05f79ac8958bcf37fa.png, https://lastfm-img2.akamaized.net/i/u/174s/56f99650b0478d05f79ac8958bcf37fa.png, https://lastfm-img2.akamaized.net/i/u/300x300/56f99650b0478d05f79ac8958bcf37fa.png, small, medium, large, extralarge
#> 11 https://lastfm-img2.akamaized.net/i/u/34s/acff6975233a56254751aae6fa23d36f.png, https://lastfm-img2.akamaized.net/i/u/64s/acff6975233a56254751aae6fa23d36f.png, https://lastfm-img2.akamaized.net/i/u/174s/acff6975233a56254751aae6fa23d36f.png, https://lastfm-img2.akamaized.net/i/u/300x300/acff6975233a56254751aae6fa23d36f.png, small, medium, large, extralarge
#> 12 https://lastfm-img2.akamaized.net/i/u/34s/9a553cad633e13c57c2174ee202462e2.png, https://lastfm-img2.akamaized.net/i/u/64s/9a553cad633e13c57c2174ee202462e2.png, https://lastfm-img2.akamaized.net/i/u/174s/9a553cad633e13c57c2174ee202462e2.png, https://lastfm-img2.akamaized.net/i/u/300x300/9a553cad633e13c57c2174ee202462e2.png, small, medium, large, extralarge
#> 13 https://lastfm-img2.akamaized.net/i/u/34s/176ec2a895494dee5291ad0507c1ed82.png, https://lastfm-img2.akamaized.net/i/u/64s/176ec2a895494dee5291ad0507c1ed82.png, https://lastfm-img2.akamaized.net/i/u/174s/176ec2a895494dee5291ad0507c1ed82.png, https://lastfm-img2.akamaized.net/i/u/300x300/176ec2a895494dee5291ad0507c1ed82.png, small, medium, large, extralarge
#> 14 https://lastfm-img2.akamaized.net/i/u/34s/5591506667b045aac54e13f41def97b2.png, https://lastfm-img2.akamaized.net/i/u/64s/5591506667b045aac54e13f41def97b2.png, https://lastfm-img2.akamaized.net/i/u/174s/5591506667b045aac54e13f41def97b2.png, https://lastfm-img2.akamaized.net/i/u/300x300/5591506667b045aac54e13f41def97b2.png, small, medium, large, extralarge
#> 15 https://lastfm-img2.akamaized.net/i/u/34s/94b8c41d192e91dbd773a4c979155acd.png, https://lastfm-img2.akamaized.net/i/u/64s/94b8c41d192e91dbd773a4c979155acd.png, https://lastfm-img2.akamaized.net/i/u/174s/94b8c41d192e91dbd773a4c979155acd.png, https://lastfm-img2.akamaized.net/i/u/300x300/94b8c41d192e91dbd773a4c979155acd.png, small, medium, large, extralarge
#> 16 https://lastfm-img2.akamaized.net/i/u/34s/b7d3ca86a4a3f91b9ce31faed5c00c91.png, https://lastfm-img2.akamaized.net/i/u/64s/b7d3ca86a4a3f91b9ce31faed5c00c91.png, https://lastfm-img2.akamaized.net/i/u/174s/b7d3ca86a4a3f91b9ce31faed5c00c91.png, https://lastfm-img2.akamaized.net/i/u/300x300/b7d3ca86a4a3f91b9ce31faed5c00c91.png, small, medium, large, extralarge
#> 17 https://lastfm-img2.akamaized.net/i/u/34s/748b3c68f962a59e20b1f2f3de792de4.png, https://lastfm-img2.akamaized.net/i/u/64s/748b3c68f962a59e20b1f2f3de792de4.png, https://lastfm-img2.akamaized.net/i/u/174s/748b3c68f962a59e20b1f2f3de792de4.png, https://lastfm-img2.akamaized.net/i/u/300x300/748b3c68f962a59e20b1f2f3de792de4.png, small, medium, large, extralarge
#> 18 https://lastfm-img2.akamaized.net/i/u/34s/3f7f29a985ef48699823eaecb39846b0.png, https://lastfm-img2.akamaized.net/i/u/64s/3f7f29a985ef48699823eaecb39846b0.png, https://lastfm-img2.akamaized.net/i/u/174s/3f7f29a985ef48699823eaecb39846b0.png, https://lastfm-img2.akamaized.net/i/u/300x300/3f7f29a985ef48699823eaecb39846b0.png, small, medium, large, extralarge
#> 19 https://lastfm-img2.akamaized.net/i/u/34s/c4708870fe25430fc3dbf679787192cf.png, https://lastfm-img2.akamaized.net/i/u/64s/c4708870fe25430fc3dbf679787192cf.png, https://lastfm-img2.akamaized.net/i/u/174s/c4708870fe25430fc3dbf679787192cf.png, https://lastfm-img2.akamaized.net/i/u/300x300/c4708870fe25430fc3dbf679787192cf.png, small, medium, large, extralarge
#> 20 https://lastfm-img2.akamaized.net/i/u/34s/f37e6bd19001468ac65cd867568cfb32.png, https://lastfm-img2.akamaized.net/i/u/64s/f37e6bd19001468ac65cd867568cfb32.png, https://lastfm-img2.akamaized.net/i/u/174s/f37e6bd19001468ac65cd867568cfb32.png, https://lastfm-img2.akamaized.net/i/u/300x300/f37e6bd19001468ac65cd867568cfb32.png, small, medium, large, extralarge
#> 21 https://lastfm-img2.akamaized.net/i/u/34s/e5417e080af3b15e157282701b16f968.png, https://lastfm-img2.akamaized.net/i/u/64s/e5417e080af3b15e157282701b16f968.png, https://lastfm-img2.akamaized.net/i/u/174s/e5417e080af3b15e157282701b16f968.png, https://lastfm-img2.akamaized.net/i/u/300x300/e5417e080af3b15e157282701b16f968.png, small, medium, large, extralarge
#> 22 https://lastfm-img2.akamaized.net/i/u/34s/1afa8189218fa56eb2570cd97b62e9bf.png, https://lastfm-img2.akamaized.net/i/u/64s/1afa8189218fa56eb2570cd97b62e9bf.png, https://lastfm-img2.akamaized.net/i/u/174s/1afa8189218fa56eb2570cd97b62e9bf.png, https://lastfm-img2.akamaized.net/i/u/300x300/1afa8189218fa56eb2570cd97b62e9bf.png, small, medium, large, extralarge
#> 23                                                                                                                                                                                                                                                                                                                              , , , , small, medium, large, extralarge
#> 24 https://lastfm-img2.akamaized.net/i/u/34s/2ab43f30ecc90c0eb6662273a8e57d2c.png, https://lastfm-img2.akamaized.net/i/u/64s/2ab43f30ecc90c0eb6662273a8e57d2c.png, https://lastfm-img2.akamaized.net/i/u/174s/2ab43f30ecc90c0eb6662273a8e57d2c.png, https://lastfm-img2.akamaized.net/i/u/300x300/2ab43f30ecc90c0eb6662273a8e57d2c.png, small, medium, large, extralarge
#> 25 https://lastfm-img2.akamaized.net/i/u/34s/c3e44abf132245845795364096dd6dad.png, https://lastfm-img2.akamaized.net/i/u/64s/c3e44abf132245845795364096dd6dad.png, https://lastfm-img2.akamaized.net/i/u/174s/c3e44abf132245845795364096dd6dad.png, https://lastfm-img2.akamaized.net/i/u/300x300/c3e44abf132245845795364096dd6dad.png, small, medium, large, extralarge
#> 26 https://lastfm-img2.akamaized.net/i/u/34s/c5d9f20f4d0c8f2ff3ae969b149ea683.png, https://lastfm-img2.akamaized.net/i/u/64s/c5d9f20f4d0c8f2ff3ae969b149ea683.png, https://lastfm-img2.akamaized.net/i/u/174s/c5d9f20f4d0c8f2ff3ae969b149ea683.png, https://lastfm-img2.akamaized.net/i/u/300x300/c5d9f20f4d0c8f2ff3ae969b149ea683.png, small, medium, large, extralarge
#> 27 https://lastfm-img2.akamaized.net/i/u/34s/1f8aedcf3ceefbb0d71cf7673d7883d0.png, https://lastfm-img2.akamaized.net/i/u/64s/1f8aedcf3ceefbb0d71cf7673d7883d0.png, https://lastfm-img2.akamaized.net/i/u/174s/1f8aedcf3ceefbb0d71cf7673d7883d0.png, https://lastfm-img2.akamaized.net/i/u/300x300/1f8aedcf3ceefbb0d71cf7673d7883d0.png, small, medium, large, extralarge
#> 28 https://lastfm-img2.akamaized.net/i/u/34s/034963140f5a9d8acfa1bf715ef6b030.png, https://lastfm-img2.akamaized.net/i/u/64s/034963140f5a9d8acfa1bf715ef6b030.png, https://lastfm-img2.akamaized.net/i/u/174s/034963140f5a9d8acfa1bf715ef6b030.png, https://lastfm-img2.akamaized.net/i/u/300x300/034963140f5a9d8acfa1bf715ef6b030.png, small, medium, large, extralarge
#> 29 https://lastfm-img2.akamaized.net/i/u/34s/70a4cdf9be234d1fcdb72bedd9b5c12a.png, https://lastfm-img2.akamaized.net/i/u/64s/70a4cdf9be234d1fcdb72bedd9b5c12a.png, https://lastfm-img2.akamaized.net/i/u/174s/70a4cdf9be234d1fcdb72bedd9b5c12a.png, https://lastfm-img2.akamaized.net/i/u/300x300/70a4cdf9be234d1fcdb72bedd9b5c12a.png, small, medium, large, extralarge
#> 30 https://lastfm-img2.akamaized.net/i/u/34s/80fac939d75e430cee4023f010828b0e.png, https://lastfm-img2.akamaized.net/i/u/64s/80fac939d75e430cee4023f010828b0e.png, https://lastfm-img2.akamaized.net/i/u/174s/80fac939d75e430cee4023f010828b0e.png, https://lastfm-img2.akamaized.net/i/u/300x300/80fac939d75e430cee4023f010828b0e.png, small, medium, large, extralarge
#> 31 https://lastfm-img2.akamaized.net/i/u/34s/3d7b04cb4ee247c5c7cbc1fcce72b281.png, https://lastfm-img2.akamaized.net/i/u/64s/3d7b04cb4ee247c5c7cbc1fcce72b281.png, https://lastfm-img2.akamaized.net/i/u/174s/3d7b04cb4ee247c5c7cbc1fcce72b281.png, https://lastfm-img2.akamaized.net/i/u/300x300/3d7b04cb4ee247c5c7cbc1fcce72b281.png, small, medium, large, extralarge
#> 32 https://lastfm-img2.akamaized.net/i/u/34s/11ac82d97d5aabc35af53eaea9fb7411.png, https://lastfm-img2.akamaized.net/i/u/64s/11ac82d97d5aabc35af53eaea9fb7411.png, https://lastfm-img2.akamaized.net/i/u/174s/11ac82d97d5aabc35af53eaea9fb7411.png, https://lastfm-img2.akamaized.net/i/u/300x300/11ac82d97d5aabc35af53eaea9fb7411.png, small, medium, large, extralarge
#> 33 https://lastfm-img2.akamaized.net/i/u/34s/79fa5c8add0448e4c8586d426c8784a8.png, https://lastfm-img2.akamaized.net/i/u/64s/79fa5c8add0448e4c8586d426c8784a8.png, https://lastfm-img2.akamaized.net/i/u/174s/79fa5c8add0448e4c8586d426c8784a8.png, https://lastfm-img2.akamaized.net/i/u/300x300/79fa5c8add0448e4c8586d426c8784a8.png, small, medium, large, extralarge
#> 34 https://lastfm-img2.akamaized.net/i/u/34s/375a9688a2cb3d4e073bacf71735a63f.png, https://lastfm-img2.akamaized.net/i/u/64s/375a9688a2cb3d4e073bacf71735a63f.png, https://lastfm-img2.akamaized.net/i/u/174s/375a9688a2cb3d4e073bacf71735a63f.png, https://lastfm-img2.akamaized.net/i/u/300x300/375a9688a2cb3d4e073bacf71735a63f.png, small, medium, large, extralarge
#> 35 https://lastfm-img2.akamaized.net/i/u/34s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/64s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/174s/d4f6f6f537a28b6b15f793515b95452b.png, https://lastfm-img2.akamaized.net/i/u/300x300/d4f6f6f537a28b6b15f793515b95452b.png, small, medium, large, extralarge
#> 36 https://lastfm-img2.akamaized.net/i/u/34s/ea77f864eff0a4283fa30b8edade7ddb.png, https://lastfm-img2.akamaized.net/i/u/64s/ea77f864eff0a4283fa30b8edade7ddb.png, https://lastfm-img2.akamaized.net/i/u/174s/ea77f864eff0a4283fa30b8edade7ddb.png, https://lastfm-img2.akamaized.net/i/u/300x300/ea77f864eff0a4283fa30b8edade7ddb.png, small, medium, large, extralarge
#> 37 https://lastfm-img2.akamaized.net/i/u/34s/4d94b168b20acc03e8ade6fd75656ea5.png, https://lastfm-img2.akamaized.net/i/u/64s/4d94b168b20acc03e8ade6fd75656ea5.png, https://lastfm-img2.akamaized.net/i/u/174s/4d94b168b20acc03e8ade6fd75656ea5.png, https://lastfm-img2.akamaized.net/i/u/300x300/4d94b168b20acc03e8ade6fd75656ea5.png, small, medium, large, extralarge
#> 38 https://lastfm-img2.akamaized.net/i/u/34s/6c0c2b3d7f95361446aaaad9c64dc798.png, https://lastfm-img2.akamaized.net/i/u/64s/6c0c2b3d7f95361446aaaad9c64dc798.png, https://lastfm-img2.akamaized.net/i/u/174s/6c0c2b3d7f95361446aaaad9c64dc798.png, https://lastfm-img2.akamaized.net/i/u/300x300/6c0c2b3d7f95361446aaaad9c64dc798.png, small, medium, large, extralarge
#> 39 https://lastfm-img2.akamaized.net/i/u/34s/66792502e9fb82551612c02cfef19f10.png, https://lastfm-img2.akamaized.net/i/u/64s/66792502e9fb82551612c02cfef19f10.png, https://lastfm-img2.akamaized.net/i/u/174s/66792502e9fb82551612c02cfef19f10.png, https://lastfm-img2.akamaized.net/i/u/300x300/66792502e9fb82551612c02cfef19f10.png, small, medium, large, extralarge
#> 40 https://lastfm-img2.akamaized.net/i/u/34s/91e69adb142c8a7a7fd9ac2cb976af45.png, https://lastfm-img2.akamaized.net/i/u/64s/91e69adb142c8a7a7fd9ac2cb976af45.png, https://lastfm-img2.akamaized.net/i/u/174s/91e69adb142c8a7a7fd9ac2cb976af45.png, https://lastfm-img2.akamaized.net/i/u/300x300/91e69adb142c8a7a7fd9ac2cb976af45.png, small, medium, large, extralarge
#> 41 https://lastfm-img2.akamaized.net/i/u/34s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/64s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/174s/86ca4b8daad31dbdbb93ef988d8b1d27.png, https://lastfm-img2.akamaized.net/i/u/300x300/86ca4b8daad31dbdbb93ef988d8b1d27.png, small, medium, large, extralarge
#> 42 https://lastfm-img2.akamaized.net/i/u/34s/199e16c1a491396fb746867642da3e9f.png, https://lastfm-img2.akamaized.net/i/u/64s/199e16c1a491396fb746867642da3e9f.png, https://lastfm-img2.akamaized.net/i/u/174s/199e16c1a491396fb746867642da3e9f.png, https://lastfm-img2.akamaized.net/i/u/300x300/199e16c1a491396fb746867642da3e9f.png, small, medium, large, extralarge
#> 43 https://lastfm-img2.akamaized.net/i/u/34s/929ad8ab026d43fbc3fe547d8e63e7f7.png, https://lastfm-img2.akamaized.net/i/u/64s/929ad8ab026d43fbc3fe547d8e63e7f7.png, https://lastfm-img2.akamaized.net/i/u/174s/929ad8ab026d43fbc3fe547d8e63e7f7.png, https://lastfm-img2.akamaized.net/i/u/300x300/929ad8ab026d43fbc3fe547d8e63e7f7.png, small, medium, large, extralarge
#> 44 https://lastfm-img2.akamaized.net/i/u/34s/1366d9c6e5bc4803ac9e266c623aee35.png, https://lastfm-img2.akamaized.net/i/u/64s/1366d9c6e5bc4803ac9e266c623aee35.png, https://lastfm-img2.akamaized.net/i/u/174s/1366d9c6e5bc4803ac9e266c623aee35.png, https://lastfm-img2.akamaized.net/i/u/300x300/1366d9c6e5bc4803ac9e266c623aee35.png, small, medium, large, extralarge
#> 45 https://lastfm-img2.akamaized.net/i/u/34s/7946f969614f46c40b124678d38dcff4.png, https://lastfm-img2.akamaized.net/i/u/64s/7946f969614f46c40b124678d38dcff4.png, https://lastfm-img2.akamaized.net/i/u/174s/7946f969614f46c40b124678d38dcff4.png, https://lastfm-img2.akamaized.net/i/u/300x300/7946f969614f46c40b124678d38dcff4.png, small, medium, large, extralarge
#> 46 https://lastfm-img2.akamaized.net/i/u/34s/7ad951be0a942eb6559c5502c356fbf2.png, https://lastfm-img2.akamaized.net/i/u/64s/7ad951be0a942eb6559c5502c356fbf2.png, https://lastfm-img2.akamaized.net/i/u/174s/7ad951be0a942eb6559c5502c356fbf2.png, https://lastfm-img2.akamaized.net/i/u/300x300/7ad951be0a942eb6559c5502c356fbf2.png, small, medium, large, extralarge
#> 47 https://lastfm-img2.akamaized.net/i/u/34s/7299006562a706e180e4b9b25732b885.png, https://lastfm-img2.akamaized.net/i/u/64s/7299006562a706e180e4b9b25732b885.png, https://lastfm-img2.akamaized.net/i/u/174s/7299006562a706e180e4b9b25732b885.png, https://lastfm-img2.akamaized.net/i/u/300x300/7299006562a706e180e4b9b25732b885.png, small, medium, large, extralarge
#> 48 https://lastfm-img2.akamaized.net/i/u/34s/4d140d8d8dde411e96d2c6168f4f083f.png, https://lastfm-img2.akamaized.net/i/u/64s/4d140d8d8dde411e96d2c6168f4f083f.png, https://lastfm-img2.akamaized.net/i/u/174s/4d140d8d8dde411e96d2c6168f4f083f.png, https://lastfm-img2.akamaized.net/i/u/300x300/4d140d8d8dde411e96d2c6168f4f083f.png, small, medium, large, extralarge
#> 49 https://lastfm-img2.akamaized.net/i/u/34s/6016c9614463408a9108f80d66051dc0.png, https://lastfm-img2.akamaized.net/i/u/64s/6016c9614463408a9108f80d66051dc0.png, https://lastfm-img2.akamaized.net/i/u/174s/6016c9614463408a9108f80d66051dc0.png, https://lastfm-img2.akamaized.net/i/u/300x300/6016c9614463408a9108f80d66051dc0.png, small, medium, large, extralarge
#> 50 https://lastfm-img2.akamaized.net/i/u/34s/9f3c07bb127efaf7ef5e6958a5ee0f59.png, https://lastfm-img2.akamaized.net/i/u/64s/9f3c07bb127efaf7ef5e6958a5ee0f59.png, https://lastfm-img2.akamaized.net/i/u/174s/9f3c07bb127efaf7ef5e6958a5ee0f59.png, https://lastfm-img2.akamaized.net/i/u/300x300/9f3c07bb127efaf7ef5e6958a5ee0f59.png, small, medium, large, extralarge
#>    rank
#> 1     1
#> 2     2
#> 3     3
#> 4     4
#> 5     5
#> 6     6
#> 7     7
#> 8     8
#> 9     9
#> 10   10
#> 11   11
#> 12   12
#> 13   13
#> 14   14
#> 15   15
#> 16   16
#> 17   17
#> 18   18
#> 19   19
#> 20   20
#> 21   21
#> 22   22
#> 23   23
#> 24   24
#> 25   25
#> 26   26
#> 27   27
#> 28   28
#> 29   29
#> 30   30
#> 31   31
#> 32   32
#> 33   33
#> 34   34
#> 35   35
#> 36   36
#> 37   37
#> 38   38
#> 39   39
#> 40   40
#> 41   41
#> 42   42
#> 43   43
#> 44   44
#> 45   45
#> 46   46
#> 47   47
#> 48   48
#> 49   49
#> 50   50
#> 
#> $`@attr`
#> $`@attr`$user
#> [1] "platyjus"
#> 
#> $`@attr`$page
#> [1] "1"
#> 
#> $`@attr`$perPage
#> [1] "50"
#> 
#> $`@attr`$totalPages
#> [1] "21"
#> 
#> $`@attr`$total
#> [1] "1013"
#> 
#>