Get the top artists listened to by a user.

Get the top artists listened to by a user. You can stipulate a time period. Sends the overall chart by default. Implementation of last.fm's user.getTopArtists API method (http://www.last.fm/api/show/user.getTopArtists).

user_getTopArtists(user, period = NA, limit = NA, page = NA,
  api_key = lastkey)

Arguments

user
The user name to fetch top artists for.

period
overall | 7day | 1month | 3month | 6month | 12month - The time period over which to retrieve top artists for.

limit
The number of results to fetch per page. Defaults to 50.

page
The page number to fetch. Defaults to first page.

api_key
A Last.fm API key.

Value

A list of the top artists listened to by a user.

Examples

user_getTopArtists("platyjus")

#> $artist
#>                                      name playcount
#> 1                            Leon Bridges       260
#> 2                   ::M∆DE::IN::HEIGHTS::       246
#> 3                             Andrew Bird       160
#> 4                           The Lumineers       158
#> 5           Trent Reznor and Atticus Ross       156
#> 6                           Lisa Hannigan       151
#> 7                           Anderson Paak       142
#> 8                        Courtney Barnett       137
#> 9                                 Oddisee       133
#> 10                      Chance the Rapper       132
#> 11                         Kendrick Lamar       130
#> 12                            The Suffers       120
#> 13                         Run the Jewels       119
#> 14                 The Head and the Heart       119
#> 15                         Hey Marseilles       101
#> 16                          Isaiah Rashad       100
#> 17                                 Beirut        97
#> 18                              Sigur Rós        95
#> 19                            Little Simz        92
#> 20                         The Wild Reeds        88
#> 21                          Vince Staples        83
#> 22                            Joey Bada$$        79
#> 23                         Carla Morrison        77
#> 24                                 NoName        76
#> 25                             Kanye West        61
#> 26                            Frank Ocean        51
#> 27                                    Sol        51
#> 28                           Mick Jenkins        48
#> 29                   A Tribe Called Quest        47
#> 30                      Father John Misty        47
#> 31                         Ólafur Arnalds        47
#> 32                         Anderson .Paak        46
#> 33                       Flatbush ZOMBiES        46
#> 34                                 Skepta        44
#> 35                            James Blake        41
#> 36                          Ramin Djawadi        41
#> 37                         Sufjan Stevens        41
#> 38                                  Logic        40
#> 39                                 Sampha        40
#> 40                       Childish Gambino        39
#> 41                             Fruit Bats        39
#> 42                              DJ Shadow        38
#> 43                            ScHoolboy Q        38
#> 44                           Joan Shelley        35
#> 45 Donnie Trumpet & The Social Experiment        34
#> 46                            Fleet Foxes        31
#> 47                           Helado Negro        31
#> 48                         Band of Horses        30
#> 49                            Beach House        29
#> 50                          Tomo Nakayama        27
#>                                    mbid
#> 1                                      
#> 2  72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 3  081b133e-ce74-42ba-92c1-c18234acb532
#> 4  bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 5  bc880664-6751-4998-984d-ebdbff15d229
#> 6  8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#> 7  d02dd67e-f655-4600-bc47-f789f59e7367
#> 8  55111838-f001-494a-a1b5-9d818db85810
#> 9  aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 10 373a4c98-a46b-48e4-86ec-f6ca65b4f438
#> 11 381086ea-f511-4aba-bdf9-71c753dc5077
#> 12 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 13 a80c5dc5-b12e-4667-9f5a-b568961f3839
#> 14 98cb416e-5271-484c-ab3b-6bed6b4aae12
#> 15 4f0ca99b-0ab2-4861-b8ac-e6dadabff23e
#> 16 37547347-ea75-45fa-9766-ff36b44d8147
#> 17 989276c4-d2e8-4599-a2a0-c9ebeba73104
#> 18 f6f2326f-6b25-4170-b89d-e235b25508e8
#> 19 3cdb40fe-a63e-4bb9-b40d-17cda5f50979
#> 20 54412058-8ed9-49a2-b931-ce8071caf851
#> 21 78e854b8-9713-4ff2-9218-6b3784893bff
#> 22 9757ae7d-44b1-4d63-81ce-b463fe27b9ae
#> 23 50ee3204-13fb-4ae3-b152-82e3fc845110
#> 24 9393cc98-b014-493c-bdf1-a643a088f662
#> 25 164f0d73-1234-4e2c-8743-d77bf2191051
#> 26 e520459c-dff4-491d-a6e4-c97be35e0044
#> 27 f7ef4d71-9f9e-4b53-9f3a-0fa557167c94
#> 28 64c4dfcc-ba2b-4036-ac01-1636ae15ebb8
#> 29 9689aa5a-4471-4fb4-9721-07cecda0fa9f
#> 30 172053e5-a7fb-4548-a64d-c53858b5e775
#> 31 6655955b-1c1e-4bcb-84e4-81bcd9efab30
#> 32                                     
#> 33 f59a5dac-939a-4f12-a35c-9a49676f6898
#> 34 38474d18-9204-4f60-a137-1fcc6750b869
#> 35 8dc08b1f-e393-4f85-a5dd-300f7693a8b8
#> 36 f089f636-6b85-4e0e-a192-0d30ed2b44d1
#> 37 01d3c51b-9b98-418a-8d8e-37f6fab59d8c
#> 38 b2f45ca6-2c92-438c-b037-ad4d8b92e90e
#> 39 130c2184-14f2-4180-a216-04b5a333be57
#> 40 7fb57fba-a6ef-44c2-abab-2fa3bdee607e
#> 41 e3f139c7-11bc-4803-82be-2bac85ca3c61
#> 42 efa2c11a-1a35-4b60-bc1b-66d37de88511
#> 43 bce6d667-cde8-485e-b078-c0a05adea36d
#> 44                                     
#> 45                                     
#> 46 fa97dd36-1b82-43d7-a6e4-2adeafd59cef
#> 47 aa0a54b3-d37e-4c1d-a53f-b3b94a3dfcad
#> 48 07b6020a-c539-4d68-aeef-f159f3befc76
#> 49 d5cc67b8-1cc4-453b-96e8-44487acdebea
#> 50                                     
#>                                                                 url streamable
#> 1                            https://www.last.fm/music/Leon+Bridges          0
#> 2           https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::          0
#> 3                             https://www.last.fm/music/Andrew+Bird          0
#> 4                           https://www.last.fm/music/The+Lumineers          0
#> 5           https://www.last.fm/music/Trent+Reznor+and+Atticus+Ross          0
#> 6                           https://www.last.fm/music/Lisa+Hannigan          0
#> 7                           https://www.last.fm/music/Anderson+Paak          0
#> 8                        https://www.last.fm/music/Courtney+Barnett          0
#> 9                                 https://www.last.fm/music/Oddisee          0
#> 10                      https://www.last.fm/music/Chance+the+Rapper          0
#> 11                         https://www.last.fm/music/Kendrick+Lamar          0
#> 12                            https://www.last.fm/music/The+Suffers          0
#> 13                         https://www.last.fm/music/Run+the+Jewels          0
#> 14                 https://www.last.fm/music/The+Head+and+the+Heart          0
#> 15                         https://www.last.fm/music/Hey+Marseilles          0
#> 16                          https://www.last.fm/music/Isaiah+Rashad          0
#> 17                                 https://www.last.fm/music/Beirut          0
#> 18                         https://www.last.fm/music/Sigur+R%C3%B3s          0
#> 19                            https://www.last.fm/music/Little+Simz          0
#> 20                         https://www.last.fm/music/The+Wild+Reeds          0
#> 21                          https://www.last.fm/music/Vince+Staples          0
#> 22                            https://www.last.fm/music/Joey+Bada$$          0
#> 23                         https://www.last.fm/music/Carla+Morrison          0
#> 24                                 https://www.last.fm/music/NoName          0
#> 25                             https://www.last.fm/music/Kanye+West          0
#> 26                            https://www.last.fm/music/Frank+Ocean          0
#> 27                                    https://www.last.fm/music/Sol          0
#> 28                           https://www.last.fm/music/Mick+Jenkins          0
#> 29                   https://www.last.fm/music/A+Tribe+Called+Quest          0
#> 30                      https://www.last.fm/music/Father+John+Misty          0
#> 31                    https://www.last.fm/music/%C3%93lafur+Arnalds          0
#> 32                         https://www.last.fm/music/Anderson+.Paak          0
#> 33                       https://www.last.fm/music/Flatbush+ZOMBiES          0
#> 34                                 https://www.last.fm/music/Skepta          0
#> 35                            https://www.last.fm/music/James+Blake          0
#> 36                          https://www.last.fm/music/Ramin+Djawadi          0
#> 37                         https://www.last.fm/music/Sufjan+Stevens          0
#> 38                                  https://www.last.fm/music/Logic          0
#> 39                                 https://www.last.fm/music/Sampha          0
#> 40                       https://www.last.fm/music/Childish+Gambino          0
#> 41                             https://www.last.fm/music/Fruit+Bats          0
#> 42                              https://www.last.fm/music/DJ+Shadow          0
#> 43                            https://www.last.fm/music/ScHoolboy+Q          0
#> 44                           https://www.last.fm/music/Joan+Shelley          0
#> 45 https://www.last.fm/music/Donnie+Trumpet+&+The+Social+Experiment          0
#> 46                            https://www.last.fm/music/Fleet+Foxes          0
#> 47                           https://www.last.fm/music/Helado+Negro          0
#> 48                         https://www.last.fm/music/Band+of+Horses          0
#> 49                            https://www.last.fm/music/Beach+House          0
#> 50                          https://www.last.fm/music/Tomo+Nakayama          0
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                      image
#> 1  https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge, mega
#> 2  https://lastfm-img2.akamaized.net/i/u/34s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/64s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/174s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/300x300/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/2b45cccc6ca247a6c78cca6ace9c2114.png, small, medium, large, extralarge, mega
#> 3  https://lastfm-img2.akamaized.net/i/u/34s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/64s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/174s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/300x300/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/73dba10b71674e029b3ddc0692899e0b.png, small, medium, large, extralarge, mega
#> 4  https://lastfm-img2.akamaized.net/i/u/34s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/64s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/174s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/300x300/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/03e320dd70c0577c42c459eac0063c16.png, small, medium, large, extralarge, mega
#> 5  https://lastfm-img2.akamaized.net/i/u/34s/79c9b637803a79ce4aeaede8ec93da38.png, https://lastfm-img2.akamaized.net/i/u/64s/79c9b637803a79ce4aeaede8ec93da38.png, https://lastfm-img2.akamaized.net/i/u/174s/79c9b637803a79ce4aeaede8ec93da38.png, https://lastfm-img2.akamaized.net/i/u/300x300/79c9b637803a79ce4aeaede8ec93da38.png, https://lastfm-img2.akamaized.net/i/u/79c9b637803a79ce4aeaede8ec93da38.png, small, medium, large, extralarge, mega
#> 6  https://lastfm-img2.akamaized.net/i/u/34s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/64s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/174s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/300x300/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/c8f8be9e530f4956a70c00d46c272fbe.png, small, medium, large, extralarge, mega
#> 7  https://lastfm-img2.akamaized.net/i/u/34s/e009139061502922de33ed960c5fac9c.png, https://lastfm-img2.akamaized.net/i/u/64s/e009139061502922de33ed960c5fac9c.png, https://lastfm-img2.akamaized.net/i/u/174s/e009139061502922de33ed960c5fac9c.png, https://lastfm-img2.akamaized.net/i/u/300x300/e009139061502922de33ed960c5fac9c.png, https://lastfm-img2.akamaized.net/i/u/e009139061502922de33ed960c5fac9c.png, small, medium, large, extralarge, mega
#> 8  https://lastfm-img2.akamaized.net/i/u/34s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/64s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/174s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/300x300/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/ac5489a01d41cb391ab5f745d157a291.png, small, medium, large, extralarge, mega
#> 9  https://lastfm-img2.akamaized.net/i/u/34s/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/64s/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/174s/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/300x300/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/f6d5cc790d1b48c5838ff4de1c6f7549.png, small, medium, large, extralarge, mega
#> 10 https://lastfm-img2.akamaized.net/i/u/34s/3d42bab35e7ad7332a27b9ff38db971a.png, https://lastfm-img2.akamaized.net/i/u/64s/3d42bab35e7ad7332a27b9ff38db971a.png, https://lastfm-img2.akamaized.net/i/u/174s/3d42bab35e7ad7332a27b9ff38db971a.png, https://lastfm-img2.akamaized.net/i/u/300x300/3d42bab35e7ad7332a27b9ff38db971a.png, https://lastfm-img2.akamaized.net/i/u/3d42bab35e7ad7332a27b9ff38db971a.png, small, medium, large, extralarge, mega
#> 11 https://lastfm-img2.akamaized.net/i/u/34s/a4a35fcdf62b8a532039012653eeea14.png, https://lastfm-img2.akamaized.net/i/u/64s/a4a35fcdf62b8a532039012653eeea14.png, https://lastfm-img2.akamaized.net/i/u/174s/a4a35fcdf62b8a532039012653eeea14.png, https://lastfm-img2.akamaized.net/i/u/300x300/a4a35fcdf62b8a532039012653eeea14.png, https://lastfm-img2.akamaized.net/i/u/a4a35fcdf62b8a532039012653eeea14.png, small, medium, large, extralarge, mega
#> 12 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/f48930d08bfc7c2caf7631d6bdcdb91c.png, small, medium, large, extralarge, mega
#> 13 https://lastfm-img2.akamaized.net/i/u/34s/fbb820614a34485ac8b0a1f2d0f02c9b.png, https://lastfm-img2.akamaized.net/i/u/64s/fbb820614a34485ac8b0a1f2d0f02c9b.png, https://lastfm-img2.akamaized.net/i/u/174s/fbb820614a34485ac8b0a1f2d0f02c9b.png, https://lastfm-img2.akamaized.net/i/u/300x300/fbb820614a34485ac8b0a1f2d0f02c9b.png, https://lastfm-img2.akamaized.net/i/u/fbb820614a34485ac8b0a1f2d0f02c9b.png, small, medium, large, extralarge, mega
#> 14 https://lastfm-img2.akamaized.net/i/u/34s/2c50aec4065c4947be519bd11f15aba8.png, https://lastfm-img2.akamaized.net/i/u/64s/2c50aec4065c4947be519bd11f15aba8.png, https://lastfm-img2.akamaized.net/i/u/174s/2c50aec4065c4947be519bd11f15aba8.png, https://lastfm-img2.akamaized.net/i/u/300x300/2c50aec4065c4947be519bd11f15aba8.png, https://lastfm-img2.akamaized.net/i/u/2c50aec4065c4947be519bd11f15aba8.png, small, medium, large, extralarge, mega
#> 15 https://lastfm-img2.akamaized.net/i/u/34s/9b79ac7edb6d41a1b32ad223016210c8.png, https://lastfm-img2.akamaized.net/i/u/64s/9b79ac7edb6d41a1b32ad223016210c8.png, https://lastfm-img2.akamaized.net/i/u/174s/9b79ac7edb6d41a1b32ad223016210c8.png, https://lastfm-img2.akamaized.net/i/u/300x300/9b79ac7edb6d41a1b32ad223016210c8.png, https://lastfm-img2.akamaized.net/i/u/9b79ac7edb6d41a1b32ad223016210c8.png, small, medium, large, extralarge, mega
#> 16 https://lastfm-img2.akamaized.net/i/u/34s/eba48bf37a654c1fccc9fda5ac562fc7.png, https://lastfm-img2.akamaized.net/i/u/64s/eba48bf37a654c1fccc9fda5ac562fc7.png, https://lastfm-img2.akamaized.net/i/u/174s/eba48bf37a654c1fccc9fda5ac562fc7.png, https://lastfm-img2.akamaized.net/i/u/300x300/eba48bf37a654c1fccc9fda5ac562fc7.png, https://lastfm-img2.akamaized.net/i/u/eba48bf37a654c1fccc9fda5ac562fc7.png, small, medium, large, extralarge, mega
#> 17 https://lastfm-img2.akamaized.net/i/u/34s/28a739aacdeb40b1acc24d1b75030fdf.png, https://lastfm-img2.akamaized.net/i/u/64s/28a739aacdeb40b1acc24d1b75030fdf.png, https://lastfm-img2.akamaized.net/i/u/174s/28a739aacdeb40b1acc24d1b75030fdf.png, https://lastfm-img2.akamaized.net/i/u/300x300/28a739aacdeb40b1acc24d1b75030fdf.png, https://lastfm-img2.akamaized.net/i/u/28a739aacdeb40b1acc24d1b75030fdf.png, small, medium, large, extralarge, mega
#> 18 https://lastfm-img2.akamaized.net/i/u/34s/1cb7c3262a664d1da6f7a548e2f3036d.png, https://lastfm-img2.akamaized.net/i/u/64s/1cb7c3262a664d1da6f7a548e2f3036d.png, https://lastfm-img2.akamaized.net/i/u/174s/1cb7c3262a664d1da6f7a548e2f3036d.png, https://lastfm-img2.akamaized.net/i/u/300x300/1cb7c3262a664d1da6f7a548e2f3036d.png, https://lastfm-img2.akamaized.net/i/u/1cb7c3262a664d1da6f7a548e2f3036d.png, small, medium, large, extralarge, mega
#> 19 https://lastfm-img2.akamaized.net/i/u/34s/fbe808618eb37ca062e9c3fc0c26492c.png, https://lastfm-img2.akamaized.net/i/u/64s/fbe808618eb37ca062e9c3fc0c26492c.png, https://lastfm-img2.akamaized.net/i/u/174s/fbe808618eb37ca062e9c3fc0c26492c.png, https://lastfm-img2.akamaized.net/i/u/300x300/fbe808618eb37ca062e9c3fc0c26492c.png, https://lastfm-img2.akamaized.net/i/u/fbe808618eb37ca062e9c3fc0c26492c.png, small, medium, large, extralarge, mega
#> 20 https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/300x300/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/85f1a1719c316244d4a7740c42246da0.png, small, medium, large, extralarge, mega
#> 21 https://lastfm-img2.akamaized.net/i/u/34s/7cdf797748268b3385a56445fc751c39.png, https://lastfm-img2.akamaized.net/i/u/64s/7cdf797748268b3385a56445fc751c39.png, https://lastfm-img2.akamaized.net/i/u/174s/7cdf797748268b3385a56445fc751c39.png, https://lastfm-img2.akamaized.net/i/u/300x300/7cdf797748268b3385a56445fc751c39.png, https://lastfm-img2.akamaized.net/i/u/7cdf797748268b3385a56445fc751c39.png, small, medium, large, extralarge, mega
#> 22 https://lastfm-img2.akamaized.net/i/u/34s/56e0b6ab50624765c7ea12fd8040f000.png, https://lastfm-img2.akamaized.net/i/u/64s/56e0b6ab50624765c7ea12fd8040f000.png, https://lastfm-img2.akamaized.net/i/u/174s/56e0b6ab50624765c7ea12fd8040f000.png, https://lastfm-img2.akamaized.net/i/u/300x300/56e0b6ab50624765c7ea12fd8040f000.png, https://lastfm-img2.akamaized.net/i/u/56e0b6ab50624765c7ea12fd8040f000.png, small, medium, large, extralarge, mega
#> 23 https://lastfm-img2.akamaized.net/i/u/34s/04eb0d30dc1f4886bb1ad0299993d67e.png, https://lastfm-img2.akamaized.net/i/u/64s/04eb0d30dc1f4886bb1ad0299993d67e.png, https://lastfm-img2.akamaized.net/i/u/174s/04eb0d30dc1f4886bb1ad0299993d67e.png, https://lastfm-img2.akamaized.net/i/u/300x300/04eb0d30dc1f4886bb1ad0299993d67e.png, https://lastfm-img2.akamaized.net/i/u/04eb0d30dc1f4886bb1ad0299993d67e.png, small, medium, large, extralarge, mega
#> 24 https://lastfm-img2.akamaized.net/i/u/34s/dbe58a33f97279ad7a6037e67d7f43f6.png, https://lastfm-img2.akamaized.net/i/u/64s/dbe58a33f97279ad7a6037e67d7f43f6.png, https://lastfm-img2.akamaized.net/i/u/174s/dbe58a33f97279ad7a6037e67d7f43f6.png, https://lastfm-img2.akamaized.net/i/u/300x300/dbe58a33f97279ad7a6037e67d7f43f6.png, https://lastfm-img2.akamaized.net/i/u/dbe58a33f97279ad7a6037e67d7f43f6.png, small, medium, large, extralarge, mega
#> 25 https://lastfm-img2.akamaized.net/i/u/34s/41f5fe0fc4d27544e7d47fee1583a374.png, https://lastfm-img2.akamaized.net/i/u/64s/41f5fe0fc4d27544e7d47fee1583a374.png, https://lastfm-img2.akamaized.net/i/u/174s/41f5fe0fc4d27544e7d47fee1583a374.png, https://lastfm-img2.akamaized.net/i/u/300x300/41f5fe0fc4d27544e7d47fee1583a374.png, https://lastfm-img2.akamaized.net/i/u/41f5fe0fc4d27544e7d47fee1583a374.png, small, medium, large, extralarge, mega
#> 26 https://lastfm-img2.akamaized.net/i/u/34s/d48ee8f148783679d9d933d0948e4e7b.png, https://lastfm-img2.akamaized.net/i/u/64s/d48ee8f148783679d9d933d0948e4e7b.png, https://lastfm-img2.akamaized.net/i/u/174s/d48ee8f148783679d9d933d0948e4e7b.png, https://lastfm-img2.akamaized.net/i/u/300x300/d48ee8f148783679d9d933d0948e4e7b.png, https://lastfm-img2.akamaized.net/i/u/d48ee8f148783679d9d933d0948e4e7b.png, small, medium, large, extralarge, mega
#> 27 https://lastfm-img2.akamaized.net/i/u/34s/c32f9c62e8ef4bd0ae2e4af11021f05a.png, https://lastfm-img2.akamaized.net/i/u/64s/c32f9c62e8ef4bd0ae2e4af11021f05a.png, https://lastfm-img2.akamaized.net/i/u/174s/c32f9c62e8ef4bd0ae2e4af11021f05a.png, https://lastfm-img2.akamaized.net/i/u/300x300/c32f9c62e8ef4bd0ae2e4af11021f05a.png, https://lastfm-img2.akamaized.net/i/u/c32f9c62e8ef4bd0ae2e4af11021f05a.png, small, medium, large, extralarge, mega
#> 28 https://lastfm-img2.akamaized.net/i/u/34s/f86c88f8e069414dc1c5152a5516200e.png, https://lastfm-img2.akamaized.net/i/u/64s/f86c88f8e069414dc1c5152a5516200e.png, https://lastfm-img2.akamaized.net/i/u/174s/f86c88f8e069414dc1c5152a5516200e.png, https://lastfm-img2.akamaized.net/i/u/300x300/f86c88f8e069414dc1c5152a5516200e.png, https://lastfm-img2.akamaized.net/i/u/f86c88f8e069414dc1c5152a5516200e.png, small, medium, large, extralarge, mega
#> 29 https://lastfm-img2.akamaized.net/i/u/34s/0fb85ae7231548f7b9d610ba38f47c75.png, https://lastfm-img2.akamaized.net/i/u/64s/0fb85ae7231548f7b9d610ba38f47c75.png, https://lastfm-img2.akamaized.net/i/u/174s/0fb85ae7231548f7b9d610ba38f47c75.png, https://lastfm-img2.akamaized.net/i/u/300x300/0fb85ae7231548f7b9d610ba38f47c75.png, https://lastfm-img2.akamaized.net/i/u/0fb85ae7231548f7b9d610ba38f47c75.png, small, medium, large, extralarge, mega
#> 30 https://lastfm-img2.akamaized.net/i/u/34s/b5e4739b38896c44f3fca387c08f4279.png, https://lastfm-img2.akamaized.net/i/u/64s/b5e4739b38896c44f3fca387c08f4279.png, https://lastfm-img2.akamaized.net/i/u/174s/b5e4739b38896c44f3fca387c08f4279.png, https://lastfm-img2.akamaized.net/i/u/300x300/b5e4739b38896c44f3fca387c08f4279.png, https://lastfm-img2.akamaized.net/i/u/b5e4739b38896c44f3fca387c08f4279.png, small, medium, large, extralarge, mega
#> 31 https://lastfm-img2.akamaized.net/i/u/34s/28aeaea416cd3aaade5d3a31ee45136b.png, https://lastfm-img2.akamaized.net/i/u/64s/28aeaea416cd3aaade5d3a31ee45136b.png, https://lastfm-img2.akamaized.net/i/u/174s/28aeaea416cd3aaade5d3a31ee45136b.png, https://lastfm-img2.akamaized.net/i/u/300x300/28aeaea416cd3aaade5d3a31ee45136b.png, https://lastfm-img2.akamaized.net/i/u/28aeaea416cd3aaade5d3a31ee45136b.png, small, medium, large, extralarge, mega
#> 32 https://lastfm-img2.akamaized.net/i/u/34s/c2171d1092ef0c995fd6165bfeffe548.png, https://lastfm-img2.akamaized.net/i/u/64s/c2171d1092ef0c995fd6165bfeffe548.png, https://lastfm-img2.akamaized.net/i/u/174s/c2171d1092ef0c995fd6165bfeffe548.png, https://lastfm-img2.akamaized.net/i/u/300x300/c2171d1092ef0c995fd6165bfeffe548.png, https://lastfm-img2.akamaized.net/i/u/c2171d1092ef0c995fd6165bfeffe548.png, small, medium, large, extralarge, mega
#> 33 https://lastfm-img2.akamaized.net/i/u/34s/b9d69347fd9e492ba7e4762e777eb86e.png, https://lastfm-img2.akamaized.net/i/u/64s/b9d69347fd9e492ba7e4762e777eb86e.png, https://lastfm-img2.akamaized.net/i/u/174s/b9d69347fd9e492ba7e4762e777eb86e.png, https://lastfm-img2.akamaized.net/i/u/300x300/b9d69347fd9e492ba7e4762e777eb86e.png, https://lastfm-img2.akamaized.net/i/u/b9d69347fd9e492ba7e4762e777eb86e.png, small, medium, large, extralarge, mega
#> 34 https://lastfm-img2.akamaized.net/i/u/34s/f6ddfd432f95b8ba4c64575eea0b359d.png, https://lastfm-img2.akamaized.net/i/u/64s/f6ddfd432f95b8ba4c64575eea0b359d.png, https://lastfm-img2.akamaized.net/i/u/174s/f6ddfd432f95b8ba4c64575eea0b359d.png, https://lastfm-img2.akamaized.net/i/u/300x300/f6ddfd432f95b8ba4c64575eea0b359d.png, https://lastfm-img2.akamaized.net/i/u/f6ddfd432f95b8ba4c64575eea0b359d.png, small, medium, large, extralarge, mega
#> 35 https://lastfm-img2.akamaized.net/i/u/34s/4b53885d58f049158ad3324d1dfa808f.png, https://lastfm-img2.akamaized.net/i/u/64s/4b53885d58f049158ad3324d1dfa808f.png, https://lastfm-img2.akamaized.net/i/u/174s/4b53885d58f049158ad3324d1dfa808f.png, https://lastfm-img2.akamaized.net/i/u/300x300/4b53885d58f049158ad3324d1dfa808f.png, https://lastfm-img2.akamaized.net/i/u/4b53885d58f049158ad3324d1dfa808f.png, small, medium, large, extralarge, mega
#> 36 https://lastfm-img2.akamaized.net/i/u/34s/aca9aeed35974ecbce7323ee5f5862be.png, https://lastfm-img2.akamaized.net/i/u/64s/aca9aeed35974ecbce7323ee5f5862be.png, https://lastfm-img2.akamaized.net/i/u/174s/aca9aeed35974ecbce7323ee5f5862be.png, https://lastfm-img2.akamaized.net/i/u/300x300/aca9aeed35974ecbce7323ee5f5862be.png, https://lastfm-img2.akamaized.net/i/u/aca9aeed35974ecbce7323ee5f5862be.png, small, medium, large, extralarge, mega
#> 37 https://lastfm-img2.akamaized.net/i/u/34s/b5676ef4915d42838f8b2cc26c7536be.png, https://lastfm-img2.akamaized.net/i/u/64s/b5676ef4915d42838f8b2cc26c7536be.png, https://lastfm-img2.akamaized.net/i/u/174s/b5676ef4915d42838f8b2cc26c7536be.png, https://lastfm-img2.akamaized.net/i/u/300x300/b5676ef4915d42838f8b2cc26c7536be.png, https://lastfm-img2.akamaized.net/i/u/b5676ef4915d42838f8b2cc26c7536be.png, small, medium, large, extralarge, mega
#> 38 https://lastfm-img2.akamaized.net/i/u/34s/dad642d49cc9ed1553737dd6e9f0e6f9.png, https://lastfm-img2.akamaized.net/i/u/64s/dad642d49cc9ed1553737dd6e9f0e6f9.png, https://lastfm-img2.akamaized.net/i/u/174s/dad642d49cc9ed1553737dd6e9f0e6f9.png, https://lastfm-img2.akamaized.net/i/u/300x300/dad642d49cc9ed1553737dd6e9f0e6f9.png, https://lastfm-img2.akamaized.net/i/u/dad642d49cc9ed1553737dd6e9f0e6f9.png, small, medium, large, extralarge, mega
#> 39 https://lastfm-img2.akamaized.net/i/u/34s/b6dbab944db8cc773ba370193b653fd6.png, https://lastfm-img2.akamaized.net/i/u/64s/b6dbab944db8cc773ba370193b653fd6.png, https://lastfm-img2.akamaized.net/i/u/174s/b6dbab944db8cc773ba370193b653fd6.png, https://lastfm-img2.akamaized.net/i/u/300x300/b6dbab944db8cc773ba370193b653fd6.png, https://lastfm-img2.akamaized.net/i/u/b6dbab944db8cc773ba370193b653fd6.png, small, medium, large, extralarge, mega
#> 40 https://lastfm-img2.akamaized.net/i/u/34s/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/64s/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/174s/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/300x300/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/95384659e8a160a327e0d097a76db1f3.png, small, medium, large, extralarge, mega
#> 41 https://lastfm-img2.akamaized.net/i/u/34s/b03375321bef44ad9a944e0a7020b4f7.png, https://lastfm-img2.akamaized.net/i/u/64s/b03375321bef44ad9a944e0a7020b4f7.png, https://lastfm-img2.akamaized.net/i/u/174s/b03375321bef44ad9a944e0a7020b4f7.png, https://lastfm-img2.akamaized.net/i/u/300x300/b03375321bef44ad9a944e0a7020b4f7.png, https://lastfm-img2.akamaized.net/i/u/b03375321bef44ad9a944e0a7020b4f7.png, small, medium, large, extralarge, mega
#> 42 https://lastfm-img2.akamaized.net/i/u/34s/33a127e674f149f18d9e27528d469302.png, https://lastfm-img2.akamaized.net/i/u/64s/33a127e674f149f18d9e27528d469302.png, https://lastfm-img2.akamaized.net/i/u/174s/33a127e674f149f18d9e27528d469302.png, https://lastfm-img2.akamaized.net/i/u/300x300/33a127e674f149f18d9e27528d469302.png, https://lastfm-img2.akamaized.net/i/u/33a127e674f149f18d9e27528d469302.png, small, medium, large, extralarge, mega
#> 43 https://lastfm-img2.akamaized.net/i/u/34s/71438a1a33fd44188127f5556bb5dee3.png, https://lastfm-img2.akamaized.net/i/u/64s/71438a1a33fd44188127f5556bb5dee3.png, https://lastfm-img2.akamaized.net/i/u/174s/71438a1a33fd44188127f5556bb5dee3.png, https://lastfm-img2.akamaized.net/i/u/300x300/71438a1a33fd44188127f5556bb5dee3.png, https://lastfm-img2.akamaized.net/i/u/71438a1a33fd44188127f5556bb5dee3.png, small, medium, large, extralarge, mega
#> 44 https://lastfm-img2.akamaized.net/i/u/34s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/64s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/174s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/300x300/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/4725042f9ec445e0ba8295d755edf41f.png, small, medium, large, extralarge, mega
#> 45 https://lastfm-img2.akamaized.net/i/u/34s/13d7fdafb4eb3217f8ce8b8bcbad4661.png, https://lastfm-img2.akamaized.net/i/u/64s/13d7fdafb4eb3217f8ce8b8bcbad4661.png, https://lastfm-img2.akamaized.net/i/u/174s/13d7fdafb4eb3217f8ce8b8bcbad4661.png, https://lastfm-img2.akamaized.net/i/u/300x300/13d7fdafb4eb3217f8ce8b8bcbad4661.png, https://lastfm-img2.akamaized.net/i/u/13d7fdafb4eb3217f8ce8b8bcbad4661.png, small, medium, large, extralarge, mega
#> 46 https://lastfm-img2.akamaized.net/i/u/34s/9f2a280863a94830bcf70f56c287bae8.png, https://lastfm-img2.akamaized.net/i/u/64s/9f2a280863a94830bcf70f56c287bae8.png, https://lastfm-img2.akamaized.net/i/u/174s/9f2a280863a94830bcf70f56c287bae8.png, https://lastfm-img2.akamaized.net/i/u/300x300/9f2a280863a94830bcf70f56c287bae8.png, https://lastfm-img2.akamaized.net/i/u/9f2a280863a94830bcf70f56c287bae8.png, small, medium, large, extralarge, mega
#> 47 https://lastfm-img2.akamaized.net/i/u/34s/a27c0c1e014643b292e499b6d9e3ce1a.png, https://lastfm-img2.akamaized.net/i/u/64s/a27c0c1e014643b292e499b6d9e3ce1a.png, https://lastfm-img2.akamaized.net/i/u/174s/a27c0c1e014643b292e499b6d9e3ce1a.png, https://lastfm-img2.akamaized.net/i/u/300x300/a27c0c1e014643b292e499b6d9e3ce1a.png, https://lastfm-img2.akamaized.net/i/u/a27c0c1e014643b292e499b6d9e3ce1a.png, small, medium, large, extralarge, mega
#> 48 https://lastfm-img2.akamaized.net/i/u/34s/ff725ce238ff48b0b72fc88beea038e1.png, https://lastfm-img2.akamaized.net/i/u/64s/ff725ce238ff48b0b72fc88beea038e1.png, https://lastfm-img2.akamaized.net/i/u/174s/ff725ce238ff48b0b72fc88beea038e1.png, https://lastfm-img2.akamaized.net/i/u/300x300/ff725ce238ff48b0b72fc88beea038e1.png, https://lastfm-img2.akamaized.net/i/u/ff725ce238ff48b0b72fc88beea038e1.png, small, medium, large, extralarge, mega
#> 49 https://lastfm-img2.akamaized.net/i/u/34s/17d7b30da49d4d76a4a5da7b0982115b.png, https://lastfm-img2.akamaized.net/i/u/64s/17d7b30da49d4d76a4a5da7b0982115b.png, https://lastfm-img2.akamaized.net/i/u/174s/17d7b30da49d4d76a4a5da7b0982115b.png, https://lastfm-img2.akamaized.net/i/u/300x300/17d7b30da49d4d76a4a5da7b0982115b.png, https://lastfm-img2.akamaized.net/i/u/17d7b30da49d4d76a4a5da7b0982115b.png, small, medium, large, extralarge, mega
#> 50 https://lastfm-img2.akamaized.net/i/u/34s/840ce96e127e406cacf9c40b88b81a0f.png, https://lastfm-img2.akamaized.net/i/u/64s/840ce96e127e406cacf9c40b88b81a0f.png, https://lastfm-img2.akamaized.net/i/u/174s/840ce96e127e406cacf9c40b88b81a0f.png, https://lastfm-img2.akamaized.net/i/u/300x300/840ce96e127e406cacf9c40b88b81a0f.png, https://lastfm-img2.akamaized.net/i/u/840ce96e127e406cacf9c40b88b81a0f.png, small, medium, large, extralarge, mega
#>    rank
#> 1     1
#> 2     2
#> 3     3
#> 4     4
#> 5     5
#> 6     6
#> 7     7
#> 8     8
#> 9     9
#> 10   10
#> 11   11
#> 12   12
#> 13   13
#> 14   14
#> 15   15
#> 16   16
#> 17   17
#> 18   18
#> 19   19
#> 20   20
#> 21   21
#> 22   22
#> 23   23
#> 24   24
#> 25   25
#> 26   26
#> 27   27
#> 28   28
#> 29   29
#> 30   30
#> 31   31
#> 32   32
#> 33   33
#> 34   34
#> 35   35
#> 36   36
#> 37   37
#> 38   38
#> 39   39
#> 40   40
#> 41   41
#> 42   42
#> 43   43
#> 44   44
#> 45   45
#> 46   46
#> 47   47
#> 48   48
#> 49   49
#> 50   50
#> 
#> $`@attr`
#> $`@attr`$user
#> [1] "platyjus"
#> 
#> $`@attr`$page
#> [1] "1"
#> 
#> $`@attr`$perPage
#> [1] "50"
#> 
#> $`@attr`$totalPages
#> [1] "12"
#> 
#> $`@attr`$total
#> [1] "594"
#> 
#>