Get the top tracks listened to by a user.

Get the top tracks listened to by a user. You can stipulate a time period. Sends the overall chart by default. Implementation of last.fm's user.getTopTracks API method (http://www.last.fm/api/show/user.getTopTracks).

user_getTopTracks(user, period = NA, limit = NA, page = NA)

Arguments

user
The user name to fetch top tracks for.

period
overall | 7day | 1month | 3month | 6month | 12month - The time period over which to retrieve top tracks for.

limit
The number of results to fetch per page. Defaults to 50.

page
The page number to fetch. Defaults to first page.

Value

A list of the top tracks listened to by a user.

Examples

user_getTopTracks("platyjus")

#>                                                name duration playcount
#> 1                                        Better Man        0        30
#> 2                                       Coming Home        0        30
#> 3                                              Fall        0        26
#> 4                                   Brown Skin Girl        0        24
#> 5               Nobody Speak (feat. Run the Jewels)        0        24
#> 6                                           Ophelia        0        23
#> 7                                Sleep on the Floor        0        23
#> 8                                         Blessings        0        21
#> 9                              Prayer For The Dying        0        21
#> 10                                         Capsized        0        20
#> 11                                      Lisa Sawyer        0        20
#> 12                                          Peanuts        0        20
#> 13                                            River        0        20
#> 14                                   Smooth Sailin'        0        20
#> 15                                        Cleopatra        0        19
#> 16                                   Make Some Room        0        19
#> 17                                            DEATH        0        18
#> 18                                             Gwan      231        18
#> 19                                        Norf Norf        0        18
#> 20                                        Pull Away        0        18
#> 21                                            Shine        0        18
#> 22                             Skylark Interabang?!      283        18
#> 23                                             Snow        0        18
#> 24                                         Gun Song      191        17
#> 25                                              Ora        0        17
#> 26                                             Stay        0        17
#> 27                                 All We Ever Knew        0        16
#> 28                                        Depreston        0        16
#> 29                                            Giver        0        16
#> 30 Nobody Really Cares If You Don't Go to the Party        0        16
#> 31                                           Angela        0        15
#> 32                                           Better        0        15
#> 33                         Blockbuster Night Part 1      152        15
#> 34           Everything Looks Better (In Hindsight)        0        15
#> 35                                         Good Day        0        15
#> 36                                        Yesterday        0        15
#> 37                                   All The Places      196        14
#> 38                                     Digging Deep        0        14
#> 39                                          Flowers        0        14
#> 40                                         LOCATION        0        14
#> 41                                      Pop It in 2        0        14
#> 42                                        ROMA FADE        0        14
#> 43                                    Þú ert jörðin        0        14
#> 44                                      4r Da Squaw        0        13
#> 45                                        95 Radios        0        13
#> 46                                         Big Fish        0        13
#> 47                                       Free Lunch        0        13
#> 48                                  Here In My Arms        0        13
#> 49                                     Hors d'œuvre      205        13
#> 50                                     I Don't Know      296        13
#>                                    mbid
#> 1                                      
#> 2                                      
#> 3                                      
#> 4                                      
#> 5                                      
#> 6                                      
#> 7                                      
#> 8                                      
#> 9                                      
#> 10                                     
#> 11                                     
#> 12                                     
#> 13                                     
#> 14                                     
#> 15                                     
#> 16                                     
#> 17                                     
#> 18 c848fff9-2434-4d5e-8f42-456381a033a2
#> 19                                     
#> 20                                     
#> 21                                     
#> 22 4d9b5da0-9a5e-4b24-97e9-d3497fc3a3c1
#> 23                                     
#> 24 f108c60c-e286-4bda-9584-dd0cb70535c9
#> 25                                     
#> 26                                     
#> 27                                     
#> 28                                     
#> 29                                     
#> 30                                     
#> 31                                     
#> 32                                     
#> 33 3532826e-5523-42d1-bf9d-8b5318f9470d
#> 34                                     
#> 35                                     
#> 36                                     
#> 37 1d6bf8bf-7030-4254-9617-2d499ffc16eb
#> 38                                     
#> 39                                     
#> 40                                     
#> 41                                     
#> 42                                     
#> 43                                     
#> 44                                     
#> 45                                     
#> 46                                     
#> 47                                     
#> 48                                     
#> 49 e3a9dfc0-ef48-4b9c-84d9-e1092530e5fa
#> 50 0033f256-c78f-4cbb-877c-20d03506fb2a
#>                                                                                                url
#> 1                                              https://www.last.fm/music/Leon+Bridges/_/Better+Man
#> 2                                             https://www.last.fm/music/Leon+Bridges/_/Coming+Home
#> 3                                                   https://www.last.fm/music/Lisa+Hannigan/_/Fall
#> 4                                         https://www.last.fm/music/Leon+Bridges/_/Brown+Skin+Girl
#> 5                        https://www.last.fm/music/DJ+Shadow/_/Nobody+Speak+(feat.+Run+the+Jewels)
#> 6                                                https://www.last.fm/music/The+Lumineers/_/Ophelia
#> 7                                     https://www.last.fm/music/The+Lumineers/_/Sleep+on+the+Floor
#> 8                                          https://www.last.fm/music/Chance+the+Rapper/_/Blessings
#> 9                                   https://www.last.fm/music/Lisa+Hannigan/_/Prayer+For+The+Dying
#> 10                                                https://www.last.fm/music/Andrew+Bird/_/Capsized
#> 11                                            https://www.last.fm/music/Leon+Bridges/_/Lisa+Sawyer
#> 12                                                 https://www.last.fm/music/The+Suffers/_/Peanuts
#> 13                                                  https://www.last.fm/music/Leon+Bridges/_/River
#> 14                                       https://www.last.fm/music/Leon+Bridges/_/Smooth+Sailin%27
#> 15                                             https://www.last.fm/music/The+Lumineers/_/Cleopatra
#> 16                                          https://www.last.fm/music/The+Suffers/_/Make+Some+Room
#> 17                                 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/_/DEATH
#> 18                                                    https://www.last.fm/music/The+Suffers/_/Gwan
#> 19                                             https://www.last.fm/music/Vince+Staples/_/Norf+Norf
#> 20                                              https://www.last.fm/music/Leon+Bridges/_/Pull+Away
#> 21                                                  https://www.last.fm/music/Leon+Bridges/_/Shine
#> 22              https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/_/Skylark+Interabang%3F%21
#> 23                                                  https://www.last.fm/music/Lisa+Hannigan/_/Snow
#> 24                                              https://www.last.fm/music/The+Lumineers/_/Gun+Song
#> 25                                                   https://www.last.fm/music/Lisa+Hannigan/_/Ora
#> 26                                                    https://www.last.fm/music/The+Suffers/_/Stay
#> 27                             https://www.last.fm/music/The+Head+and+the+Heart/_/All+We+Ever+Knew
#> 28                                          https://www.last.fm/music/Courtney+Barnett/_/Depreston
#> 29                                                   https://www.last.fm/music/The+Suffers/_/Giver
#> 30 https://www.last.fm/music/Courtney+Barnett/_/Nobody+Really+Cares+If+You+Don%27t+Go+to+the+Party
#> 31                                                https://www.last.fm/music/The+Lumineers/_/Angela
#> 32                                                  https://www.last.fm/music/The+Suffers/_/Better
#> 33                             https://www.last.fm/music/Run+the+Jewels/_/Blockbuster+Night+Part+1
#> 34               https://www.last.fm/music/The+Wild+Reeds/_/Everything+Looks+Better+(In+Hindsight)
#> 35                                                https://www.last.fm/music/The+Suffers/_/Good+Day
#> 36                                                    https://www.last.fm/music/NoName/_/Yesterday
#> 37                        https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/_/All+The+Places
#> 38                                                https://www.last.fm/music/Oddisee/_/Digging+Deep
#> 39                                                https://www.last.fm/music/Leon+Bridges/_/Flowers
#> 40                                                     https://www.last.fm/music/Khalid/_/LOCATION
#> 41                           https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/_/Pop+It+in+2
#> 42                                               https://www.last.fm/music/Andrew+Bird/_/ROMA+FADE
#> 43               https://www.last.fm/music/%C3%93lafur+Arnalds/_/%C3%9E%C3%BA+ert+j%C3%B6r%C3%B0in
#> 44                                           https://www.last.fm/music/Isaiah+Rashad/_/4r+Da+Squaw
#> 45                                           https://www.last.fm/music/Open+Mike+Eagle/_/95+Radios
#> 46                                              https://www.last.fm/music/Vince+Staples/_/Big+Fish
#> 47                                            https://www.last.fm/music/Isaiah+Rashad/_/Free+Lunch
#> 48                                        https://www.last.fm/music/Leon+Bridges/_/Here+In+My+Arms
#> 49                   https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/_/Hors+d%27%C5%93uvre
#> 50                                        https://www.last.fm/music/Lisa+Hannigan/_/I+Don%27t+Know
#>    streamable.#text streamable.fulltrack            artist.name
#> 1                 0                    0           Leon Bridges
#> 2                 0                    0           Leon Bridges
#> 3                 0                    0          Lisa Hannigan
#> 4                 0                    0           Leon Bridges
#> 5                 0                    0              DJ Shadow
#> 6                 0                    0          The Lumineers
#> 7                 0                    0          The Lumineers
#> 8                 0                    0      Chance the Rapper
#> 9                 0                    0          Lisa Hannigan
#> 10                0                    0            Andrew Bird
#> 11                0                    0           Leon Bridges
#> 12                0                    0            The Suffers
#> 13                0                    0           Leon Bridges
#> 14                0                    0           Leon Bridges
#> 15                0                    0          The Lumineers
#> 16                0                    0            The Suffers
#> 17                0                    0  ::M∆DE::IN::HEIGHTS::
#> 18                0                    0            The Suffers
#> 19                0                    0          Vince Staples
#> 20                0                    0           Leon Bridges
#> 21                0                    0           Leon Bridges
#> 22                0                    0  ::M∆DE::IN::HEIGHTS::
#> 23                0                    0          Lisa Hannigan
#> 24                0                    0          The Lumineers
#> 25                0                    0          Lisa Hannigan
#> 26                0                    0            The Suffers
#> 27                0                    0 The Head and the Heart
#> 28                0                    0       Courtney Barnett
#> 29                0                    0            The Suffers
#> 30                0                    0       Courtney Barnett
#> 31                0                    0          The Lumineers
#> 32                0                    0            The Suffers
#> 33                0                    0         Run the Jewels
#> 34                0                    0         The Wild Reeds
#> 35                0                    0            The Suffers
#> 36                0                    0                 NoName
#> 37                0                    0  ::M∆DE::IN::HEIGHTS::
#> 38                0                    0                Oddisee
#> 39                0                    0           Leon Bridges
#> 40                0                    0                 Khalid
#> 41                0                    0  ::M∆DE::IN::HEIGHTS::
#> 42                0                    0            Andrew Bird
#> 43                0                    0         Ólafur Arnalds
#> 44                0                    0          Isaiah Rashad
#> 45                0                    0        Open Mike Eagle
#> 46                0                    0          Vince Staples
#> 47                0                    0          Isaiah Rashad
#> 48                0                    0           Leon Bridges
#> 49                0                    0  ::M∆DE::IN::HEIGHTS::
#> 50                0                    0          Lisa Hannigan
#>                             artist.mbid
#> 1                                      
#> 2                                      
#> 3  8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#> 4                                      
#> 5  efa2c11a-1a35-4b60-bc1b-66d37de88511
#> 6  bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 7  bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 8  373a4c98-a46b-48e4-86ec-f6ca65b4f438
#> 9  8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#> 10 081b133e-ce74-42ba-92c1-c18234acb532
#> 11                                     
#> 12 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 13                                     
#> 14                                     
#> 15 bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 16 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 17 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 18 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 19 78e854b8-9713-4ff2-9218-6b3784893bff
#> 20                                     
#> 21                                     
#> 22 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 23 8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#> 24 bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 25 8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#> 26 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 27 98cb416e-5271-484c-ab3b-6bed6b4aae12
#> 28 55111838-f001-494a-a1b5-9d818db85810
#> 29 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 30 55111838-f001-494a-a1b5-9d818db85810
#> 31 bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 32 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 33 a80c5dc5-b12e-4667-9f5a-b568961f3839
#> 34 54412058-8ed9-49a2-b931-ce8071caf851
#> 35 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 36 9393cc98-b014-493c-bdf1-a643a088f662
#> 37 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 38 aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 39                                     
#> 40 c6bfb05d-f570-46c8-98e1-e25441189770
#> 41 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 42 081b133e-ce74-42ba-92c1-c18234acb532
#> 43 6655955b-1c1e-4bcb-84e4-81bcd9efab30
#> 44 37547347-ea75-45fa-9766-ff36b44d8147
#> 45 4508a423-daaa-4d43-a839-aa1a9be88a9c
#> 46 78e854b8-9713-4ff2-9218-6b3784893bff
#> 47 37547347-ea75-45fa-9766-ff36b44d8147
#> 48                                     
#> 49 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 50 8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#>                                                 artist.url
#> 1                   https://www.last.fm/music/Leon+Bridges
#> 2                   https://www.last.fm/music/Leon+Bridges
#> 3                  https://www.last.fm/music/Lisa+Hannigan
#> 4                   https://www.last.fm/music/Leon+Bridges
#> 5                      https://www.last.fm/music/DJ+Shadow
#> 6                  https://www.last.fm/music/The+Lumineers
#> 7                  https://www.last.fm/music/The+Lumineers
#> 8              https://www.last.fm/music/Chance+the+Rapper
#> 9                  https://www.last.fm/music/Lisa+Hannigan
#> 10                   https://www.last.fm/music/Andrew+Bird
#> 11                  https://www.last.fm/music/Leon+Bridges
#> 12                   https://www.last.fm/music/The+Suffers
#> 13                  https://www.last.fm/music/Leon+Bridges
#> 14                  https://www.last.fm/music/Leon+Bridges
#> 15                 https://www.last.fm/music/The+Lumineers
#> 16                   https://www.last.fm/music/The+Suffers
#> 17 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 18                   https://www.last.fm/music/The+Suffers
#> 19                 https://www.last.fm/music/Vince+Staples
#> 20                  https://www.last.fm/music/Leon+Bridges
#> 21                  https://www.last.fm/music/Leon+Bridges
#> 22 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 23                 https://www.last.fm/music/Lisa+Hannigan
#> 24                 https://www.last.fm/music/The+Lumineers
#> 25                 https://www.last.fm/music/Lisa+Hannigan
#> 26                   https://www.last.fm/music/The+Suffers
#> 27        https://www.last.fm/music/The+Head+and+the+Heart
#> 28              https://www.last.fm/music/Courtney+Barnett
#> 29                   https://www.last.fm/music/The+Suffers
#> 30              https://www.last.fm/music/Courtney+Barnett
#> 31                 https://www.last.fm/music/The+Lumineers
#> 32                   https://www.last.fm/music/The+Suffers
#> 33                https://www.last.fm/music/Run+the+Jewels
#> 34                https://www.last.fm/music/The+Wild+Reeds
#> 35                   https://www.last.fm/music/The+Suffers
#> 36                        https://www.last.fm/music/NoName
#> 37 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 38                       https://www.last.fm/music/Oddisee
#> 39                  https://www.last.fm/music/Leon+Bridges
#> 40                        https://www.last.fm/music/Khalid
#> 41 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 42                   https://www.last.fm/music/Andrew+Bird
#> 43           https://www.last.fm/music/%C3%93lafur+Arnalds
#> 44                 https://www.last.fm/music/Isaiah+Rashad
#> 45               https://www.last.fm/music/Open+Mike+Eagle
#> 46                 https://www.last.fm/music/Vince+Staples
#> 47                 https://www.last.fm/music/Isaiah+Rashad
#> 48                  https://www.last.fm/music/Leon+Bridges
#> 49 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 50                 https://www.last.fm/music/Lisa+Hannigan
#>                                                                      image.extralarge
#> 1  https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 2  https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 3  https://lastfm-img2.akamaized.net/i/u/300x300/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 4  https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 5  https://lastfm-img2.akamaized.net/i/u/300x300/5656a077826e470596509c5a7e9a2087.png
#> 6  https://lastfm-img2.akamaized.net/i/u/300x300/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 7  https://lastfm-img2.akamaized.net/i/u/300x300/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 8  https://lastfm-img2.akamaized.net/i/u/300x300/b5670f1ef6be2b1f784752d139494419.png
#> 9  https://lastfm-img2.akamaized.net/i/u/300x300/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 10 https://lastfm-img2.akamaized.net/i/u/300x300/5f879adfe74f40fc8af223cc68da341e.png
#> 11 https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 12 https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 13 https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 14 https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 15 https://lastfm-img2.akamaized.net/i/u/300x300/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 16 https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 17 https://lastfm-img2.akamaized.net/i/u/300x300/eaa53023be9574ffe48d80ae4d51d37f.png
#> 18 https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 19 https://lastfm-img2.akamaized.net/i/u/300x300/520852e6f27411710da9eaa8721f7514.png
#> 20 https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 21 https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 22 https://lastfm-img2.akamaized.net/i/u/300x300/eaa53023be9574ffe48d80ae4d51d37f.png
#> 23 https://lastfm-img2.akamaized.net/i/u/300x300/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 24 https://lastfm-img2.akamaized.net/i/u/300x300/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 25 https://lastfm-img2.akamaized.net/i/u/300x300/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 26 https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 27 https://lastfm-img2.akamaized.net/i/u/300x300/2c50aec4065c4947be519bd11f15aba8.png
#> 28 https://lastfm-img2.akamaized.net/i/u/300x300/cbdbd65e390ae988adccd9c91952b693.png
#> 29 https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 30 https://lastfm-img2.akamaized.net/i/u/300x300/cbdbd65e390ae988adccd9c91952b693.png
#> 31 https://lastfm-img2.akamaized.net/i/u/300x300/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 32 https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 33 https://lastfm-img2.akamaized.net/i/u/300x300/c9a4052e724b4211c2dd2c1be4b09a6c.png
#> 34 https://lastfm-img2.akamaized.net/i/u/300x300/85f1a1719c316244d4a7740c42246da0.png
#> 35 https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 36 https://lastfm-img2.akamaized.net/i/u/300x300/dbe58a33f97279ad7a6037e67d7f43f6.png
#> 37 https://lastfm-img2.akamaized.net/i/u/300x300/eaa53023be9574ffe48d80ae4d51d37f.png
#> 38 https://lastfm-img2.akamaized.net/i/u/300x300/b4cd2591df054c5e96f448eea98ae95e.png
#> 39 https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 40 https://lastfm-img2.akamaized.net/i/u/300x300/5fa19194d7f4daf3ad1649dd930c7454.png
#> 41 https://lastfm-img2.akamaized.net/i/u/300x300/eaa53023be9574ffe48d80ae4d51d37f.png
#> 42 https://lastfm-img2.akamaized.net/i/u/300x300/5f879adfe74f40fc8af223cc68da341e.png
#> 43 https://lastfm-img2.akamaized.net/i/u/300x300/b1dcb02ddd420147d569a6a4d846424c.png
#> 44 https://lastfm-img2.akamaized.net/i/u/300x300/ec3984e07b02169b9d6bf7d8dfdb3ff3.png
#> 45 https://lastfm-img2.akamaized.net/i/u/300x300/9cd6d39ffb570806a0779c88817f74eb.png
#> 46 https://lastfm-img2.akamaized.net/i/u/300x300/520852e6f27411710da9eaa8721f7514.png
#> 47 https://lastfm-img2.akamaized.net/i/u/300x300/ec3984e07b02169b9d6bf7d8dfdb3ff3.png
#> 48 https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 49 https://lastfm-img2.akamaized.net/i/u/300x300/eaa53023be9574ffe48d80ae4d51d37f.png
#> 50 https://lastfm-img2.akamaized.net/i/u/300x300/9a87d69ae0aa4d668f98e559ac7725e5.png
#>                                                                        image.large
#> 1  https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 2  https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 3  https://lastfm-img2.akamaized.net/i/u/174s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 4  https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 5  https://lastfm-img2.akamaized.net/i/u/174s/5656a077826e470596509c5a7e9a2087.png
#> 6  https://lastfm-img2.akamaized.net/i/u/174s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 7  https://lastfm-img2.akamaized.net/i/u/174s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 8  https://lastfm-img2.akamaized.net/i/u/174s/b5670f1ef6be2b1f784752d139494419.png
#> 9  https://lastfm-img2.akamaized.net/i/u/174s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 10 https://lastfm-img2.akamaized.net/i/u/174s/5f879adfe74f40fc8af223cc68da341e.png
#> 11 https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 12 https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 13 https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 14 https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 15 https://lastfm-img2.akamaized.net/i/u/174s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 16 https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 17 https://lastfm-img2.akamaized.net/i/u/174s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 18 https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 19 https://lastfm-img2.akamaized.net/i/u/174s/520852e6f27411710da9eaa8721f7514.png
#> 20 https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 21 https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 22 https://lastfm-img2.akamaized.net/i/u/174s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 23 https://lastfm-img2.akamaized.net/i/u/174s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 24 https://lastfm-img2.akamaized.net/i/u/174s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 25 https://lastfm-img2.akamaized.net/i/u/174s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 26 https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 27 https://lastfm-img2.akamaized.net/i/u/174s/2c50aec4065c4947be519bd11f15aba8.png
#> 28 https://lastfm-img2.akamaized.net/i/u/174s/cbdbd65e390ae988adccd9c91952b693.png
#> 29 https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 30 https://lastfm-img2.akamaized.net/i/u/174s/cbdbd65e390ae988adccd9c91952b693.png
#> 31 https://lastfm-img2.akamaized.net/i/u/174s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 32 https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 33 https://lastfm-img2.akamaized.net/i/u/174s/c9a4052e724b4211c2dd2c1be4b09a6c.png
#> 34 https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png
#> 35 https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 36 https://lastfm-img2.akamaized.net/i/u/174s/dbe58a33f97279ad7a6037e67d7f43f6.png
#> 37 https://lastfm-img2.akamaized.net/i/u/174s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 38 https://lastfm-img2.akamaized.net/i/u/174s/b4cd2591df054c5e96f448eea98ae95e.png
#> 39 https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 40 https://lastfm-img2.akamaized.net/i/u/174s/5fa19194d7f4daf3ad1649dd930c7454.png
#> 41 https://lastfm-img2.akamaized.net/i/u/174s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 42 https://lastfm-img2.akamaized.net/i/u/174s/5f879adfe74f40fc8af223cc68da341e.png
#> 43 https://lastfm-img2.akamaized.net/i/u/174s/b1dcb02ddd420147d569a6a4d846424c.png
#> 44 https://lastfm-img2.akamaized.net/i/u/174s/ec3984e07b02169b9d6bf7d8dfdb3ff3.png
#> 45 https://lastfm-img2.akamaized.net/i/u/174s/9cd6d39ffb570806a0779c88817f74eb.png
#> 46 https://lastfm-img2.akamaized.net/i/u/174s/520852e6f27411710da9eaa8721f7514.png
#> 47 https://lastfm-img2.akamaized.net/i/u/174s/ec3984e07b02169b9d6bf7d8dfdb3ff3.png
#> 48 https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 49 https://lastfm-img2.akamaized.net/i/u/174s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 50 https://lastfm-img2.akamaized.net/i/u/174s/9a87d69ae0aa4d668f98e559ac7725e5.png
#>                                                                      image.medium
#> 1  https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 2  https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 3  https://lastfm-img2.akamaized.net/i/u/64s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 4  https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 5  https://lastfm-img2.akamaized.net/i/u/64s/5656a077826e470596509c5a7e9a2087.png
#> 6  https://lastfm-img2.akamaized.net/i/u/64s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 7  https://lastfm-img2.akamaized.net/i/u/64s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 8  https://lastfm-img2.akamaized.net/i/u/64s/b5670f1ef6be2b1f784752d139494419.png
#> 9  https://lastfm-img2.akamaized.net/i/u/64s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 10 https://lastfm-img2.akamaized.net/i/u/64s/5f879adfe74f40fc8af223cc68da341e.png
#> 11 https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 12 https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 13 https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 14 https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 15 https://lastfm-img2.akamaized.net/i/u/64s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 16 https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 17 https://lastfm-img2.akamaized.net/i/u/64s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 18 https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 19 https://lastfm-img2.akamaized.net/i/u/64s/520852e6f27411710da9eaa8721f7514.png
#> 20 https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 21 https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 22 https://lastfm-img2.akamaized.net/i/u/64s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 23 https://lastfm-img2.akamaized.net/i/u/64s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 24 https://lastfm-img2.akamaized.net/i/u/64s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 25 https://lastfm-img2.akamaized.net/i/u/64s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 26 https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 27 https://lastfm-img2.akamaized.net/i/u/64s/2c50aec4065c4947be519bd11f15aba8.png
#> 28 https://lastfm-img2.akamaized.net/i/u/64s/cbdbd65e390ae988adccd9c91952b693.png
#> 29 https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 30 https://lastfm-img2.akamaized.net/i/u/64s/cbdbd65e390ae988adccd9c91952b693.png
#> 31 https://lastfm-img2.akamaized.net/i/u/64s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 32 https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 33 https://lastfm-img2.akamaized.net/i/u/64s/c9a4052e724b4211c2dd2c1be4b09a6c.png
#> 34 https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png
#> 35 https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 36 https://lastfm-img2.akamaized.net/i/u/64s/dbe58a33f97279ad7a6037e67d7f43f6.png
#> 37 https://lastfm-img2.akamaized.net/i/u/64s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 38 https://lastfm-img2.akamaized.net/i/u/64s/b4cd2591df054c5e96f448eea98ae95e.png
#> 39 https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 40 https://lastfm-img2.akamaized.net/i/u/64s/5fa19194d7f4daf3ad1649dd930c7454.png
#> 41 https://lastfm-img2.akamaized.net/i/u/64s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 42 https://lastfm-img2.akamaized.net/i/u/64s/5f879adfe74f40fc8af223cc68da341e.png
#> 43 https://lastfm-img2.akamaized.net/i/u/64s/b1dcb02ddd420147d569a6a4d846424c.png
#> 44 https://lastfm-img2.akamaized.net/i/u/64s/ec3984e07b02169b9d6bf7d8dfdb3ff3.png
#> 45 https://lastfm-img2.akamaized.net/i/u/64s/9cd6d39ffb570806a0779c88817f74eb.png
#> 46 https://lastfm-img2.akamaized.net/i/u/64s/520852e6f27411710da9eaa8721f7514.png
#> 47 https://lastfm-img2.akamaized.net/i/u/64s/ec3984e07b02169b9d6bf7d8dfdb3ff3.png
#> 48 https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 49 https://lastfm-img2.akamaized.net/i/u/64s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 50 https://lastfm-img2.akamaized.net/i/u/64s/9a87d69ae0aa4d668f98e559ac7725e5.png
#>                                                                       image.small
#> 1  https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 2  https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 3  https://lastfm-img2.akamaized.net/i/u/34s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 4  https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 5  https://lastfm-img2.akamaized.net/i/u/34s/5656a077826e470596509c5a7e9a2087.png
#> 6  https://lastfm-img2.akamaized.net/i/u/34s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 7  https://lastfm-img2.akamaized.net/i/u/34s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 8  https://lastfm-img2.akamaized.net/i/u/34s/b5670f1ef6be2b1f784752d139494419.png
#> 9  https://lastfm-img2.akamaized.net/i/u/34s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 10 https://lastfm-img2.akamaized.net/i/u/34s/5f879adfe74f40fc8af223cc68da341e.png
#> 11 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 12 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 13 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 14 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 15 https://lastfm-img2.akamaized.net/i/u/34s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 16 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 17 https://lastfm-img2.akamaized.net/i/u/34s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 18 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 19 https://lastfm-img2.akamaized.net/i/u/34s/520852e6f27411710da9eaa8721f7514.png
#> 20 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 21 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 22 https://lastfm-img2.akamaized.net/i/u/34s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 23 https://lastfm-img2.akamaized.net/i/u/34s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 24 https://lastfm-img2.akamaized.net/i/u/34s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 25 https://lastfm-img2.akamaized.net/i/u/34s/9a87d69ae0aa4d668f98e559ac7725e5.png
#> 26 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 27 https://lastfm-img2.akamaized.net/i/u/34s/2c50aec4065c4947be519bd11f15aba8.png
#> 28 https://lastfm-img2.akamaized.net/i/u/34s/cbdbd65e390ae988adccd9c91952b693.png
#> 29 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 30 https://lastfm-img2.akamaized.net/i/u/34s/cbdbd65e390ae988adccd9c91952b693.png
#> 31 https://lastfm-img2.akamaized.net/i/u/34s/a42a15b3b8914afdaefa8f8f1e1c9263.png
#> 32 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 33 https://lastfm-img2.akamaized.net/i/u/34s/c9a4052e724b4211c2dd2c1be4b09a6c.png
#> 34 https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png
#> 35 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png
#> 36 https://lastfm-img2.akamaized.net/i/u/34s/dbe58a33f97279ad7a6037e67d7f43f6.png
#> 37 https://lastfm-img2.akamaized.net/i/u/34s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 38 https://lastfm-img2.akamaized.net/i/u/34s/b4cd2591df054c5e96f448eea98ae95e.png
#> 39 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 40 https://lastfm-img2.akamaized.net/i/u/34s/5fa19194d7f4daf3ad1649dd930c7454.png
#> 41 https://lastfm-img2.akamaized.net/i/u/34s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 42 https://lastfm-img2.akamaized.net/i/u/34s/5f879adfe74f40fc8af223cc68da341e.png
#> 43 https://lastfm-img2.akamaized.net/i/u/34s/b1dcb02ddd420147d569a6a4d846424c.png
#> 44 https://lastfm-img2.akamaized.net/i/u/34s/ec3984e07b02169b9d6bf7d8dfdb3ff3.png
#> 45 https://lastfm-img2.akamaized.net/i/u/34s/9cd6d39ffb570806a0779c88817f74eb.png
#> 46 https://lastfm-img2.akamaized.net/i/u/34s/520852e6f27411710da9eaa8721f7514.png
#> 47 https://lastfm-img2.akamaized.net/i/u/34s/ec3984e07b02169b9d6bf7d8dfdb3ff3.png
#> 48 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png
#> 49 https://lastfm-img2.akamaized.net/i/u/34s/eaa53023be9574ffe48d80ae4d51d37f.png
#> 50 https://lastfm-img2.akamaized.net/i/u/34s/9a87d69ae0aa4d668f98e559ac7725e5.png
#>    @attr.rank
#> 1           1
#> 2           2
#> 3           3
#> 4           4
#> 5           5
#> 6           6
#> 7           7
#> 8           8
#> 9           9
#> 10         10
#> 11         11
#> 12         12
#> 13         13
#> 14         14
#> 15         15
#> 16         16
#> 17         17
#> 18         18
#> 19         19
#> 20         20
#> 21         21
#> 22         22
#> 23         23
#> 24         24
#> 25         25
#> 26         26
#> 27         27
#> 28         28
#> 29         29
#> 30         30
#> 31         31
#> 32         32
#> 33         33
#> 34         34
#> 35         35
#> 36         36
#> 37         37
#> 38         38
#> 39         39
#> 40         40
#> 41         41
#> 42         42
#> 43         43
#> 44         44
#> 45         45
#> 46         46
#> 47         47
#> 48         48
#> 49         49
#> 50         50