Get an album chart for a user profile.

Get an album chart for a user profile, for a given date range. If no date range is supplied, it will return the most recent album chart for this user. Implementation of last.fm's user.getWeeklyAlbumChart API method (http://www.last.fm/api/show/user.getWeeklyAlbumChart).

user_getWeeklyAlbumChart(user, from = NA, to = NA, api_key = lastkey)

Arguments

user
The last.fm username to fetch the charts of.

from
The date at which the chart should start from. See user_getChartsList for more.

to
The date at which the chart should end on. See user_getChartsList for more.

api_key
A Last.fm API key.

Value

A list of an album chart for a user profile.

Examples

user_getWeeklyAlbumChart("platyjus")

#> $album
#>       artist.#text                          artist.mbid
#> 1   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 2     Joan Shelley                                     
#> 3 Childish Gambino 7fb57fba-a6ef-44c2-abab-2fa3bdee607e
#> 4     Joan Shelley                                     
#> 5          Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 6        Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993
#> 7           Khalid c6bfb05d-f570-46c8-98e1-e25441189770
#> 8        Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993
#> 9        Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993
#>                                               name
#> 1                               The World We Built
#> 2                                     Joan Shelley
#> 3                               "Awaken, My Love!"
#> 4                                Over and Even + 3
#> 5                                      The Iceberg
#> 6 30 Greatest Hits: Portrait of a Legend 1951-1964
#> 7                                    American Teen
#> 8                Ain't That Good News (Remastered)
#> 9                                        Swing Low
#>                                   mbid playcount
#> 1                                             19
#> 2                                              5
#> 3                                              4
#> 4                                              4
#> 5                                              4
#> 6                                              2
#> 7                                              2
#> 8                                              1
#> 9 8ac8fd98-f1eb-4049-8cea-9be6e2268272         1
#>                                                                                    url
#> 1                          https://www.last.fm/music/The+Wild+Reeds/The+World+We+Built
#> 2                                  https://www.last.fm/music/Joan+Shelley/Joan+Shelley
#> 3                  https://www.last.fm/music/Childish+Gambino/%22Awaken,+My+Love%21%22
#> 4                         https://www.last.fm/music/Joan+Shelley/Over+and+Even+%252B+3
#> 5                                        https://www.last.fm/music/Oddisee/The+Iceberg
#> 6 https://www.last.fm/music/Sam+Cooke/30+Greatest+Hits:+Portrait+of+a+Legend+1951-1964
#> 7                                       https://www.last.fm/music/Khalid/American+Teen
#> 8              https://www.last.fm/music/Sam+Cooke/Ain%27t+That+Good+News+(Remastered)
#> 9                                        https://www.last.fm/music/Sam+Cooke/Swing+Low
#>   rank
#> 1    1
#> 2    2
#> 3    3
#> 4    4
#> 5    5
#> 6    6
#> 7    7
#> 8    8
#> 9    9
#> 
#> $`@attr`
#> $`@attr`$user
#> [1] "platyjus"
#> 
#> $`@attr`$from
#> [1] "1494504000"
#> 
#> $`@attr`$to
#> [1] "1495108800"
#> 
#>