Get an artist chart for a user profile.

Get an artist chart for a user profile, for a given date range. If no date range is supplied, it will return the most recent artist chart for this user. Implementation of last.fm's user.getWeeklyArtistChart API method (http://www.last.fm/api/show/user.getWeeklyArtistChart).

user_getWeeklyArtistChart(user, from = NA, to = NA)

Arguments

user
The last.fm username to fetch the charts of.

from
The date at which the chart should start from.

to
The date at which the chart should end on.

Value

A list of an artist chart for a user profile.

Examples

user_getWeeklyArtistChart("platyjus")

#>                                 name                                 mbid
#> 1                            Wye Oak 3c8ff4c9-b5b9-4e78-ad9d-5157c323de24
#> 2                           Jay Rock 9880800a-f6f8-4f8f-a00e-b4b664776c0c
#> 3                       Moses Sumney 89c081d4-2ab2-4d3e-8589-ad77dfc40384
#> 4                         Feng Suave                                     
#> 5                 Natalia Lafourcade 2ceb4e66-4eaa-4dba-ad3a-30df3b742557
#> 6                               Saba c2311581-df24-451e-8232-a9dcee3bfa62
#> 7                            Ab-Soul acc12019-10e3-40aa-b4bf-66fbb4a692af
#> 8                       Big K.R.I.T. 7e966957-9b0b-4e19-b8a0-4206535d8052
#> 9                       Blood Orange e43daa78-3cf3-4cf9-bc27-af4d4e675a02
#> 10                     Isaiah Rashad 37547347-ea75-45fa-9766-ff36b44d8147
#> 11                        King Krule c0daae21-ec2a-4dd0-a323-4f53f5945d04
#> 12                     Little Dragon c8a9ae11-63bf-409b-8bf3-a96b3712a2ff
#> 13                      Mick Jenkins 64c4dfcc-ba2b-4036-ac01-1636ae15ebb8
#> 14                           Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 15                   Open Mike Eagle 4508a423-daaa-4d43-a839-aa1a9be88a9c
#> 16                            Sampha 130c2184-14f2-4180-a216-04b5a333be57
#> 17                             Sango 150aacdf-a15f-46b8-8a99-25cde68e90de
#> 18                         Sevdaliza                                     
#> 19                               Sir 96966179-d9ef-46f4-9179-5545cd0e6279
#> 20                        Thundercat 044fd265-79dd-43eb-afc4-8b20becf7e17
#> 21                     Vince Staples 78e854b8-9713-4ff2-9218-6b3784893bff
#> 22                    Anderson .Paak                                     
#> 23                     BADBADNOTGOOD 754294d5-d7d2-4ea2-8184-1dcaaf55a56f
#> 24                            Barrie                                     
#> 25              Benjamin Lazar Davis                                     
#> 26                      Bishop Nehru 0735e073-b829-4c10-9144-8ca286ec68b2
#> 27 Brian Blade & The Fellowship Band 4d54e19b-f78d-4d2b-bdcd-a8d0348d6e6f
#> 28                     Cornelia Murr                                     
#> 29                          Czarface 66120159-5fe7-4cdb-8c9a-7ac68e393c6f
#> 30                    Denmark Vessey                                     
#> 31                          Ed Mount                                     
#> 32                     Freddie Gibbs 21645c31-fe1c-45a4-955c-3e172b12c3f9
#> 33            Gabriel Garzón-Montano 4cc212e1-0c06-4675-9c8b-bf6c4fb05769
#> 34                     Hugh Masekela 83a2e0fb-3d9b-46ff-b383-3dde54c15297
#> 35                       James Blake 8dc08b1f-e393-4f85-a5dd-300f7693a8b8
#> 36                      Jamila Woods                                     
#> 37                          Jonwayne 70bd369d-f58a-48ab-96eb-299d9b89a3c7
#> 38                       JUNGLEPUSSY 85f40e9b-c639-4783-b621-80d873615e04
#> 39                        Kali Uchis                                     
#> 40                 Kamasi Washington                                     
#> 41                            Kelela 3599cf21-2ed3-405d-a1f1-b52ff7b28df8
#> 42                    Kevin Abstract                                     
#> 43                     Kevin Krauter                                     
#> 44                         Knxwledge 7df4df53-17ba-4099-90ff-c5528762c5a7
#> 45                      Les Louanges                                     
#> 46                         Lion Babe c9dda914-1eb2-4762-8006-208010b89ee3
#> 47                          Lykke Li cc7d4686-ea02-45fd-956e-94c1a322558c
#> 48                             Matty                                     
#> 49                       Monophonics f1745cca-1cae-4696-a0d3-62553c742b0c
#> 50                  Morabeza Tobacco                                     
#> 51                 Nitai Hershkovits                                     
#> 52                            NoName 9393cc98-b014-493c-bdf1-a643a088f662
#> 53                      Nubya Garcia                                     
#> 54                  Olivier St.Louis                                     
#> 55                         Orion Sun                                     
#> 56                             Ought a039f1f1-e664-4aee-a917-426514734aa0
#> 57                          Palatine                                     
#> 58                  Patrick Paige II                                     
#> 59                       Pomplamoose 63d61461-ecfe-4057-896b-3ba3ff3111eb
#> 60         Robert Glasper Experiment 0ec3c9dc-45e7-4bc1-9813-a866acaae772
#> 61                      Snoh Aalegra 636e0fb6-f2ad-45fc-b385-4d994665e7e3
#> 62                     Sons of Kemet fee950a2-e68c-4e76-b1ae-ff6efdf68b11
#> 63             Sunny & The Sunliners                                     
#> 64         The California Honeydrops c144dbec-b0f1-4aae-b2e6-5087e44806c8
#> 65                      The Internet 45804465-4271-4e6a-b881-ce668ef09530
#> 66                  The Sea and Cake 74171cb6-1ff8-42bd-be8a-96660c22e3e0
#> 67                         The Voidz                                     
#> 68                            Tirzah f697bed6-3b4e-4352-b6fb-70550afd8dcb
#> 69                        Umi Copper                                     
#> 70                    Ólafur Arnalds 6655955b-1c1e-4bcb-84e4-81bcd9efab30
#>    playcount                                                         url
#> 1         24                           https://www.last.fm/music/Wye+Oak
#> 2         18                          https://www.last.fm/music/Jay+Rock
#> 3         10                      https://www.last.fm/music/Moses+Sumney
#> 4          9                        https://www.last.fm/music/Feng+Suave
#> 5          6                https://www.last.fm/music/Natalia+Lafourcade
#> 6          3                              https://www.last.fm/music/Saba
#> 7          2                           https://www.last.fm/music/Ab-Soul
#> 8          2                      https://www.last.fm/music/Big+K.R.I.T.
#> 9          2                      https://www.last.fm/music/Blood+Orange
#> 10         2                     https://www.last.fm/music/Isaiah+Rashad
#> 11         2                        https://www.last.fm/music/King+Krule
#> 12         2                     https://www.last.fm/music/Little+Dragon
#> 13         2                      https://www.last.fm/music/Mick+Jenkins
#> 14         2                           https://www.last.fm/music/Oddisee
#> 15         2                   https://www.last.fm/music/Open+Mike+Eagle
#> 16         2                            https://www.last.fm/music/Sampha
#> 17         2                             https://www.last.fm/music/Sango
#> 18         2                         https://www.last.fm/music/Sevdaliza
#> 19         2                               https://www.last.fm/music/Sir
#> 20         2                        https://www.last.fm/music/Thundercat
#> 21         2                     https://www.last.fm/music/Vince+Staples
#> 22         1                    https://www.last.fm/music/Anderson+.Paak
#> 23         1                     https://www.last.fm/music/BADBADNOTGOOD
#> 24         1                            https://www.last.fm/music/Barrie
#> 25         1              https://www.last.fm/music/Benjamin+Lazar+Davis
#> 26         1                      https://www.last.fm/music/Bishop+Nehru
#> 27         1 https://www.last.fm/music/Brian+Blade+&+The+Fellowship+Band
#> 28         1                     https://www.last.fm/music/Cornelia+Murr
#> 29         1                          https://www.last.fm/music/Czarface
#> 30         1                    https://www.last.fm/music/Denmark+Vessey
#> 31         1                          https://www.last.fm/music/Ed+Mount
#> 32         1                     https://www.last.fm/music/Freddie+Gibbs
#> 33         1       https://www.last.fm/music/Gabriel+Garz%C3%B3n-Montano
#> 34         1                     https://www.last.fm/music/Hugh+Masekela
#> 35         1                       https://www.last.fm/music/James+Blake
#> 36         1                      https://www.last.fm/music/Jamila+Woods
#> 37         1                          https://www.last.fm/music/Jonwayne
#> 38         1                       https://www.last.fm/music/JUNGLEPUSSY
#> 39         1                        https://www.last.fm/music/Kali+Uchis
#> 40         1                 https://www.last.fm/music/Kamasi+Washington
#> 41         1                            https://www.last.fm/music/Kelela
#> 42         1                    https://www.last.fm/music/Kevin+Abstract
#> 43         1                     https://www.last.fm/music/Kevin+Krauter
#> 44         1                         https://www.last.fm/music/Knxwledge
#> 45         1                      https://www.last.fm/music/Les+Louanges
#> 46         1                         https://www.last.fm/music/Lion+Babe
#> 47         1                          https://www.last.fm/music/Lykke+Li
#> 48         1                             https://www.last.fm/music/Matty
#> 49         1                       https://www.last.fm/music/Monophonics
#> 50         1                  https://www.last.fm/music/Morabeza+Tobacco
#> 51         1                 https://www.last.fm/music/Nitai+Hershkovits
#> 52         1                            https://www.last.fm/music/NoName
#> 53         1                      https://www.last.fm/music/Nubya+Garcia
#> 54         1                  https://www.last.fm/music/Olivier+St.Louis
#> 55         1                         https://www.last.fm/music/Orion+Sun
#> 56         1                             https://www.last.fm/music/Ought
#> 57         1                          https://www.last.fm/music/Palatine
#> 58         1                  https://www.last.fm/music/Patrick+Paige+II
#> 59         1                       https://www.last.fm/music/Pomplamoose
#> 60         1         https://www.last.fm/music/Robert+Glasper+Experiment
#> 61         1                      https://www.last.fm/music/Snoh+Aalegra
#> 62         1                     https://www.last.fm/music/Sons+of+Kemet
#> 63         1             https://www.last.fm/music/Sunny+&+The+Sunliners
#> 64         1         https://www.last.fm/music/The+California+Honeydrops
#> 65         1                      https://www.last.fm/music/The+Internet
#> 66         1                  https://www.last.fm/music/The+Sea+and+Cake
#> 67         1                         https://www.last.fm/music/The+Voidz
#> 68         1                            https://www.last.fm/music/Tirzah
#> 69         1                        https://www.last.fm/music/Umi+Copper
#> 70         1               https://www.last.fm/music/%C3%93lafur+Arnalds
#>    @attr.rank
#> 1           1
#> 2           2
#> 3           3
#> 4           4
#> 5           5
#> 6           6
#> 7           7
#> 8           8
#> 9           9
#> 10         10
#> 11         11
#> 12         12
#> 13         13
#> 14         14
#> 15         15
#> 16         16
#> 17         17
#> 18         18
#> 19         19
#> 20         20
#> 21         21
#> 22         22
#> 23         23
#> 24         24
#> 25         25
#> 26         26
#> 27         27
#> 28         28
#> 29         29
#> 30         30
#> 31         31
#> 32         32
#> 33         33
#> 34         34
#> 35         35
#> 36         36
#> 37         37
#> 38         38
#> 39         39
#> 40         40
#> 41         41
#> 42         42
#> 43         43
#> 44         44
#> 45         45
#> 46         46
#> 47         47
#> 48         48
#> 49         49
#> 50         50
#> 51         51
#> 52         52
#> 53         53
#> 54         54
#> 55         55
#> 56         56
#> 57         57
#> 58         58
#> 59         59
#> 60         60
#> 61         61
#> 62         62
#> 63         63
#> 64         64
#> 65         65
#> 66         66
#> 67         67
#> 68         68
#> 69         69
#> 70         70