Get a list of available charts for this user.

Get a list of available charts for this user, expressed as date ranges which can be sent to the chart services. Implementation of last.fm's user.getWeeklyChartList API method (http://www.last.fm/api/show/user.getWeeklyChartList).

user_getWeeklyChartList(user, api_key = lastkey)

Arguments

user
TThe last.fm username to fetch the charts list for.

api_key
A Last.fm API key.

Value

A list of an artist chart for a user profile.

Examples

user_getWeeklyChartList("platyjus")

#> $chart
#>     #text       from         to
#> 1         1108296000 1108900800
#> 2         1108900800 1109505600
#> 3         1109505600 1110110400
#> 4         1110110400 1110715200
#> 5         1110715200 1111320000
#> 6         1111320000 1111924800
#> 7         1111924800 1112529600
#> 8         1112529600 1113134400
#> 9         1113134400 1113739200
#> 10        1113739200 1114344000
#> 11        1114344000 1114948800
#> 12        1114948800 1115553600
#> 13        1115553600 1116158400
#> 14        1116158400 1116763200
#> 15        1116763200 1117368000
#> 16        1117368000 1117972800
#> 17        1117972800 1118577600
#> 18        1118577600 1119182400
#> 19        1119182400 1119787200
#> 20        1119787200 1120392000
#> 21        1120392000 1120996800
#> 22        1120996800 1121601600
#> 23        1121601600 1122206400
#> 24        1122206400 1122811200
#> 25        1122811200 1123416000
#> 26        1123416000 1124020800
#> 27        1124020800 1124625600
#> 28        1124625600 1125230400
#> 29        1125230400 1125835200
#> 30        1125835200 1126440000
#> 31        1126440000 1127044800
#> 32        1127044800 1127649600
#> 33        1127649600 1128254400
#> 34        1128254400 1128859200
#> 35        1128859200 1129464000
#> 36        1129464000 1130068800
#> 37        1130068800 1130673600
#> 38        1130673600 1131278400
#> 39        1131278400 1131883200
#> 40        1131883200 1132488000
#> 41        1132488000 1133092800
#> 42        1133092800 1133697600
#> 43        1133697600 1134302400
#> 44        1134302400 1134907200
#> 45        1134907200 1135512000
#> 46        1135512000 1136116800
#> 47        1136116800 1136721600
#> 48        1136721600 1137326400
#> 49        1137326400 1137931200
#> 50        1137931200 1138536000
#> 51        1138536000 1139140800
#> 52        1139140800 1139745600
#> 53        1139745600 1140350400
#> 54        1140350400 1140955200
#> 55        1140955200 1141560000
#> 56        1141560000 1142164800
#> 57        1142164800 1142769600
#> 58        1142769600 1143374400
#> 59        1143374400 1143979200
#> 60        1143979200 1144584000
#> 61        1144584000 1145188800
#> 62        1145188800 1145793600
#> 63        1145793600 1146398400
#> 64        1146398400 1147003200
#> 65        1147003200 1147608000
#> 66        1147608000 1148212800
#> 67        1148212800 1148817600
#> 68        1148817600 1149422400
#> 69        1149422400 1150027200
#> 70        1150027200 1150632000
#> 71        1150632000 1151236800
#> 72        1151236800 1151841600
#> 73        1151841600 1152446400
#> 74        1152446400 1153051200
#> 75        1153051200 1153656000
#> 76        1153656000 1154260800
#> 77        1154260800 1154865600
#> 78        1154865600 1155470400
#> 79        1155470400 1156075200
#> 80        1156075200 1156680000
#> 81        1156680000 1157284800
#> 82        1157284800 1157889600
#> 83        1157889600 1158494400
#> 84        1158494400 1159099200
#> 85        1159099200 1159704000
#> 86        1159704000 1160308800
#> 87        1160308800 1160913600
#> 88        1160913600 1161518400
#> 89        1161518400 1162123200
#> 90        1162123200 1162728000
#> 91        1162728000 1163332800
#> 92        1163332800 1163937600
#> 93        1163937600 1164542400
#> 94        1164542400 1165147200
#> 95        1165147200 1165752000
#> 96        1165752000 1166356800
#> 97        1166356800 1166961600
#> 98        1166961600 1167566400
#> 99        1167566400 1168171200
#> 100       1168171200 1168776000
#> 101       1168776000 1169380800
#> 102       1169380800 1169985600
#> 103       1169985600 1170590400
#> 104       1170590400 1171195200
#> 105       1171195200 1171800000
#> 106       1171800000 1172404800
#> 107       1172404800 1173009600
#> 108       1173009600 1173614400
#> 109       1173614400 1174219200
#> 110       1174219200 1174824000
#> 111       1174824000 1175428800
#> 112       1175428800 1176033600
#> 113       1176033600 1176638400
#> 114       1176638400 1177243200
#> 115       1177243200 1177848000
#> 116       1177848000 1178452800
#> 117       1178452800 1179057600
#> 118       1179057600 1179662400
#> 119       1179662400 1180267200
#> 120       1180267200 1180872000
#> 121       1180872000 1181476800
#> 122       1181476800 1182081600
#> 123       1182081600 1182686400
#> 124       1182686400 1183291200
#> 125       1183291200 1183896000
#> 126       1183896000 1184500800
#> 127       1184500800 1185105600
#> 128       1185105600 1185710400
#> 129       1185710400 1186315200
#> 130       1186315200 1186920000
#> 131       1186920000 1187524800
#> 132       1187524800 1188129600
#> 133       1188129600 1188734400
#> 134       1188734400 1189339200
#> 135       1189339200 1189944000
#> 136       1189944000 1190548800
#> 137       1190548800 1191153600
#> 138       1191153600 1191758400
#> 139       1191758400 1192363200
#> 140       1192363200 1192968000
#> 141       1192968000 1193572800
#> 142       1193572800 1194177600
#> 143       1194177600 1194782400
#> 144       1194782400 1195387200
#> 145       1195387200 1195992000
#> 146       1195992000 1196596800
#> 147       1196596800 1197201600
#> 148       1197201600 1197806400
#> 149       1197806400 1198411200
#> 150       1198411200 1199016000
#> 151       1199016000 1199620800
#> 152       1199620800 1200225600
#> 153       1200225600 1200830400
#> 154       1200830400 1201435200
#> 155       1201435200 1202040000
#> 156       1202040000 1202644800
#> 157       1202644800 1203249600
#> 158       1203249600 1203854400
#> 159       1203854400 1204459200
#> 160       1204459200 1205064000
#> 161       1205064000 1205668800
#> 162       1205668800 1206273600
#> 163       1206273600 1206878400
#> 164       1206878400 1207483200
#> 165       1207483200 1208088000
#> 166       1208088000 1208692800
#> 167       1208692800 1209297600
#> 168       1209297600 1209902400
#> 169       1209902400 1210507200
#> 170       1210507200 1211112000
#> 171       1211112000 1211716800
#> 172       1211716800 1212321600
#> 173       1212321600 1212926400
#> 174       1212926400 1213531200
#> 175       1213531200 1214136000
#> 176       1214136000 1214740800
#> 177       1214740800 1215345600
#> 178       1215345600 1215950400
#> 179       1215950400 1216555200
#> 180       1216555200 1217160000
#> 181       1217160000 1217764800
#> 182       1217764800 1218369600
#> 183       1218369600 1218974400
#> 184       1218974400 1219579200
#> 185       1219579200 1220184000
#> 186       1220184000 1220788800
#> 187       1220788800 1221393600
#> 188       1221393600 1221998400
#> 189       1221998400 1222603200
#> 190       1222603200 1223208000
#> 191       1223208000 1223812800
#> 192       1223812800 1224417600
#> 193       1224417600 1225022400
#> 194       1225022400 1225627200
#> 195       1225627200 1226232000
#> 196       1226232000 1226836800
#> 197       1226836800 1227441600
#> 198       1227441600 1228046400
#> 199       1228046400 1228651200
#> 200       1228651200 1229256000
#> 201       1229256000 1229860800
#> 202       1229860800 1230465600
#> 203       1230465600 1231070400
#> 204       1231070400 1231675200
#> 205       1231675200 1232280000
#> 206       1232280000 1232884800
#> 207       1232884800 1233489600
#> 208       1233489600 1234094400
#> 209       1234094400 1234699200
#> 210       1234699200 1235304000
#> 211       1235304000 1235908800
#> 212       1235908800 1236513600
#> 213       1236513600 1237118400
#> 214       1237118400 1237723200
#> 215       1237723200 1238328000
#> 216       1238328000 1238932800
#> 217       1238932800 1239537600
#> 218       1239537600 1240142400
#> 219       1240142400 1240747200
#> 220       1240747200 1241352000
#> 221       1241352000 1241956800
#> 222       1241956800 1242561600
#> 223       1242561600 1243166400
#> 224       1243166400 1243771200
#> 225       1243771200 1244376000
#> 226       1244376000 1244980800
#> 227       1244980800 1245585600
#> 228       1245585600 1246190400
#> 229       1246190400 1246795200
#> 230       1246795200 1247400000
#> 231       1247400000 1248004800
#> 232       1248004800 1248609600
#> 233       1248609600 1249214400
#> 234       1249214400 1249819200
#> 235       1249819200 1250424000
#> 236       1250424000 1251028800
#> 237       1251028800 1251633600
#> 238       1251633600 1252238400
#> 239       1252238400 1252843200
#> 240       1252843200 1253448000
#> 241       1253448000 1254052800
#> 242       1254052800 1254657600
#> 243       1254657600 1255262400
#> 244       1255262400 1255867200
#> 245       1255867200 1256472000
#> 246       1256472000 1257076800
#> 247       1257076800 1257681600
#> 248       1257681600 1258286400
#> 249       1258286400 1258891200
#> 250       1258891200 1259496000
#> 251       1259496000 1260100800
#> 252       1260100800 1260705600
#> 253       1260705600 1261310400
#> 254       1261310400 1261915200
#> 255       1261915200 1262520000
#> 256       1262520000 1263124800
#> 257       1263124800 1263729600
#> 258       1263729600 1264334400
#> 259       1264334400 1264939200
#> 260       1264939200 1265544000
#> 261       1265544000 1266148800
#> 262       1266148800 1266753600
#> 263       1266753600 1267358400
#> 264       1267358400 1267963200
#> 265       1267963200 1268568000
#> 266       1268568000 1269172800
#> 267       1269172800 1269777600
#> 268       1269777600 1270382400
#> 269       1270382400 1270987200
#> 270       1270987200 1271592000
#> 271       1271592000 1272196800
#> 272       1272196800 1272801600
#> 273       1272801600 1273406400
#> 274       1273406400 1274011200
#> 275       1274011200 1274616000
#> 276       1274616000 1275220800
#> 277       1275220800 1275825600
#> 278       1275825600 1276430400
#> 279       1276430400 1277035200
#> 280       1277035200 1277640000
#> 281       1277640000 1278244800
#> 282       1278244800 1278849600
#> 283       1278849600 1279454400
#> 284       1279454400 1280059200
#> 285       1280059200 1280664000
#> 286       1280664000 1281268800
#> 287       1281268800 1281873600
#> 288       1281873600 1282478400
#> 289       1282478400 1283083200
#> 290       1283083200 1283688000
#> 291       1283688000 1284292800
#> 292       1284292800 1284897600
#> 293       1284897600 1285502400
#> 294       1285502400 1286107200
#> 295       1286107200 1286712000
#> 296       1286712000 1287316800
#> 297       1287316800 1287921600
#> 298       1287921600 1288526400
#> 299       1288526400 1289131200
#> 300       1289131200 1289736000
#> 301       1289736000 1290340800
#> 302       1290340800 1290945600
#> 303       1290945600 1291550400
#> 304       1291550400 1292155200
#> 305       1292155200 1292760000
#> 306       1292760000 1293364800
#> 307       1293364800 1293969600
#> 308       1293969600 1294574400
#> 309       1294574400 1295179200
#> 310       1295179200 1295784000
#> 311       1295784000 1296388800
#> 312       1296388800 1296993600
#> 313       1296993600 1297598400
#> 314       1297598400 1298203200
#> 315       1298203200 1298808000
#> 316       1298808000 1299412800
#> 317       1299412800 1300017600
#> 318       1300017600 1300622400
#> 319       1300622400 1301227200
#> 320       1301227200 1301832000
#> 321       1301832000 1302436800
#> 322       1302436800 1303041600
#> 323       1303041600 1303646400
#> 324       1303646400 1304251200
#> 325       1304251200 1304856000
#> 326       1304856000 1305460800
#> 327       1305460800 1306065600
#> 328       1306065600 1306670400
#> 329       1306670400 1307275200
#> 330       1307275200 1307880000
#> 331       1307880000 1308484800
#> 332       1308484800 1309089600
#> 333       1309089600 1309694400
#>  [ reached getOption("max.print") -- omitted 306 rows ]
#> 
#> $`@attr`
#> $`@attr`$user
#> [1] "platyjus"
#> 
#>