Get an track chart for a user profile.

Get an track chart for a user profile, for a given date range. If no date range is supplied, it will return the most recent track chart for this user. Implementation of last.fm's user.getWeeklyTrackChart API method (http://www.last.fm/api/show/user.getWeeklyTrackChart).

user_getWeeklyTrackChart(user, from = NA, to = NA, api_key = lastkey)

Arguments

user
The last.fm username to fetch the charts of.

from
The date at which the chart should start from. See user_getChartsList for more.

to
The date at which the chart should end on. See user_getChartsList for more.

api_key
A Last.fm API key.

Value

A list of an track chart for a user profile.

Examples

user_getWeeklyTrackChart("platyjus")

#> $track
#>        artist.#text                          artist.mbid
#> 1    The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 2    The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 3    The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 4    The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 5    The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 6         Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993
#> 7  Childish Gambino 7fb57fba-a6ef-44c2-abab-2fa3bdee607e
#> 8    The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 9      Joan Shelley                                     
#> 10          Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 11   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 12   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 13        Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993
#> 14        Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993
#> 15          Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 16   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 17          Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 18     Joan Shelley                                     
#> 19           Khalid c6bfb05d-f570-46c8-98e1-e25441189770
#> 20     Joan Shelley                                     
#> 21   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 22   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 23     Joan Shelley                                     
#> 24 Childish Gambino 7fb57fba-a6ef-44c2-abab-2fa3bdee607e
#> 25     Joan Shelley                                     
#> 26 Childish Gambino 7fb57fba-a6ef-44c2-abab-2fa3bdee607e
#> 27 Childish Gambino 7fb57fba-a6ef-44c2-abab-2fa3bdee607e
#> 28     Joan Shelley                                     
#> 29   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851
#> 30          Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 31     Joan Shelley                                     
#> 32     Joan Shelley                                     
#> 33     Joan Shelley                                     
#> 34        Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993
#> 35           Khalid c6bfb05d-f570-46c8-98e1-e25441189770
#>                                      name                                 mbid
#> 1                              Fix You Up                                     
#> 2                                Patience                                     
#> 3  Everything Looks Better (In Hindsight)                                     
#> 4                           Fall To Sleep                                     
#> 5                           Not An Option                                     
#> 6                  A Change Is Gonna Come 2f82aad5-728a-4094-b266-a201e065390d
#> 7                                Baby Boy                                     
#> 8                           Back to Earth                                     
#> 9                 Brighter Than The Blues                                     
#> 10                      Built by Pictures                                     
#> 11                                Capable                                     
#> 12                      Catch And Release                                     
#> 13                Chain Gang - Remastered                                     
#> 14                                  Cupid 8086b120-296a-4a08-91eb-0cac99dcfc38
#> 15                           Digging Deep                                     
#> 16                               Fruition                                     
#> 17                           Hold It Back                                     
#> 18               If the Storms Never Came                                     
#> 19                               LOCATION                                     
#> 20                       Not Over By Half                                     
#> 21                             Only Songs                                     
#> 22                Only Songs (Radio Edit)                                     
#> 23                          Over and Even                                     
#> 24                             Stand Tall                                     
#> 25                       Stay On My Shore                                     
#> 26                              terrified                                     
#> 27         The Night Me And Your Mama Met                                     
#> 28                      The Push and Pull                                     
#> 29                     The World We Built                                     
#> 30                                 Things                                     
#> 31                           We’d Be Home                                     
#> 32                    Where I’ll Find You                                     
#> 33                      Wild Indifference                                     
#> 34                            You Send Me 6a7a997c-bb78-408c-8b06-7a39243f416e
#> 35                     Young Dumb & Broke                                     
#>    playcount
#> 1          3
#> 2          3
#> 3          2
#> 4          2
#> 5          2
#> 6          1
#> 7          1
#> 8          1
#> 9          1
#> 10         1
#> 11         1
#> 12         1
#> 13         1
#> 14         1
#> 15         1
#> 16         1
#> 17         1
#> 18         1
#> 19         1
#> 20         1
#> 21         1
#> 22         1
#> 23         1
#> 24         1
#> 25         1
#> 26         1
#> 27         1
#> 28         1
#> 29         1
#> 30         1
#> 31         1
#> 32         1
#> 33         1
#> 34         1
#> 35         1
#>                                                                                                                                                                                                                                                                    image
#> 1  https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 2  https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 3  https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 4  https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 5  https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 6  https://lastfm-img2.akamaized.net/i/u/34s/f7ae03e9f4574566b35bc06e537cbd28.png, https://lastfm-img2.akamaized.net/i/u/64s/f7ae03e9f4574566b35bc06e537cbd28.png, https://lastfm-img2.akamaized.net/i/u/174s/f7ae03e9f4574566b35bc06e537cbd28.png, small, medium, large
#> 7  https://lastfm-img2.akamaized.net/i/u/34s/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/64s/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/174s/95384659e8a160a327e0d097a76db1f3.png, small, medium, large
#> 8  https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 9  https://lastfm-img2.akamaized.net/i/u/34s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/64s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/174s/4725042f9ec445e0ba8295d755edf41f.png, small, medium, large
#> 10 https://lastfm-img2.akamaized.net/i/u/34s/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/64s/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/174s/f6d5cc790d1b48c5838ff4de1c6f7549.png, small, medium, large
#> 11 https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 12 https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 13 https://lastfm-img2.akamaized.net/i/u/34s/f7ae03e9f4574566b35bc06e537cbd28.png, https://lastfm-img2.akamaized.net/i/u/64s/f7ae03e9f4574566b35bc06e537cbd28.png, https://lastfm-img2.akamaized.net/i/u/174s/f7ae03e9f4574566b35bc06e537cbd28.png, small, medium, large
#> 14 https://lastfm-img2.akamaized.net/i/u/34s/f7ae03e9f4574566b35bc06e537cbd28.png, https://lastfm-img2.akamaized.net/i/u/64s/f7ae03e9f4574566b35bc06e537cbd28.png, https://lastfm-img2.akamaized.net/i/u/174s/f7ae03e9f4574566b35bc06e537cbd28.png, small, medium, large
#> 15 https://lastfm-img2.akamaized.net/i/u/34s/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/64s/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/174s/f6d5cc790d1b48c5838ff4de1c6f7549.png, small, medium, large
#> 16 https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 17 https://lastfm-img2.akamaized.net/i/u/34s/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/64s/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/174s/f6d5cc790d1b48c5838ff4de1c6f7549.png, small, medium, large
#> 18 https://lastfm-img2.akamaized.net/i/u/34s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/64s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/174s/4725042f9ec445e0ba8295d755edf41f.png, small, medium, large
#> 19 https://lastfm-img2.akamaized.net/i/u/34s/1311d1c4986848db01fc71fd85aa133f.png, https://lastfm-img2.akamaized.net/i/u/64s/1311d1c4986848db01fc71fd85aa133f.png, https://lastfm-img2.akamaized.net/i/u/174s/1311d1c4986848db01fc71fd85aa133f.png, small, medium, large
#> 20 https://lastfm-img2.akamaized.net/i/u/34s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/64s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/174s/4725042f9ec445e0ba8295d755edf41f.png, small, medium, large
#> 21 https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 22 https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 23 https://lastfm-img2.akamaized.net/i/u/34s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/64s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/174s/4725042f9ec445e0ba8295d755edf41f.png, small, medium, large
#> 24 https://lastfm-img2.akamaized.net/i/u/34s/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/64s/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/174s/95384659e8a160a327e0d097a76db1f3.png, small, medium, large
#> 25 https://lastfm-img2.akamaized.net/i/u/34s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/64s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/174s/4725042f9ec445e0ba8295d755edf41f.png, small, medium, large
#> 26 https://lastfm-img2.akamaized.net/i/u/34s/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/64s/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/174s/95384659e8a160a327e0d097a76db1f3.png, small, medium, large
#> 27 https://lastfm-img2.akamaized.net/i/u/34s/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/64s/95384659e8a160a327e0d097a76db1f3.png, https://lastfm-img2.akamaized.net/i/u/174s/95384659e8a160a327e0d097a76db1f3.png, small, medium, large
#> 28 https://lastfm-img2.akamaized.net/i/u/34s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/64s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/174s/4725042f9ec445e0ba8295d755edf41f.png, small, medium, large
#> 29 https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, small, medium, large
#> 30 https://lastfm-img2.akamaized.net/i/u/34s/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/64s/f6d5cc790d1b48c5838ff4de1c6f7549.png, https://lastfm-img2.akamaized.net/i/u/174s/f6d5cc790d1b48c5838ff4de1c6f7549.png, small, medium, large
#> 31 https://lastfm-img2.akamaized.net/i/u/34s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/64s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/174s/4725042f9ec445e0ba8295d755edf41f.png, small, medium, large
#> 32 https://lastfm-img2.akamaized.net/i/u/34s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/64s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/174s/4725042f9ec445e0ba8295d755edf41f.png, small, medium, large
#> 33 https://lastfm-img2.akamaized.net/i/u/34s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/64s/4725042f9ec445e0ba8295d755edf41f.png, https://lastfm-img2.akamaized.net/i/u/174s/4725042f9ec445e0ba8295d755edf41f.png, small, medium, large
#> 34 https://lastfm-img2.akamaized.net/i/u/34s/f7ae03e9f4574566b35bc06e537cbd28.png, https://lastfm-img2.akamaized.net/i/u/64s/f7ae03e9f4574566b35bc06e537cbd28.png, https://lastfm-img2.akamaized.net/i/u/174s/f7ae03e9f4574566b35bc06e537cbd28.png, small, medium, large
#> 35 https://lastfm-img2.akamaized.net/i/u/34s/1311d1c4986848db01fc71fd85aa133f.png, https://lastfm-img2.akamaized.net/i/u/64s/1311d1c4986848db01fc71fd85aa133f.png, https://lastfm-img2.akamaized.net/i/u/174s/1311d1c4986848db01fc71fd85aa133f.png, small, medium, large
#>                                                                                  url
#> 1                              https://www.last.fm/music/The+Wild+Reeds/_/Fix+You+Up
#> 2                                https://www.last.fm/music/The+Wild+Reeds/_/Patience
#> 3  https://www.last.fm/music/The+Wild+Reeds/_/Everything+Looks+Better+(In+Hindsight)
#> 4                           https://www.last.fm/music/The+Wild+Reeds/_/Fall+To+Sleep
#> 5                           https://www.last.fm/music/The+Wild+Reeds/_/Not+An+Option
#> 6                       https://www.last.fm/music/Sam+Cooke/_/A+Change+Is+Gonna+Come
#> 7                              https://www.last.fm/music/Childish+Gambino/_/Baby+Boy
#> 8                           https://www.last.fm/music/The+Wild+Reeds/_/Back+to+Earth
#> 9                   https://www.last.fm/music/Joan+Shelley/_/Brighter+Than+The+Blues
#> 10                             https://www.last.fm/music/Oddisee/_/Built+by+Pictures
#> 11                                https://www.last.fm/music/The+Wild+Reeds/_/Capable
#> 12                      https://www.last.fm/music/The+Wild+Reeds/_/Catch+And+Release
#> 13                     https://www.last.fm/music/Sam+Cooke/_/Chain+Gang+-+Remastered
#> 14                                       https://www.last.fm/music/Sam+Cooke/_/Cupid
#> 15                                  https://www.last.fm/music/Oddisee/_/Digging+Deep
#> 16                               https://www.last.fm/music/The+Wild+Reeds/_/Fruition
#> 17                                  https://www.last.fm/music/Oddisee/_/Hold+It+Back
#> 18                 https://www.last.fm/music/Joan+Shelley/_/If+the+Storms+Never+Came
#> 19                                       https://www.last.fm/music/Khalid/_/LOCATION
#> 20                         https://www.last.fm/music/Joan+Shelley/_/Not+Over+By+Half
#> 21                             https://www.last.fm/music/The+Wild+Reeds/_/Only+Songs
#> 22                https://www.last.fm/music/The+Wild+Reeds/_/Only+Songs+(Radio+Edit)
#> 23                            https://www.last.fm/music/Joan+Shelley/_/Over+and+Even
#> 24                           https://www.last.fm/music/Childish+Gambino/_/Stand+Tall
#> 25                         https://www.last.fm/music/Joan+Shelley/_/Stay+On+My+Shore
#> 26                            https://www.last.fm/music/Childish+Gambino/_/terrified
#> 27       https://www.last.fm/music/Childish+Gambino/_/The+Night+Me+And+Your+Mama+Met
#> 28                        https://www.last.fm/music/Joan+Shelley/_/The+Push+and+Pull
#> 29                     https://www.last.fm/music/The+Wild+Reeds/_/The+World+We+Built
#> 30                                        https://www.last.fm/music/Oddisee/_/Things
#> 31                     https://www.last.fm/music/Joan+Shelley/_/We%E2%80%99d+Be+Home
#> 32              https://www.last.fm/music/Joan+Shelley/_/Where+I%E2%80%99ll+Find+You
#> 33                        https://www.last.fm/music/Joan+Shelley/_/Wild+Indifference
#> 34                                 https://www.last.fm/music/Sam+Cooke/_/You+Send+Me
#> 35                             https://www.last.fm/music/Khalid/_/Young+Dumb+&+Broke
#>    rank
#> 1     1
#> 2     2
#> 3     3
#> 4     4
#> 5     5
#> 6     6
#> 7     7
#> 8     8
#> 9     9
#> 10   10
#> 11   11
#> 12   12
#> 13   13
#> 14   14
#> 15   15
#> 16   16
#> 17   17
#> 18   18
#> 19   19
#> 20   20
#> 21   21
#> 22   22
#> 23   23
#> 24   24
#> 25   25
#> 26   26
#> 27   27
#> 28   28
#> 29   29
#> 30   30
#> 31   31
#> 32   32
#> 33   33
#> 34   34
#> 35   35
#> 
#> $`@attr`
#> $`@attr`$user
#> [1] "platyjus"
#> 
#> $`@attr`$from
#> [1] "1494504000"
#> 
#> $`@attr`$to
#> [1] "1495108800"
#> 
#>