Get the metadata and tracklist for an album.

Get the metadata and tracklist for an album on Last.fm using the album name or a musicbrainz id. Implementation of last.fm's album.getInfo API method (http://www.last.fm/api/show/album.getInfo)

album_getInfo(artist, album, mbid = NA, autocorrect = NA, username = NA,
  lang = NA, api_key = lastkey)

Arguments

artist
The artist name.

album
The album name.

mbid
The musicbrainz id for the album.

autocorrect
Transform misspelled artist names into correct artist names, returning the correct version instead. The corrected artist name will be returned in the response. [0|1]

username
The username for the context of the request. If supplied, the user's playcount for this album is included in the response.

lang
The language to return the biography in, expressed as an ISO 639 alpha-2 code.

api_key
A Last.fm API key.

Value

A list of the metadata and tracklist for an album.

Examples

album_getInfo("Father John Misty", "Fear Fun")

#> $name
#> [1] "Fear Fun"
#> 
#> $artist
#> [1] "Father John Misty"
#> 
#> $mbid
#> [1] "da07a173-bf3d-4e7e-995d-ce473ce83030"
#> 
#> $url
#> [1] "https://www.last.fm/music/Father+John+Misty/Fear+Fun"
#> 
#> $image
#>                                                                                #text
#> 1     https://lastfm-img2.akamaized.net/i/u/34s/d9839e4cfadb49909637e9b6e39caa95.png
#> 2     https://lastfm-img2.akamaized.net/i/u/64s/d9839e4cfadb49909637e9b6e39caa95.png
#> 3    https://lastfm-img2.akamaized.net/i/u/174s/d9839e4cfadb49909637e9b6e39caa95.png
#> 4 https://lastfm-img2.akamaized.net/i/u/300x300/d9839e4cfadb49909637e9b6e39caa95.png
#> 5         https://lastfm-img2.akamaized.net/i/u/d9839e4cfadb49909637e9b6e39caa95.png
#> 6     https://lastfm-img2.akamaized.net/i/u/arQ/d9839e4cfadb49909637e9b6e39caa95.png
#>         size
#> 1      small
#> 2     medium
#> 3      large
#> 4 extralarge
#> 5       mega
#> 6           
#> 
#> $listeners
#> [1] "217420"
#> 
#> $playcount
#> [1] "4429644"
#> 
#> $tracks
#> $tracks$track
#>                                    name
#> 1                   Funtimes in Babylon
#> 2                     Nancy From Now On
#> 3      Hollywood Forever Cemetery Sings
#> 4                   I'm Writing a Novel
#> 5  O I Long to Feel Your Arms Around Me
#> 6              Misty's Nightmares 1 & 2
#> 7             Only Son of the Ladiesman
#> 8                 This Is Sally Hatchet
#> 9        Well, You Can Do It Without Me
#> 10     Now I'm Learning to Love the War
#> 11                        Tee Pees 1-12
#> 12           Everyman Needs a Companion
#>                                                                                   url
#> 1                   https://www.last.fm/music/Father+John+Misty/_/Funtimes+in+Babylon
#> 2                     https://www.last.fm/music/Father+John+Misty/_/Nancy+From+Now+On
#> 3      https://www.last.fm/music/Father+John+Misty/_/Hollywood+Forever+Cemetery+Sings
#> 4                 https://www.last.fm/music/Father+John+Misty/_/I%27m+Writing+a+Novel
#> 5  https://www.last.fm/music/Father+John+Misty/_/O+I+Long+to+Feel+Your+Arms+Around+Me
#> 6            https://www.last.fm/music/Father+John+Misty/_/Misty%27s+Nightmares+1+&+2
#> 7             https://www.last.fm/music/Father+John+Misty/_/Only+Son+of+the+Ladiesman
#> 8                 https://www.last.fm/music/Father+John+Misty/_/This+Is+Sally+Hatchet
#> 9        https://www.last.fm/music/Father+John+Misty/_/Well,+You+Can+Do+It+Without+Me
#> 10   https://www.last.fm/music/Father+John+Misty/_/Now+I%27m+Learning+to+Love+the+War
#> 11                        https://www.last.fm/music/Father+John+Misty/_/Tee+Pees+1-12
#> 12           https://www.last.fm/music/Father+John+Misty/_/Everyman+Needs+a+Companion
#>    duration rank streamable.#text streamable.fulltrack       artist.name
#> 1       246    1                0                    0 Father John Misty
#> 2       229    2                0                    0 Father John Misty
#> 3       190    3                0                    0 Father John Misty
#> 4       215    4                0                    0 Father John Misty
#> 5       143    5                0                    0 Father John Misty
#> 6       193    6                0                    0 Father John Misty
#> 7       249    7                0                    0 Father John Misty
#> 8       237    8                0                    0 Father John Misty
#> 9       163    9                0                    0 Father John Misty
#> 10      261   10                0                    0 Father John Misty
#> 11      197   11                0                    0 Father John Misty
#> 12      302   12                0                    0 Father John Misty
#>                             artist.mbid
#> 1  172053e5-a7fb-4548-a64d-c53858b5e775
#> 2  172053e5-a7fb-4548-a64d-c53858b5e775
#> 3  172053e5-a7fb-4548-a64d-c53858b5e775
#> 4  172053e5-a7fb-4548-a64d-c53858b5e775
#> 5  172053e5-a7fb-4548-a64d-c53858b5e775
#> 6  172053e5-a7fb-4548-a64d-c53858b5e775
#> 7  172053e5-a7fb-4548-a64d-c53858b5e775
#> 8  172053e5-a7fb-4548-a64d-c53858b5e775
#> 9  172053e5-a7fb-4548-a64d-c53858b5e775
#> 10 172053e5-a7fb-4548-a64d-c53858b5e775
#> 11 172053e5-a7fb-4548-a64d-c53858b5e775
#> 12 172053e5-a7fb-4548-a64d-c53858b5e775
#>                                     artist.url
#> 1  https://www.last.fm/music/Father+John+Misty
#> 2  https://www.last.fm/music/Father+John+Misty
#> 3  https://www.last.fm/music/Father+John+Misty
#> 4  https://www.last.fm/music/Father+John+Misty
#> 5  https://www.last.fm/music/Father+John+Misty
#> 6  https://www.last.fm/music/Father+John+Misty
#> 7  https://www.last.fm/music/Father+John+Misty
#> 8  https://www.last.fm/music/Father+John+Misty
#> 9  https://www.last.fm/music/Father+John+Misty
#> 10 https://www.last.fm/music/Father+John+Misty
#> 11 https://www.last.fm/music/Father+John+Misty
#> 12 https://www.last.fm/music/Father+John+Misty
#> 
#> 
#> $tags
#> $tags$tag
#>                name                                       url
#> 1              2012              https://www.last.fm/tag/2012
#> 2      best of 2012      https://www.last.fm/tag/best+of+2012
#> 3 availableonemusic https://www.last.fm/tag/availableonemusic
#> 4              rock              https://www.last.fm/tag/rock
#> 5       alternative       https://www.last.fm/tag/alternative
#> 
#>