Get the metadata for a track.

Get the metadata for a track on Last.fm using the artist/track name or a musicbrainz id. Implementation of last.fm's track.getInfo API method (http://www.last.fm/api/show/track.getInfo)

track_getInfo(track, artist, mbid = NA, autocorrect = NA, username = NA)

Arguments

track
The track name.

artist
The artist name.

mbid
The musicbrainz id for the track

autocorrect
Transform misspelled artist and track names into correct artist and track names, returning the correct version instead. The corrected artist and track name will be returned in the response. [0|1]

username
The username for the context of the request. If supplied, the user's playcount for this track and whether they have loved the track is included in the response.

Value

A list of the metadata for a track.

Examples

track_getInfo("King Kunta", "Kendrick Lamar")

#>         name                                                   url duration
#> 1 King Kunta https://www.last.fm/music/Kendrick+Lamar/_/King+Kunta   234000
#>   streamable..text streamable.fulltrack listeners playcount    artist.name
#> 1                0                    0    323801   2705559 Kendrick Lamar
#>                            artist.mbid                               artist.url
#> 1 381086ea-f511-4aba-bdf9-71c753dc5077 https://www.last.fm/music/Kendrick+Lamar
#>     album.artist         album.title
#> 1 Kendrick Lamar To Pimp a Butterfly
#>                                                      album.url
#> 1 https://www.last.fm/music/Kendrick+Lamar/To+Pimp+a+Butterfly
#>                                                               album.image.extralarge
#> 1 https://lastfm-img2.akamaized.net/i/u/300x300/7ec7c79af7bbc92709ff625f369d80bf.png
#>                                                                 album.image.large
#> 1 https://lastfm-img2.akamaized.net/i/u/174s/7ec7c79af7bbc92709ff625f369d80bf.png
#>                                                               album.image.medium
#> 1 https://lastfm-img2.akamaized.net/i/u/64s/7ec7c79af7bbc92709ff625f369d80bf.png
#>                                                                album.image.small
#> 1 https://lastfm-img2.akamaized.net/i/u/34s/7ec7c79af7bbc92709ff625f369d80bf.png
#>                                                                                                                                                                                                            toptags
#> 1 Hip-Hop, best of 2015, funk, hip hop, funky, https://www.last.fm/tag/Hip-Hop, https://www.last.fm/tag/best+of+2015, https://www.last.fm/tag/funk, https://www.last.fm/tag/hip+hop, https://www.last.fm/tag/funky
#>       wiki.published
#> 1 19 Mar 2015, 17:13
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     wiki.summary
#> 1 This song contains interpolations of "Get Nekkid", written by Johnny Burns, as performed by Mausberg; resung lyrics from "Smooth Criminal", written and performed by Michael Jackson; elements of "The Payback", written by James Brown, Fred Wesley, and John Starks, as performed by James Brown; and samples of "We Want the Funk", written by Ahmad Lewis, as performed by Ahmad. It also contains additional vocals performed by Whitney Alford. <a href="http://www.last.fm/music/Kendrick+Lamar/_/King+Kunta">Read more on Last.fm</a>.
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              wiki.content
#> 1 This song contains interpolations of "Get Nekkid", written by Johnny Burns, as performed by Mausberg; resung lyrics from "Smooth Criminal", written and performed by Michael Jackson; elements of "The Payback", written by James Brown, Fred Wesley, and John Starks, as performed by James Brown; and samples of "We Want the Funk", written by Ahmad Lewis, as performed by Ahmad. It also contains additional vocals performed by Whitney Alford. <a href="http://www.last.fm/music/Kendrick+Lamar/_/King+Kunta">Read more on Last.fm</a>. User-contributed text is available under the Creative Commons By-SA License; additional terms may apply.