Get the top tracks listened to by a user.

Get the top tracks listened to by a user. You can stipulate a time period. Sends the overall chart by default. Implementation of last.fm's user.getTopTracks API method (http://www.last.fm/api/show/user.getTopTracks).

user_getTopTracks(user, period = NA, limit = NA, page = NA,
  api_key = lastkey)

Arguments

user
The user name to fetch top tracks for.

period
overall | 7day | 1month | 3month | 6month | 12month - The time period over which to retrieve top tracks for.

limit
The number of results to fetch per page. Defaults to 50.

page
The page number to fetch. Defaults to first page.

api_key
A Last.fm API key.

Value

A list of the top tracks listened to by a user.

Examples

user_getTopTracks("platyjus")

#> $track
#>                                                name duration playcount
#> 1                                       Coming Home        0        24
#> 2                                        Better Man        0        23
#> 3               Nobody Speak (feat. Run the Jewels)        0        23
#> 4                                   Brown Skin Girl        0        22
#> 5                                         Blessings        0        20
#> 6                                          Capsized        0        19
#> 7                                       Lisa Sawyer        0        19
#> 8                                             River        0        19
#> 9                                              Fall        0        18
#> 10                                   Smooth Sailin'        0        18
#> 11                                        Pull Away        0        17
#> 12                                            Shine        0        17
#> 13 Nobody Really Cares If You Don't Go to the Party        0        16
#> 14                                          Ophelia        0        16
#> 15                                          Peanuts        0        16
#> 16                                 All We Ever Knew        0        15
#> 17                             Prayer For The Dying        0        15
#> 18                               Sleep on the Floor        0        15
#> 19                                        Cleopatra        0        14
#> 20                                         Good Day        0        14
#> 21                                             Stay        0        14
#> 22                         Blockbuster Night Part 1      152        13
#> 23                                        Depreston        0        13
#> 24                                             Gwan      231        13
#> 25                                  Here In My Arms        0        13
#> 26                                   Make Some Room        0        13
#> 27                                        Norf Norf        0        13
#> 28                                              Ora        0        13
#> 29                             Skylark Interabang?!      283        13
#> 30                                   All The Places      196        12
#> 31                              Angels (feat. Saba)        0        12
#> 32                                            DEATH        0        12
#> 33           Everything Looks Better (In Hindsight)        0        12
#> 34                                          Flowers        0        12
#> 35                                            Giver        0        12
#> 36                                       Lift Me Up        0        12
#> 37                                     Paper Trail$      194        12
#> 38                                        ROMA FADE        0        12
#> 39                                             Snow        0        12
#> 40                              Twistin' & Groovin'        0        12
#> 41                                           Better        0        11
#> 42                                         Gun Song      191        11
#> 43                       Heart Don't Stand a Chance        0        11
#> 44                                     Hors d'œuvre      205        11
#> 45                               Left Handed Kisses        0        11
#> 46                                               Lo        0        11
#> 47                                          Midtown        0        11
#> 48                               Pedestrian at Best        0        11
#> 49                                        Pirouette        0        11
#> 50                                         The Bird        0        11
#>                                    mbid
#> 1                                      
#> 2                                      
#> 3                                      
#> 4                                      
#> 5                                      
#> 6                                      
#> 7                                      
#> 8                                      
#> 9                                      
#> 10                                     
#> 11                                     
#> 12                                     
#> 13                                     
#> 14                                     
#> 15                                     
#> 16                                     
#> 17                                     
#> 18                                     
#> 19                                     
#> 20                                     
#> 21                                     
#> 22 3532826e-5523-42d1-bf9d-8b5318f9470d
#> 23                                     
#> 24 c848fff9-2434-4d5e-8f42-456381a033a2
#> 25                                     
#> 26                                     
#> 27                                     
#> 28                                     
#> 29 4d9b5da0-9a5e-4b24-97e9-d3497fc3a3c1
#> 30 1d6bf8bf-7030-4254-9617-2d499ffc16eb
#> 31                                     
#> 32                                     
#> 33                                     
#> 34                                     
#> 35                                     
#> 36                                     
#> 37 8ead7130-7ff9-4491-9837-e16610d3ebdb
#> 38                                     
#> 39                                     
#> 40                                     
#> 41                                     
#> 42 f108c60c-e286-4bda-9584-dd0cb70535c9
#> 43                                     
#> 44 e3a9dfc0-ef48-4b9c-84d9-e1092530e5fa
#> 45                                     
#> 46                                     
#> 47                                     
#> 48                                     
#> 49                                     
#> 50                                     
#>                                                                                                url
#> 1                                             https://www.last.fm/music/Leon+Bridges/_/Coming+Home
#> 2                                              https://www.last.fm/music/Leon+Bridges/_/Better+Man
#> 3                        https://www.last.fm/music/DJ+Shadow/_/Nobody+Speak+(feat.+Run+the+Jewels)
#> 4                                         https://www.last.fm/music/Leon+Bridges/_/Brown+Skin+Girl
#> 5                                          https://www.last.fm/music/Chance+the+Rapper/_/Blessings
#> 6                                                 https://www.last.fm/music/Andrew+Bird/_/Capsized
#> 7                                             https://www.last.fm/music/Leon+Bridges/_/Lisa+Sawyer
#> 8                                                   https://www.last.fm/music/Leon+Bridges/_/River
#> 9                                                   https://www.last.fm/music/Lisa+Hannigan/_/Fall
#> 10                                       https://www.last.fm/music/Leon+Bridges/_/Smooth+Sailin%27
#> 11                                              https://www.last.fm/music/Leon+Bridges/_/Pull+Away
#> 12                                                  https://www.last.fm/music/Leon+Bridges/_/Shine
#> 13 https://www.last.fm/music/Courtney+Barnett/_/Nobody+Really+Cares+If+You+Don%27t+Go+to+the+Party
#> 14                                               https://www.last.fm/music/The+Lumineers/_/Ophelia
#> 15                                                 https://www.last.fm/music/The+Suffers/_/Peanuts
#> 16                             https://www.last.fm/music/The+Head+and+the+Heart/_/All+We+Ever+Knew
#> 17                                  https://www.last.fm/music/Lisa+Hannigan/_/Prayer+For+The+Dying
#> 18                                    https://www.last.fm/music/The+Lumineers/_/Sleep+on+the+Floor
#> 19                                             https://www.last.fm/music/The+Lumineers/_/Cleopatra
#> 20                                                https://www.last.fm/music/The+Suffers/_/Good+Day
#> 21                                                    https://www.last.fm/music/The+Suffers/_/Stay
#> 22                             https://www.last.fm/music/Run+the+Jewels/_/Blockbuster+Night+Part+1
#> 23                                          https://www.last.fm/music/Courtney+Barnett/_/Depreston
#> 24                                                    https://www.last.fm/music/The+Suffers/_/Gwan
#> 25                                        https://www.last.fm/music/Leon+Bridges/_/Here+In+My+Arms
#> 26                                          https://www.last.fm/music/The+Suffers/_/Make+Some+Room
#> 27                                             https://www.last.fm/music/Vince+Staples/_/Norf+Norf
#> 28                                                   https://www.last.fm/music/Lisa+Hannigan/_/Ora
#> 29              https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/_/Skylark+Interabang%3F%21
#> 30                        https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/_/All+The+Places
#> 31                               https://www.last.fm/music/Chance+the+Rapper/_/Angels+(feat.+Saba)
#> 32                                 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/_/DEATH
#> 33               https://www.last.fm/music/The+Wild+Reeds/_/Everything+Looks+Better+(In+Hindsight)
#> 34                                                https://www.last.fm/music/Leon+Bridges/_/Flowers
#> 35                                                   https://www.last.fm/music/The+Suffers/_/Giver
#> 36                                            https://www.last.fm/music/Vince+Staples/_/Lift+Me+Up
#> 37                                            https://www.last.fm/music/Joey+Bada$$/_/Paper+Trail$
#> 38                                               https://www.last.fm/music/Andrew+Bird/_/ROMA+FADE
#> 39                                                  https://www.last.fm/music/Lisa+Hannigan/_/Snow
#> 40                                https://www.last.fm/music/Leon+Bridges/_/Twistin%27+&+Groovin%27
#> 41                                                  https://www.last.fm/music/The+Suffers/_/Better
#> 42                                              https://www.last.fm/music/The+Lumineers/_/Gun+Song
#> 43                          https://www.last.fm/music/Anderson+Paak/_/Heart+Don%27t+Stand+a+Chance
#> 44                   https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/_/Hors+d%27%C5%93uvre
#> 45                                      https://www.last.fm/music/Andrew+Bird/_/Left+Handed+Kisses
#> 46                                                    https://www.last.fm/music/Lisa+Hannigan/_/Lo
#> 47                                                 https://www.last.fm/music/The+Suffers/_/Midtown
#> 48                                 https://www.last.fm/music/Courtney+Barnett/_/Pedestrian+at+Best
#> 49                             https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::/_/Pirouette
#> 50                                              https://www.last.fm/music/Anderson+Paak/_/The+Bird
#>    streamable.#text streamable.fulltrack            artist.name
#> 1                 0                    0           Leon Bridges
#> 2                 0                    0           Leon Bridges
#> 3                 0                    0              DJ Shadow
#> 4                 0                    0           Leon Bridges
#> 5                 0                    0      Chance the Rapper
#> 6                 0                    0            Andrew Bird
#> 7                 0                    0           Leon Bridges
#> 8                 0                    0           Leon Bridges
#> 9                 0                    0          Lisa Hannigan
#> 10                0                    0           Leon Bridges
#> 11                0                    0           Leon Bridges
#> 12                0                    0           Leon Bridges
#> 13                0                    0       Courtney Barnett
#> 14                0                    0          The Lumineers
#> 15                0                    0            The Suffers
#> 16                0                    0 The Head and the Heart
#> 17                0                    0          Lisa Hannigan
#> 18                0                    0          The Lumineers
#> 19                0                    0          The Lumineers
#> 20                0                    0            The Suffers
#> 21                0                    0            The Suffers
#> 22                0                    0         Run the Jewels
#> 23                0                    0       Courtney Barnett
#> 24                0                    0            The Suffers
#> 25                0                    0           Leon Bridges
#> 26                0                    0            The Suffers
#> 27                0                    0          Vince Staples
#> 28                0                    0          Lisa Hannigan
#> 29                0                    0  ::M∆DE::IN::HEIGHTS::
#> 30                0                    0  ::M∆DE::IN::HEIGHTS::
#> 31                0                    0      Chance the Rapper
#> 32                0                    0  ::M∆DE::IN::HEIGHTS::
#> 33                0                    0         The Wild Reeds
#> 34                0                    0           Leon Bridges
#> 35                0                    0            The Suffers
#> 36                0                    0          Vince Staples
#> 37                0                    0            Joey Bada$$
#> 38                0                    0            Andrew Bird
#> 39                0                    0          Lisa Hannigan
#> 40                0                    0           Leon Bridges
#> 41                0                    0            The Suffers
#> 42                0                    0          The Lumineers
#> 43                0                    0          Anderson Paak
#> 44                0                    0  ::M∆DE::IN::HEIGHTS::
#> 45                0                    0            Andrew Bird
#> 46                0                    0          Lisa Hannigan
#> 47                0                    0            The Suffers
#> 48                0                    0       Courtney Barnett
#> 49                0                    0  ::M∆DE::IN::HEIGHTS::
#> 50                0                    0          Anderson Paak
#>                             artist.mbid
#> 1                                      
#> 2                                      
#> 3  efa2c11a-1a35-4b60-bc1b-66d37de88511
#> 4                                      
#> 5  373a4c98-a46b-48e4-86ec-f6ca65b4f438
#> 6  081b133e-ce74-42ba-92c1-c18234acb532
#> 7                                      
#> 8                                      
#> 9  8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#> 10                                     
#> 11                                     
#> 12                                     
#> 13 55111838-f001-494a-a1b5-9d818db85810
#> 14 bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 15 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 16 98cb416e-5271-484c-ab3b-6bed6b4aae12
#> 17 8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#> 18 bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 19 bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 20 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 21 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 22 a80c5dc5-b12e-4667-9f5a-b568961f3839
#> 23 55111838-f001-494a-a1b5-9d818db85810
#> 24 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 25                                     
#> 26 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 27 78e854b8-9713-4ff2-9218-6b3784893bff
#> 28 8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#> 29 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 30 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 31 373a4c98-a46b-48e4-86ec-f6ca65b4f438
#> 32 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 33 54412058-8ed9-49a2-b931-ce8071caf851
#> 34                                     
#> 35 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 36 78e854b8-9713-4ff2-9218-6b3784893bff
#> 37 9757ae7d-44b1-4d63-81ce-b463fe27b9ae
#> 38 081b133e-ce74-42ba-92c1-c18234acb532
#> 39 8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#> 40                                     
#> 41 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 42 bfcb6630-9b31-4e63-b11f-7b0363be72b5
#> 43 d02dd67e-f655-4600-bc47-f789f59e7367
#> 44 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 45 081b133e-ce74-42ba-92c1-c18234acb532
#> 46 8d0d65cf-3d01-4110-bf36-7e1aa27b68f2
#> 47 bbfd82bf-6f5f-43cf-ad53-82f13cbc1322
#> 48 55111838-f001-494a-a1b5-9d818db85810
#> 49 72c9f2ac-12ca-4a2a-ae79-6b19bd1e5b73
#> 50 d02dd67e-f655-4600-bc47-f789f59e7367
#>                                                 artist.url
#> 1                   https://www.last.fm/music/Leon+Bridges
#> 2                   https://www.last.fm/music/Leon+Bridges
#> 3                      https://www.last.fm/music/DJ+Shadow
#> 4                   https://www.last.fm/music/Leon+Bridges
#> 5              https://www.last.fm/music/Chance+the+Rapper
#> 6                    https://www.last.fm/music/Andrew+Bird
#> 7                   https://www.last.fm/music/Leon+Bridges
#> 8                   https://www.last.fm/music/Leon+Bridges
#> 9                  https://www.last.fm/music/Lisa+Hannigan
#> 10                  https://www.last.fm/music/Leon+Bridges
#> 11                  https://www.last.fm/music/Leon+Bridges
#> 12                  https://www.last.fm/music/Leon+Bridges
#> 13              https://www.last.fm/music/Courtney+Barnett
#> 14                 https://www.last.fm/music/The+Lumineers
#> 15                   https://www.last.fm/music/The+Suffers
#> 16        https://www.last.fm/music/The+Head+and+the+Heart
#> 17                 https://www.last.fm/music/Lisa+Hannigan
#> 18                 https://www.last.fm/music/The+Lumineers
#> 19                 https://www.last.fm/music/The+Lumineers
#> 20                   https://www.last.fm/music/The+Suffers
#> 21                   https://www.last.fm/music/The+Suffers
#> 22                https://www.last.fm/music/Run+the+Jewels
#> 23              https://www.last.fm/music/Courtney+Barnett
#> 24                   https://www.last.fm/music/The+Suffers
#> 25                  https://www.last.fm/music/Leon+Bridges
#> 26                   https://www.last.fm/music/The+Suffers
#> 27                 https://www.last.fm/music/Vince+Staples
#> 28                 https://www.last.fm/music/Lisa+Hannigan
#> 29 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 30 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 31             https://www.last.fm/music/Chance+the+Rapper
#> 32 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 33                https://www.last.fm/music/The+Wild+Reeds
#> 34                  https://www.last.fm/music/Leon+Bridges
#> 35                   https://www.last.fm/music/The+Suffers
#> 36                 https://www.last.fm/music/Vince+Staples
#> 37                   https://www.last.fm/music/Joey+Bada$$
#> 38                   https://www.last.fm/music/Andrew+Bird
#> 39                 https://www.last.fm/music/Lisa+Hannigan
#> 40                  https://www.last.fm/music/Leon+Bridges
#> 41                   https://www.last.fm/music/The+Suffers
#> 42                 https://www.last.fm/music/The+Lumineers
#> 43                 https://www.last.fm/music/Anderson+Paak
#> 44 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 45                   https://www.last.fm/music/Andrew+Bird
#> 46                 https://www.last.fm/music/Lisa+Hannigan
#> 47                   https://www.last.fm/music/The+Suffers
#> 48              https://www.last.fm/music/Courtney+Barnett
#> 49 https://www.last.fm/music/::M%E2%88%86DE::IN::HEIGHTS::
#> 50                 https://www.last.fm/music/Anderson+Paak
#>                                                                                                                                                                                                                                                                                                                                                                    image
#> 1  https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge
#> 2  https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge
#> 3  https://lastfm-img2.akamaized.net/i/u/34s/33a127e674f149f18d9e27528d469302.png, https://lastfm-img2.akamaized.net/i/u/64s/33a127e674f149f18d9e27528d469302.png, https://lastfm-img2.akamaized.net/i/u/174s/33a127e674f149f18d9e27528d469302.png, https://lastfm-img2.akamaized.net/i/u/300x300/33a127e674f149f18d9e27528d469302.png, small, medium, large, extralarge
#> 4  https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge
#> 5  https://lastfm-img2.akamaized.net/i/u/34s/3d42bab35e7ad7332a27b9ff38db971a.png, https://lastfm-img2.akamaized.net/i/u/64s/3d42bab35e7ad7332a27b9ff38db971a.png, https://lastfm-img2.akamaized.net/i/u/174s/3d42bab35e7ad7332a27b9ff38db971a.png, https://lastfm-img2.akamaized.net/i/u/300x300/3d42bab35e7ad7332a27b9ff38db971a.png, small, medium, large, extralarge
#> 6  https://lastfm-img2.akamaized.net/i/u/34s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/64s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/174s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/300x300/73dba10b71674e029b3ddc0692899e0b.png, small, medium, large, extralarge
#> 7  https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge
#> 8  https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge
#> 9  https://lastfm-img2.akamaized.net/i/u/34s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/64s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/174s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/300x300/c8f8be9e530f4956a70c00d46c272fbe.png, small, medium, large, extralarge
#> 10 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge
#> 11 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge
#> 12 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge
#> 13 https://lastfm-img2.akamaized.net/i/u/34s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/64s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/174s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/300x300/ac5489a01d41cb391ab5f745d157a291.png, small, medium, large, extralarge
#> 14 https://lastfm-img2.akamaized.net/i/u/34s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/64s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/174s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/300x300/03e320dd70c0577c42c459eac0063c16.png, small, medium, large, extralarge
#> 15 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png, small, medium, large, extralarge
#> 16 https://lastfm-img2.akamaized.net/i/u/34s/2c50aec4065c4947be519bd11f15aba8.png, https://lastfm-img2.akamaized.net/i/u/64s/2c50aec4065c4947be519bd11f15aba8.png, https://lastfm-img2.akamaized.net/i/u/174s/2c50aec4065c4947be519bd11f15aba8.png, https://lastfm-img2.akamaized.net/i/u/300x300/2c50aec4065c4947be519bd11f15aba8.png, small, medium, large, extralarge
#> 17 https://lastfm-img2.akamaized.net/i/u/34s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/64s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/174s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/300x300/c8f8be9e530f4956a70c00d46c272fbe.png, small, medium, large, extralarge
#> 18 https://lastfm-img2.akamaized.net/i/u/34s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/64s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/174s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/300x300/03e320dd70c0577c42c459eac0063c16.png, small, medium, large, extralarge
#> 19 https://lastfm-img2.akamaized.net/i/u/34s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/64s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/174s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/300x300/03e320dd70c0577c42c459eac0063c16.png, small, medium, large, extralarge
#> 20 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png, small, medium, large, extralarge
#> 21 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png, small, medium, large, extralarge
#> 22 https://lastfm-img2.akamaized.net/i/u/34s/fbb820614a34485ac8b0a1f2d0f02c9b.png, https://lastfm-img2.akamaized.net/i/u/64s/fbb820614a34485ac8b0a1f2d0f02c9b.png, https://lastfm-img2.akamaized.net/i/u/174s/fbb820614a34485ac8b0a1f2d0f02c9b.png, https://lastfm-img2.akamaized.net/i/u/300x300/fbb820614a34485ac8b0a1f2d0f02c9b.png, small, medium, large, extralarge
#> 23 https://lastfm-img2.akamaized.net/i/u/34s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/64s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/174s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/300x300/ac5489a01d41cb391ab5f745d157a291.png, small, medium, large, extralarge
#> 24 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png, small, medium, large, extralarge
#> 25 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge
#> 26 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png, small, medium, large, extralarge
#> 27 https://lastfm-img2.akamaized.net/i/u/34s/7cdf797748268b3385a56445fc751c39.png, https://lastfm-img2.akamaized.net/i/u/64s/7cdf797748268b3385a56445fc751c39.png, https://lastfm-img2.akamaized.net/i/u/174s/7cdf797748268b3385a56445fc751c39.png, https://lastfm-img2.akamaized.net/i/u/300x300/7cdf797748268b3385a56445fc751c39.png, small, medium, large, extralarge
#> 28 https://lastfm-img2.akamaized.net/i/u/34s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/64s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/174s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/300x300/c8f8be9e530f4956a70c00d46c272fbe.png, small, medium, large, extralarge
#> 29 https://lastfm-img2.akamaized.net/i/u/34s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/64s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/174s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/300x300/2b45cccc6ca247a6c78cca6ace9c2114.png, small, medium, large, extralarge
#> 30 https://lastfm-img2.akamaized.net/i/u/34s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/64s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/174s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/300x300/2b45cccc6ca247a6c78cca6ace9c2114.png, small, medium, large, extralarge
#> 31 https://lastfm-img2.akamaized.net/i/u/34s/3d42bab35e7ad7332a27b9ff38db971a.png, https://lastfm-img2.akamaized.net/i/u/64s/3d42bab35e7ad7332a27b9ff38db971a.png, https://lastfm-img2.akamaized.net/i/u/174s/3d42bab35e7ad7332a27b9ff38db971a.png, https://lastfm-img2.akamaized.net/i/u/300x300/3d42bab35e7ad7332a27b9ff38db971a.png, small, medium, large, extralarge
#> 32 https://lastfm-img2.akamaized.net/i/u/34s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/64s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/174s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/300x300/2b45cccc6ca247a6c78cca6ace9c2114.png, small, medium, large, extralarge
#> 33 https://lastfm-img2.akamaized.net/i/u/34s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/64s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/174s/85f1a1719c316244d4a7740c42246da0.png, https://lastfm-img2.akamaized.net/i/u/300x300/85f1a1719c316244d4a7740c42246da0.png, small, medium, large, extralarge
#> 34 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge
#> 35 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png, small, medium, large, extralarge
#> 36 https://lastfm-img2.akamaized.net/i/u/34s/7cdf797748268b3385a56445fc751c39.png, https://lastfm-img2.akamaized.net/i/u/64s/7cdf797748268b3385a56445fc751c39.png, https://lastfm-img2.akamaized.net/i/u/174s/7cdf797748268b3385a56445fc751c39.png, https://lastfm-img2.akamaized.net/i/u/300x300/7cdf797748268b3385a56445fc751c39.png, small, medium, large, extralarge
#> 37 https://lastfm-img2.akamaized.net/i/u/34s/56e0b6ab50624765c7ea12fd8040f000.png, https://lastfm-img2.akamaized.net/i/u/64s/56e0b6ab50624765c7ea12fd8040f000.png, https://lastfm-img2.akamaized.net/i/u/174s/56e0b6ab50624765c7ea12fd8040f000.png, https://lastfm-img2.akamaized.net/i/u/300x300/56e0b6ab50624765c7ea12fd8040f000.png, small, medium, large, extralarge
#> 38 https://lastfm-img2.akamaized.net/i/u/34s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/64s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/174s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/300x300/73dba10b71674e029b3ddc0692899e0b.png, small, medium, large, extralarge
#> 39 https://lastfm-img2.akamaized.net/i/u/34s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/64s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/174s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/300x300/c8f8be9e530f4956a70c00d46c272fbe.png, small, medium, large, extralarge
#> 40 https://lastfm-img2.akamaized.net/i/u/34s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/64s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/174s/4e867d7c2077fe42e03df7d9d1e7ac78.png, https://lastfm-img2.akamaized.net/i/u/300x300/4e867d7c2077fe42e03df7d9d1e7ac78.png, small, medium, large, extralarge
#> 41 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png, small, medium, large, extralarge
#> 42 https://lastfm-img2.akamaized.net/i/u/34s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/64s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/174s/03e320dd70c0577c42c459eac0063c16.png, https://lastfm-img2.akamaized.net/i/u/300x300/03e320dd70c0577c42c459eac0063c16.png, small, medium, large, extralarge
#> 43 https://lastfm-img2.akamaized.net/i/u/34s/e009139061502922de33ed960c5fac9c.png, https://lastfm-img2.akamaized.net/i/u/64s/e009139061502922de33ed960c5fac9c.png, https://lastfm-img2.akamaized.net/i/u/174s/e009139061502922de33ed960c5fac9c.png, https://lastfm-img2.akamaized.net/i/u/300x300/e009139061502922de33ed960c5fac9c.png, small, medium, large, extralarge
#> 44 https://lastfm-img2.akamaized.net/i/u/34s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/64s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/174s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/300x300/2b45cccc6ca247a6c78cca6ace9c2114.png, small, medium, large, extralarge
#> 45 https://lastfm-img2.akamaized.net/i/u/34s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/64s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/174s/73dba10b71674e029b3ddc0692899e0b.png, https://lastfm-img2.akamaized.net/i/u/300x300/73dba10b71674e029b3ddc0692899e0b.png, small, medium, large, extralarge
#> 46 https://lastfm-img2.akamaized.net/i/u/34s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/64s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/174s/c8f8be9e530f4956a70c00d46c272fbe.png, https://lastfm-img2.akamaized.net/i/u/300x300/c8f8be9e530f4956a70c00d46c272fbe.png, small, medium, large, extralarge
#> 47 https://lastfm-img2.akamaized.net/i/u/34s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/64s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/174s/f48930d08bfc7c2caf7631d6bdcdb91c.png, https://lastfm-img2.akamaized.net/i/u/300x300/f48930d08bfc7c2caf7631d6bdcdb91c.png, small, medium, large, extralarge
#> 48 https://lastfm-img2.akamaized.net/i/u/34s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/64s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/174s/ac5489a01d41cb391ab5f745d157a291.png, https://lastfm-img2.akamaized.net/i/u/300x300/ac5489a01d41cb391ab5f745d157a291.png, small, medium, large, extralarge
#> 49 https://lastfm-img2.akamaized.net/i/u/34s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/64s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/174s/2b45cccc6ca247a6c78cca6ace9c2114.png, https://lastfm-img2.akamaized.net/i/u/300x300/2b45cccc6ca247a6c78cca6ace9c2114.png, small, medium, large, extralarge
#> 50 https://lastfm-img2.akamaized.net/i/u/34s/e009139061502922de33ed960c5fac9c.png, https://lastfm-img2.akamaized.net/i/u/64s/e009139061502922de33ed960c5fac9c.png, https://lastfm-img2.akamaized.net/i/u/174s/e009139061502922de33ed960c5fac9c.png, https://lastfm-img2.akamaized.net/i/u/300x300/e009139061502922de33ed960c5fac9c.png, small, medium, large, extralarge
#>    rank
#> 1     1
#> 2     2
#> 3     3
#> 4     4
#> 5     5
#> 6     6
#> 7     7
#> 8     8
#> 9     9
#> 10   10
#> 11   11
#> 12   12
#> 13   13
#> 14   14
#> 15   15
#> 16   16
#> 17   17
#> 18   18
#> 19   19
#> 20   20
#> 21   21
#> 22   22
#> 23   23
#> 24   24
#> 25   25
#> 26   26
#> 27   27
#> 28   28
#> 29   29
#> 30   30
#> 31   31
#> 32   32
#> 33   33
#> 34   34
#> 35   35
#> 36   36
#> 37   37
#> 38   38
#> 39   39
#> 40   40
#> 41   41
#> 42   42
#> 43   43
#> 44   44
#> 45   45
#> 46   46
#> 47   47
#> 48   48
#> 49   49
#> 50   50
#> 
#> $`@attr`
#> $`@attr`$user
#> [1] "platyjus"
#> 
#> $`@attr`$page
#> [1] "1"
#> 
#> $`@attr`$perPage
#> [1] "50"
#> 
#> $`@attr`$totalPages
#> [1] "53"
#> 
#> $`@attr`$total
#> [1] "2644"
#> 
#>