Get an album chart for a user profile.

Get an album chart for a user profile, for a given date range. If no date range is supplied, it will return the most recent album chart for this user. Implementation of last.fm's user.getWeeklyAlbumChart API method (http://www.last.fm/api/show/user.getWeeklyAlbumChart).

user_getWeeklyAlbumChart(user, from = NA, to = NA)

Arguments

user
The last.fm username to fetch the charts of.

from
The date at which the chart should start from.

to
The date at which the chart should end on.

Value

A list of an album chart for a user profile.

Examples

user_getWeeklyAlbumChart("platyjus")

#>                                                                                 name
#> 1                                              The Louder I Call, the Faster It Runs
#> 2                                                                         Redemption
#> 3                                                                       Aromanticism
#> 4                                                                Sink into the Floor
#> 5                                                                         Feng Suave
#> 6                                                   Hasta la Raíz (Edición Especial)
#> 7                                                            Shriek (Deluxe Version)
#> 8                                                         4eva Is a Mighty Long Time
#> 9                                 Black Panther The Album Music From And Inspired By
#> 10                                                    Brick Body Kids Still Daydream
#> 11                                                               Bucket List Project
#> 12                                                                             Drunk
#> 13                                                                 In The Comfort Of
#> 14                                                                              ISON
#> 15                                                                          November
#> 16                                                                           Process
#> 17                                                                       The Iceberg
#> 18                                                                           The OOZ
#> 19                                                                    'Til It's Over
#> 20                                    A Collection of Fleeting Moments and Daydreams
#> 21                       After The Storm (feat. Tyler, The Creator & Bootsy Collins)
#> 22                                                                               All
#> 23                                         American Boyfriend: A Suburban Love Story
#> 24                                                                           Any Day
#> 25                                                 Bishounè: Alma Del Huila (Deluxe)
#> 26                                                                       Black Radio
#> 27                                                                   Body And Shadow
#> 28                                                          Call It Home: Vol. 1 & 2
#> 29                                                                           Canyons
#> 30                                                                       care for me
#> 31                                                                       Cilvia Demo
#> 32                                                                   Coastal Grooves
#> 33                                                         Czarface Meets Metal Face
#> 34                                                                            Déjàvu
#> 35                                                             Elevators: Act I & II
#> 36                                                                             Feels
#> 37                                                                    Follow Me Home
#> 38                                                                           Freddie
#> 39                                                                    Freetown Sound
#> 40                                                          Get The Fuck Off My Dick
#> 41                                                                            Gladly
#> 42                                                                         Gladwemet
#> 43                                                                Grand Paon de Nuit
#> 44                                                                  Heaven and Earth
#> 45                                                                             Heavn
#> 46                                                                   Hit the Ceiling
#> 47                                                                   I Am Not Afraid
#> 48                                                                               III
#> 49                                                                               Jp3
#> 50                                                            Letters of Irrelevance
#> 51                                                                     Little Dragon
#> 52                                                                    Man On My Mind
#> 53                                                                           Mirrors
#> 54 Musas (Un Homenaje al Folclore Latinoamericano en Manos de Los Macorinos, Vol. 1)
#> 55                                                                Natalia Lafourcade
#> 56                                                                   Nothing Matters
#> 57                                                                     Peace of Mind
#> 58                                                                   Puzzles, Vol. 3
#> 59                                                                     Rap Album One
#> 60                                                             Room Inside the World
#> 61                                                                    So sad so sexy
#> 62                                                                       Space Cries
#> 63                                                                    Summertime '06
#> 64                                                                       Sun Go Nova
#> 65                                                                     Take Me Apart
#> 66                                                                          Telefone
#> 67                                                            The Colour In Anything
#> 68                                                             The Healing Component
#> 69                                                                  The Sun's Tirade
#> 70                                                                           Toss Up
#> 71                                                                              TTYL
#> 72                                                                            Virtue
#> 73                                                                           Wave[s]
#> 74                                                                       When We Are
#> 75                                                                        Wrap It Up
#> 76                                                           Your Queen Is a Reptile
#> 77                                                                         Come Over
#> 78                                           Should I Take You Home (Keyloc Version)
#> 79                                                                            Tercel
#> 80                                                                             saman
#> 81                                                                Wondering Wanderer
#>                                    mbid playcount
#> 1                                              20
#> 2                                              17
#> 3                                              10
#> 4                                               5
#> 5                                               4
#> 6                                               4
#> 7                                               4
#> 8                                               2
#> 9                                               2
#> 10                                              2
#> 11                                              2
#> 12                                              2
#> 13                                              2
#> 14                                              2
#> 15                                              2
#> 16                                              2
#> 17                                              2
#> 18                                              2
#> 19                                              1
#> 20                                              1
#> 21                                              1
#> 22                                              1
#> 23                                              1
#> 24                                              1
#> 25                                              1
#> 26 b2cca2a2-4150-4560-a8ac-3fb4e4f3a239         1
#> 27                                              1
#> 28                                              1
#> 29                                              1
#> 30                                              1
#> 31 20aa9b51-08b4-4f09-b88c-7017eaaea971         1
#> 32 9137d3e7-23fb-47ce-900b-004059170e9d         1
#> 33                                              1
#> 34                                              1
#> 35                                              1
#> 36                                              1
#> 37 b0777432-5282-4f97-ac2a-de8aca33187b         1
#> 38                                              1
#> 39                                              1
#> 40                                              1
#> 41                                              1
#> 42                                              1
#> 43                                              1
#> 44                                              1
#> 45                                              1
#> 46                                              1
#> 47 aa20be57-486e-41d5-ab10-e6393cb0a289         1
#> 48 4a681db6-3146-4166-b997-6db77bf796dc         1
#> 49                                              1
#> 50                                              1
#> 51 5c1f9dc2-fb74-4e32-8f00-4fc1c9d530b2         1
#> 52                                              1
#> 53                                              1
#> 54                                              1
#> 55 1e52db35-52c9-43f3-ba7e-550e4b878b12         1
#> 56                                              1
#> 57                                              1
#> 58                                              1
#> 59 89dc83dd-f6eb-4912-882a-b5ebd47f3a15         1
#> 60                                              1
#> 61                                              1
#> 62                                              1
#> 63                                              1
#> 64                                              1
#> 65                                              1
#> 66                                              1
#> 67                                              1
#> 68                                              1
#> 69                                              1
#> 70                                              1
#> 71                                              1
#> 72                                              1
#> 73                                              1
#> 74                                              1
#> 75                                              1
#> 76                                              1
#> 77                                              1
#> 78                                              1
#> 79                                              1
#> 80                                              1
#> 81                                              1
#>                                                                                                                               url
#> 1                                                         https://www.last.fm/music/Wye+Oak/The+Louder+I+Call,+the+Faster+It+Runs
#> 2                                                                                   https://www.last.fm/music/Jay+Rock/Redemption
#> 3                                                                             https://www.last.fm/music/Moses+Sumney/Aromanticism
#> 4                                                                        https://www.last.fm/music/Feng+Suave/Sink+into+the+Floor
#> 5                                                                                 https://www.last.fm/music/Feng+Suave/Feng+Suave
#> 6                                         https://www.last.fm/music/Natalia+Lafourcade/Hasta+la+Ra%C3%ADz+(Edici%C3%B3n+Especial)
#> 7                                                                       https://www.last.fm/music/Wye+Oak/Shriek+(Deluxe+Version)
#> 8                                                               https://www.last.fm/music/Big+K.R.I.T./4eva+Is+a+Mighty+Long+Time
#> 9                                            https://www.last.fm/music/Ab-Soul/Black+Panther+The+Album+Music+From+And+Inspired+By
#> 10                                                       https://www.last.fm/music/Open+Mike+Eagle/Brick+Body+Kids+Still+Daydream
#> 11                                                                             https://www.last.fm/music/Saba/Bucket+List+Project
#> 12                                                                                     https://www.last.fm/music/Thundercat/Drunk
#> 13                                                                              https://www.last.fm/music/Sango/In+The+Comfort+Of
#> 14                                                                                       https://www.last.fm/music/Sevdaliza/ISON
#> 15                                                                                         https://www.last.fm/music/Sir/November
#> 16                                                                                       https://www.last.fm/music/Sampha/Process
#> 17                                                                                  https://www.last.fm/music/Oddisee/The+Iceberg
#> 18                                                                                   https://www.last.fm/music/King+Krule/The+OOZ
#> 19                                                                    https://www.last.fm/music/Anderson+.Paak/%27Til+It%27s+Over
#> 20                                             https://www.last.fm/music/Orion+Sun/A+Collection+of+Fleeting+Moments+and+Daydreams
#> 21                               https://www.last.fm/music/Kali+Uchis/After+The+Storm+(feat.+Tyler,+The+Creator+&+Bootsy+Collins)
#> 22                                                                                       https://www.last.fm/music/Umi+Copper/All
#> 23                                             https://www.last.fm/music/Kevin+Abstract/American+Boyfriend:+A+Suburban+Love+Story
#> 24                                                                             https://www.last.fm/music/The+Sea+and+Cake/Any+Day
#> 25                                   https://www.last.fm/music/Gabriel+Garz%C3%B3n-Montano/Bishoun%C3%A8:+Alma+Del+Huila+(Deluxe)
#> 26                                                                https://www.last.fm/music/Robert+Glasper+Experiment/Black+Radio
#> 27                                                    https://www.last.fm/music/Brian+Blade+&+The+Fellowship+Band/Body+And+Shadow
#> 28                                                   https://www.last.fm/music/The+California+Honeydrops/Call+It+Home:+Vol.+1+&+2
#> 29                                                                                       https://www.last.fm/music/Barrie/Canyons
#> 30                                                                                     https://www.last.fm/music/Saba/care+for+me
#> 31                                                                            https://www.last.fm/music/Isaiah+Rashad/Cilvia+Demo
#> 32                                                                         https://www.last.fm/music/Blood+Orange/Coastal+Grooves
#> 33                                                                   https://www.last.fm/music/Czarface/Czarface+Meets+Metal+Face
#> 34                                                                               https://www.last.fm/music/Matty/D%C3%A9j%C3%A0vu
#> 35                                                                   https://www.last.fm/music/Bishop+Nehru/Elevators:+Act+I+&+II
#> 36                                                                                   https://www.last.fm/music/Snoh+Aalegra/Feels
#> 37                                                                              https://www.last.fm/music/Jay+Rock/Follow+Me+Home
#> 38                                                                                https://www.last.fm/music/Freddie+Gibbs/Freddie
#> 39                                                                          https://www.last.fm/music/Blood+Orange/Freetown+Sound
#> 40                                                               https://www.last.fm/music/Vince+Staples/Get+The+Fuck+Off+My+Dick
#> 41                                                                                        https://www.last.fm/music/Tirzah/Gladly
#> 42                                                                                  https://www.last.fm/music/Knxwledge/Gladwemet
#> 43                                                                          https://www.last.fm/music/Palatine/Grand+Paon+de+Nuit
#> 44                                                                   https://www.last.fm/music/Kamasi+Washington/Heaven+and+Earth
#> 45                                                                                   https://www.last.fm/music/Jamila+Woods/Heavn
#> 46                                                                            https://www.last.fm/music/Lion+Babe/Hit+the+Ceiling
#> 47                                                                        https://www.last.fm/music/Hugh+Masekela/I+Am+Not+Afraid
#> 48                                                                                    https://www.last.fm/music/BADBADNOTGOOD/III
#> 49                                                                                      https://www.last.fm/music/JUNGLEPUSSY/Jp3
#> 50                                                              https://www.last.fm/music/Patrick+Paige+II/Letters+of+Irrelevance
#> 51                                                                          https://www.last.fm/music/Little+Dragon/Little+Dragon
#> 52                                                                         https://www.last.fm/music/Cornelia+Murr/Man+On+My+Mind
#> 53                                                                                  https://www.last.fm/music/Monophonics/Mirrors
#> 54 https://www.last.fm/music/Natalia+Lafourcade/Musas+(Un+Homenaje+al+Folclore+Latinoamericano+en+Manos+de+Los+Macorinos,+Vol.+1)
#> 55                                                                https://www.last.fm/music/Natalia+Lafourcade/Natalia+Lafourcade
#> 56                                                                 https://www.last.fm/music/Benjamin+Lazar+Davis/Nothing+Matters
#> 57                                                                          https://www.last.fm/music/Little+Dragon/Peace+of+Mind
#> 58                                                                    https://www.last.fm/music/Nitai+Hershkovits/Puzzles,+Vol.+3
#> 59                                                                               https://www.last.fm/music/Jonwayne/Rap+Album+One
#> 60                                                                          https://www.last.fm/music/Ought/Room+Inside+the+World
#> 61                                                                              https://www.last.fm/music/Lykke+Li/So+sad+so+sexy
#> 62                                                                                 https://www.last.fm/music/Ed+Mount/Space+Cries
#> 63                                                                       https://www.last.fm/music/Vince+Staples/Summertime+%2706
#> 64                                                                           https://www.last.fm/music/Denmark+Vessey/Sun+Go+Nova
#> 65                                                                                 https://www.last.fm/music/Kelela/Take+Me+Apart
#> 66                                                                                      https://www.last.fm/music/NoName/Telefone
#> 67                                                                   https://www.last.fm/music/James+Blake/The+Colour+In+Anything
#> 68                                                                   https://www.last.fm/music/Mick+Jenkins/The+Healing+Component
#> 69                                                                     https://www.last.fm/music/Isaiah+Rashad/The+Sun%27s+Tirade
#> 70                                                                                https://www.last.fm/music/Kevin+Krauter/Toss+Up
#> 71                                                                                https://www.last.fm/music/Morabeza+Tobacco/TTYL
#> 72                                                                                     https://www.last.fm/music/The+Voidz/Virtue
#> 73                                                                             https://www.last.fm/music/Mick+Jenkins/Wave%5Bs%5D
#> 74                                                                             https://www.last.fm/music/Nubya+Garcia/When+We+Are
#> 75                                                                               https://www.last.fm/music/Pomplamoose/Wrap+It+Up
#> 76                                                                https://www.last.fm/music/Sons+of+Kemet/Your+Queen+Is+a+Reptile
#> 77                                                                               https://www.last.fm/music/The+Internet/Come+Over
#> 78                                        https://www.last.fm/music/Sunny+&+The+Sunliners/Should+I+Take+You+Home+(Keyloc+Version)
#> 79                                                                                  https://www.last.fm/music/Les+Louanges/Tercel
#> 80                                                                            https://www.last.fm/music/%C3%93lafur+Arnalds/saman
#> 81                                                                  https://www.last.fm/music/Olivier+St.Louis/Wondering+Wanderer
#>                         artist.#text                          artist.mbid
#> 1                            Wye Oak 3c8ff4c9-b5b9-4e78-ad9d-5157c323de24
#> 2                           Jay Rock 9880800a-f6f8-4f8f-a00e-b4b664776c0c
#> 3                       Moses Sumney 89c081d4-2ab2-4d3e-8589-ad77dfc40384
#> 4                         Feng Suave                                     
#> 5                         Feng Suave                                     
#> 6                 Natalia Lafourcade 2ceb4e66-4eaa-4dba-ad3a-30df3b742557
#> 7                            Wye Oak 3c8ff4c9-b5b9-4e78-ad9d-5157c323de24
#> 8                       Big K.R.I.T. 7e966957-9b0b-4e19-b8a0-4206535d8052
#> 9                            Ab-Soul acc12019-10e3-40aa-b4bf-66fbb4a692af
#> 10                   Open Mike Eagle 4508a423-daaa-4d43-a839-aa1a9be88a9c
#> 11                              Saba c2311581-df24-451e-8232-a9dcee3bfa62
#> 12                        Thundercat 044fd265-79dd-43eb-afc4-8b20becf7e17
#> 13                             Sango 150aacdf-a15f-46b8-8a99-25cde68e90de
#> 14                         Sevdaliza                                     
#> 15                               Sir 96966179-d9ef-46f4-9179-5545cd0e6279
#> 16                            Sampha 130c2184-14f2-4180-a216-04b5a333be57
#> 17                           Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83
#> 18                        King Krule c0daae21-ec2a-4dd0-a323-4f53f5945d04
#> 19                    Anderson .Paak                                     
#> 20                         Orion Sun                                     
#> 21                        Kali Uchis                                     
#> 22                        Umi Copper                                     
#> 23                    Kevin Abstract                                     
#> 24                  The Sea and Cake 74171cb6-1ff8-42bd-be8a-96660c22e3e0
#> 25            Gabriel Garzón-Montano 4cc212e1-0c06-4675-9c8b-bf6c4fb05769
#> 26         Robert Glasper Experiment 0ec3c9dc-45e7-4bc1-9813-a866acaae772
#> 27 Brian Blade & The Fellowship Band 4d54e19b-f78d-4d2b-bdcd-a8d0348d6e6f
#> 28         The California Honeydrops c144dbec-b0f1-4aae-b2e6-5087e44806c8
#> 29                            Barrie                                     
#> 30                              Saba c2311581-df24-451e-8232-a9dcee3bfa62
#> 31                     Isaiah Rashad 37547347-ea75-45fa-9766-ff36b44d8147
#> 32                      Blood Orange e43daa78-3cf3-4cf9-bc27-af4d4e675a02
#> 33                          Czarface 66120159-5fe7-4cdb-8c9a-7ac68e393c6f
#> 34                             Matty                                     
#> 35                      Bishop Nehru 0735e073-b829-4c10-9144-8ca286ec68b2
#> 36                      Snoh Aalegra 636e0fb6-f2ad-45fc-b385-4d994665e7e3
#> 37                          Jay Rock 9880800a-f6f8-4f8f-a00e-b4b664776c0c
#> 38                     Freddie Gibbs 21645c31-fe1c-45a4-955c-3e172b12c3f9
#> 39                      Blood Orange e43daa78-3cf3-4cf9-bc27-af4d4e675a02
#> 40                     Vince Staples 78e854b8-9713-4ff2-9218-6b3784893bff
#> 41                            Tirzah f697bed6-3b4e-4352-b6fb-70550afd8dcb
#> 42                         Knxwledge 7df4df53-17ba-4099-90ff-c5528762c5a7
#> 43                          Palatine                                     
#> 44                 Kamasi Washington                                     
#> 45                      Jamila Woods                                     
#> 46                         Lion Babe c9dda914-1eb2-4762-8006-208010b89ee3
#> 47                     Hugh Masekela 83a2e0fb-3d9b-46ff-b383-3dde54c15297
#> 48                     BADBADNOTGOOD 754294d5-d7d2-4ea2-8184-1dcaaf55a56f
#> 49                       JUNGLEPUSSY 85f40e9b-c639-4783-b621-80d873615e04
#> 50                  Patrick Paige II                                     
#> 51                     Little Dragon c8a9ae11-63bf-409b-8bf3-a96b3712a2ff
#> 52                     Cornelia Murr                                     
#> 53                       Monophonics f1745cca-1cae-4696-a0d3-62553c742b0c
#> 54                Natalia Lafourcade 2ceb4e66-4eaa-4dba-ad3a-30df3b742557
#> 55                Natalia Lafourcade 2ceb4e66-4eaa-4dba-ad3a-30df3b742557
#> 56              Benjamin Lazar Davis                                     
#> 57                     Little Dragon c8a9ae11-63bf-409b-8bf3-a96b3712a2ff
#> 58                 Nitai Hershkovits                                     
#> 59                          Jonwayne 70bd369d-f58a-48ab-96eb-299d9b89a3c7
#> 60                             Ought a039f1f1-e664-4aee-a917-426514734aa0
#> 61                          Lykke Li cc7d4686-ea02-45fd-956e-94c1a322558c
#> 62                          Ed Mount                                     
#> 63                     Vince Staples 78e854b8-9713-4ff2-9218-6b3784893bff
#> 64                    Denmark Vessey                                     
#> 65                            Kelela 3599cf21-2ed3-405d-a1f1-b52ff7b28df8
#> 66                            NoName 9393cc98-b014-493c-bdf1-a643a088f662
#> 67                       James Blake 8dc08b1f-e393-4f85-a5dd-300f7693a8b8
#> 68                      Mick Jenkins 64c4dfcc-ba2b-4036-ac01-1636ae15ebb8
#> 69                     Isaiah Rashad 37547347-ea75-45fa-9766-ff36b44d8147
#> 70                     Kevin Krauter                                     
#> 71                  Morabeza Tobacco                                     
#> 72                         The Voidz                                     
#> 73                      Mick Jenkins 64c4dfcc-ba2b-4036-ac01-1636ae15ebb8
#> 74                      Nubya Garcia                                     
#> 75                       Pomplamoose 63d61461-ecfe-4057-896b-3ba3ff3111eb
#> 76                     Sons of Kemet fee950a2-e68c-4e76-b1ae-ff6efdf68b11
#> 77                      The Internet 45804465-4271-4e6a-b881-ce668ef09530
#> 78             Sunny & The Sunliners                                     
#> 79                      Les Louanges                                     
#> 80                    Ólafur Arnalds 6655955b-1c1e-4bcb-84e4-81bcd9efab30
#> 81                  Olivier St.Louis                                     
#>    @attr.rank
#> 1           1
#> 2           2
#> 3           3
#> 4           4
#> 5           5
#> 6           6
#> 7           7
#> 8           8
#> 9           9
#> 10         10
#> 11         11
#> 12         12
#> 13         13
#> 14         14
#> 15         15
#> 16         16
#> 17         17
#> 18         18
#> 19         19
#> 20         20
#> 21         21
#> 22         22
#> 23         23
#> 24         24
#> 25         25
#> 26         26
#> 27         27
#> 28         28
#> 29         29
#> 30         30
#> 31         31
#> 32         32
#> 33         33
#> 34         34
#> 35         35
#> 36         36
#> 37         37
#> 38         38
#> 39         39
#> 40         40
#> 41         41
#> 42         42
#> 43         43
#> 44         44
#> 45         45
#> 46         46
#> 47         47
#> 48         48
#> 49         49
#> 50         50
#> 51         51
#> 52         52
#> 53         53
#> 54         54
#> 55         55
#> 56         56
#> 57         57
#> 58         58
#> 59         59
#> 60         60
#> 61         61
#> 62         62
#> 63         63
#> 64         64
#> 65         65
#> 66         66
#> 67         67
#> 68         68
#> 69         69
#> 70         70
#> 71         71
#> 72         72
#> 73         73
#> 74         74
#> 75         75
#> 76         76
#> 77         77
#> 78         78
#> 79         79
#> 80         80
#> 81         81