Get an artist chart for a user profile.

Get an artist chart for a user profile, for a given date range. If no date range is supplied, it will return the most recent artist chart for this user. Implementation of last.fm's user.getWeeklyArtistChart API method (http://www.last.fm/api/show/user.getWeeklyArtistChart).

user_getWeeklyArtistChart(user, from = NA, to = NA, api_key = lastkey)

Arguments

user
The last.fm username to fetch the charts of.

from
The date at which the chart should start from. See user_getChartsList for more.

to
The date at which the chart should end on. See user_getChartsList for more.

api_key
A Last.fm API key.

Value

A list of an artist chart for a user profile.

Examples

user_getWeeklyArtistChart("platyjus")

#> $artist
#>               name                                 mbid playcount
#> 1   The Wild Reeds 54412058-8ed9-49a2-b931-ce8071caf851        19
#> 2     Joan Shelley                                              9
#> 3 Childish Gambino 7fb57fba-a6ef-44c2-abab-2fa3bdee607e         4
#> 4          Oddisee aaae09b3-70d9-4a65-bb3f-57e85dfe9b83         4
#> 5        Sam Cooke 6c8fd0be-d961-454c-aee0-4366a6dbc993         4
#> 6           Khalid c6bfb05d-f570-46c8-98e1-e25441189770         2
#>                                          url rank
#> 1   https://www.last.fm/music/The+Wild+Reeds    1
#> 2     https://www.last.fm/music/Joan+Shelley    2
#> 3 https://www.last.fm/music/Childish+Gambino    3
#> 4          https://www.last.fm/music/Oddisee    4
#> 5        https://www.last.fm/music/Sam+Cooke    5
#> 6           https://www.last.fm/music/Khalid    6
#> 
#> $`@attr`
#> $`@attr`$user
#> [1] "platyjus"
#> 
#> $`@attr`$from
#> [1] "1494504000"
#> 
#> $`@attr`$to
#> [1] "1495108800"
#> 
#>