Get a list of available charts for this user.

Get a list of available charts for this user, expressed as date ranges which can be sent to the chart services. Implementation of last.fm's user.getWeeklyChartList API method (http://www.last.fm/api/show/user.getWeeklyChartList).

user_getWeeklyChartList(user)

Arguments

user
TThe last.fm username to fetch the charts list for.

Value

A list of an artist chart for a user profile.

Examples

user_getWeeklyChartList("platyjus")

#>     #text       from         to
#> 1         1108296000 1108900800
#> 2         1108900800 1109505600
#> 3         1109505600 1110110400
#> 4         1110110400 1110715200
#> 5         1110715200 1111320000
#> 6         1111320000 1111924800
#> 7         1111924800 1112529600
#> 8         1112529600 1113134400
#> 9         1113134400 1113739200
#> 10        1113739200 1114344000
#> 11        1114344000 1114948800
#> 12        1114948800 1115553600
#> 13        1115553600 1116158400
#> 14        1116158400 1116763200
#> 15        1116763200 1117368000
#> 16        1117368000 1117972800
#> 17        1117972800 1118577600
#> 18        1118577600 1119182400
#> 19        1119182400 1119787200
#> 20        1119787200 1120392000
#> 21        1120392000 1120996800
#> 22        1120996800 1121601600
#> 23        1121601600 1122206400
#> 24        1122206400 1122811200
#> 25        1122811200 1123416000
#> 26        1123416000 1124020800
#> 27        1124020800 1124625600
#> 28        1124625600 1125230400
#> 29        1125230400 1125835200
#> 30        1125835200 1126440000
#> 31        1126440000 1127044800
#> 32        1127044800 1127649600
#> 33        1127649600 1128254400
#> 34        1128254400 1128859200
#> 35        1128859200 1129464000
#> 36        1129464000 1130068800
#> 37        1130068800 1130673600
#> 38        1130673600 1131278400
#> 39        1131278400 1131883200
#> 40        1131883200 1132488000
#> 41        1132488000 1133092800
#> 42        1133092800 1133697600
#> 43        1133697600 1134302400
#> 44        1134302400 1134907200
#> 45        1134907200 1135512000
#> 46        1135512000 1136116800
#> 47        1136116800 1136721600
#> 48        1136721600 1137326400
#> 49        1137326400 1137931200
#> 50        1137931200 1138536000
#> 51        1138536000 1139140800
#> 52        1139140800 1139745600
#> 53        1139745600 1140350400
#> 54        1140350400 1140955200
#> 55        1140955200 1141560000
#> 56        1141560000 1142164800
#> 57        1142164800 1142769600
#> 58        1142769600 1143374400
#> 59        1143374400 1143979200
#> 60        1143979200 1144584000
#> 61        1144584000 1145188800
#> 62        1145188800 1145793600
#> 63        1145793600 1146398400
#> 64        1146398400 1147003200
#> 65        1147003200 1147608000
#> 66        1147608000 1148212800
#> 67        1148212800 1148817600
#> 68        1148817600 1149422400
#> 69        1149422400 1150027200
#> 70        1150027200 1150632000
#> 71        1150632000 1151236800
#> 72        1151236800 1151841600
#> 73        1151841600 1152446400
#> 74        1152446400 1153051200
#> 75        1153051200 1153656000
#> 76        1153656000 1154260800
#> 77        1154260800 1154865600
#> 78        1154865600 1155470400
#> 79        1155470400 1156075200
#> 80        1156075200 1156680000
#> 81        1156680000 1157284800
#> 82        1157284800 1157889600
#> 83        1157889600 1158494400
#> 84        1158494400 1159099200
#> 85        1159099200 1159704000
#> 86        1159704000 1160308800
#> 87        1160308800 1160913600
#> 88        1160913600 1161518400
#> 89        1161518400 1162123200
#> 90        1162123200 1162728000
#> 91        1162728000 1163332800
#> 92        1163332800 1163937600
#> 93        1163937600 1164542400
#> 94        1164542400 1165147200
#> 95        1165147200 1165752000
#> 96        1165752000 1166356800
#> 97        1166356800 1166961600
#> 98        1166961600 1167566400
#> 99        1167566400 1168171200
#> 100       1168171200 1168776000
#> 101       1168776000 1169380800
#> 102       1169380800 1169985600
#> 103       1169985600 1170590400
#> 104       1170590400 1171195200
#> 105       1171195200 1171800000
#> 106       1171800000 1172404800
#> 107       1172404800 1173009600
#> 108       1173009600 1173614400
#> 109       1173614400 1174219200
#> 110       1174219200 1174824000
#> 111       1174824000 1175428800
#> 112       1175428800 1176033600
#> 113       1176033600 1176638400
#> 114       1176638400 1177243200
#> 115       1177243200 1177848000
#> 116       1177848000 1178452800
#> 117       1178452800 1179057600
#> 118       1179057600 1179662400
#> 119       1179662400 1180267200
#> 120       1180267200 1180872000
#> 121       1180872000 1181476800
#> 122       1181476800 1182081600
#> 123       1182081600 1182686400
#> 124       1182686400 1183291200
#> 125       1183291200 1183896000
#> 126       1183896000 1184500800
#> 127       1184500800 1185105600
#> 128       1185105600 1185710400
#> 129       1185710400 1186315200
#> 130       1186315200 1186920000
#> 131       1186920000 1187524800
#> 132       1187524800 1188129600
#> 133       1188129600 1188734400
#> 134       1188734400 1189339200
#> 135       1189339200 1189944000
#> 136       1189944000 1190548800
#> 137       1190548800 1191153600
#> 138       1191153600 1191758400
#> 139       1191758400 1192363200
#> 140       1192363200 1192968000
#> 141       1192968000 1193572800
#> 142       1193572800 1194177600
#> 143       1194177600 1194782400
#> 144       1194782400 1195387200
#> 145       1195387200 1195992000
#> 146       1195992000 1196596800
#> 147       1196596800 1197201600
#> 148       1197201600 1197806400
#> 149       1197806400 1198411200
#> 150       1198411200 1199016000
#> 151       1199016000 1199620800
#> 152       1199620800 1200225600
#> 153       1200225600 1200830400
#> 154       1200830400 1201435200
#> 155       1201435200 1202040000
#> 156       1202040000 1202644800
#> 157       1202644800 1203249600
#> 158       1203249600 1203854400
#> 159       1203854400 1204459200
#> 160       1204459200 1205064000
#> 161       1205064000 1205668800
#> 162       1205668800 1206273600
#> 163       1206273600 1206878400
#> 164       1206878400 1207483200
#> 165       1207483200 1208088000
#> 166       1208088000 1208692800
#> 167       1208692800 1209297600
#> 168       1209297600 1209902400
#> 169       1209902400 1210507200
#> 170       1210507200 1211112000
#> 171       1211112000 1211716800
#> 172       1211716800 1212321600
#> 173       1212321600 1212926400
#> 174       1212926400 1213531200
#> 175       1213531200 1214136000
#> 176       1214136000 1214740800
#> 177       1214740800 1215345600
#> 178       1215345600 1215950400
#> 179       1215950400 1216555200
#> 180       1216555200 1217160000
#> 181       1217160000 1217764800
#> 182       1217764800 1218369600
#> 183       1218369600 1218974400
#> 184       1218974400 1219579200
#> 185       1219579200 1220184000
#> 186       1220184000 1220788800
#> 187       1220788800 1221393600
#> 188       1221393600 1221998400
#> 189       1221998400 1222603200
#> 190       1222603200 1223208000
#> 191       1223208000 1223812800
#> 192       1223812800 1224417600
#> 193       1224417600 1225022400
#> 194       1225022400 1225627200
#> 195       1225627200 1226232000
#> 196       1226232000 1226836800
#> 197       1226836800 1227441600
#> 198       1227441600 1228046400
#> 199       1228046400 1228651200
#> 200       1228651200 1229256000
#> 201       1229256000 1229860800
#> 202       1229860800 1230465600
#> 203       1230465600 1231070400
#> 204       1231070400 1231675200
#> 205       1231675200 1232280000
#> 206       1232280000 1232884800
#> 207       1232884800 1233489600
#> 208       1233489600 1234094400
#> 209       1234094400 1234699200
#> 210       1234699200 1235304000
#> 211       1235304000 1235908800
#> 212       1235908800 1236513600
#> 213       1236513600 1237118400
#> 214       1237118400 1237723200
#> 215       1237723200 1238328000
#> 216       1238328000 1238932800
#> 217       1238932800 1239537600
#> 218       1239537600 1240142400
#> 219       1240142400 1240747200
#> 220       1240747200 1241352000
#> 221       1241352000 1241956800
#> 222       1241956800 1242561600
#> 223       1242561600 1243166400
#> 224       1243166400 1243771200
#> 225       1243771200 1244376000
#> 226       1244376000 1244980800
#> 227       1244980800 1245585600
#> 228       1245585600 1246190400
#> 229       1246190400 1246795200
#> 230       1246795200 1247400000
#> 231       1247400000 1248004800
#> 232       1248004800 1248609600
#> 233       1248609600 1249214400
#> 234       1249214400 1249819200
#> 235       1249819200 1250424000
#> 236       1250424000 1251028800
#> 237       1251028800 1251633600
#> 238       1251633600 1252238400
#> 239       1252238400 1252843200
#> 240       1252843200 1253448000
#> 241       1253448000 1254052800
#> 242       1254052800 1254657600
#> 243       1254657600 1255262400
#> 244       1255262400 1255867200
#> 245       1255867200 1256472000
#> 246       1256472000 1257076800
#> 247       1257076800 1257681600
#> 248       1257681600 1258286400
#> 249       1258286400 1258891200
#> 250       1258891200 1259496000
#> 251       1259496000 1260100800
#> 252       1260100800 1260705600
#> 253       1260705600 1261310400
#> 254       1261310400 1261915200
#> 255       1261915200 1262520000
#> 256       1262520000 1263124800
#> 257       1263124800 1263729600
#> 258       1263729600 1264334400
#> 259       1264334400 1264939200
#> 260       1264939200 1265544000
#> 261       1265544000 1266148800
#> 262       1266148800 1266753600
#> 263       1266753600 1267358400
#> 264       1267358400 1267963200
#> 265       1267963200 1268568000
#> 266       1268568000 1269172800
#> 267       1269172800 1269777600
#> 268       1269777600 1270382400
#> 269       1270382400 1270987200
#> 270       1270987200 1271592000
#> 271       1271592000 1272196800
#> 272       1272196800 1272801600
#> 273       1272801600 1273406400
#> 274       1273406400 1274011200
#> 275       1274011200 1274616000
#> 276       1274616000 1275220800
#> 277       1275220800 1275825600
#> 278       1275825600 1276430400
#> 279       1276430400 1277035200
#> 280       1277035200 1277640000
#> 281       1277640000 1278244800
#> 282       1278244800 1278849600
#> 283       1278849600 1279454400
#> 284       1279454400 1280059200
#> 285       1280059200 1280664000
#> 286       1280664000 1281268800
#> 287       1281268800 1281873600
#> 288       1281873600 1282478400
#> 289       1282478400 1283083200
#> 290       1283083200 1283688000
#> 291       1283688000 1284292800
#> 292       1284292800 1284897600
#> 293       1284897600 1285502400
#> 294       1285502400 1286107200
#> 295       1286107200 1286712000
#> 296       1286712000 1287316800
#> 297       1287316800 1287921600
#> 298       1287921600 1288526400
#> 299       1288526400 1289131200
#> 300       1289131200 1289736000
#> 301       1289736000 1290340800
#> 302       1290340800 1290945600
#> 303       1290945600 1291550400
#> 304       1291550400 1292155200
#> 305       1292155200 1292760000
#> 306       1292760000 1293364800
#> 307       1293364800 1293969600
#> 308       1293969600 1294574400
#> 309       1294574400 1295179200
#> 310       1295179200 1295784000
#> 311       1295784000 1296388800
#> 312       1296388800 1296993600
#> 313       1296993600 1297598400
#> 314       1297598400 1298203200
#> 315       1298203200 1298808000
#> 316       1298808000 1299412800
#> 317       1299412800 1300017600
#> 318       1300017600 1300622400
#> 319       1300622400 1301227200
#> 320       1301227200 1301832000
#> 321       1301832000 1302436800
#> 322       1302436800 1303041600
#> 323       1303041600 1303646400
#> 324       1303646400 1304251200
#> 325       1304251200 1304856000
#> 326       1304856000 1305460800
#> 327       1305460800 1306065600
#> 328       1306065600 1306670400
#> 329       1306670400 1307275200
#> 330       1307275200 1307880000
#> 331       1307880000 1308484800
#> 332       1308484800 1309089600
#> 333       1309089600 1309694400
#> 334       1309694400 1310299200
#> 335       1310299200 1310904000
#> 336       1310904000 1311508800
#> 337       1311508800 1312113600
#> 338       1312113600 1312718400
#> 339       1312718400 1313323200
#> 340       1313323200 1313928000
#> 341       1313928000 1314532800
#> 342       1314532800 1315137600
#> 343       1315137600 1315742400
#> 344       1315742400 1316347200
#> 345       1316347200 1316952000
#> 346       1316952000 1317556800
#> 347       1317556800 1318161600
#> 348       1318161600 1318766400
#> 349       1318766400 1319371200
#> 350       1319371200 1319976000
#> 351       1319976000 1320580800
#> 352       1320580800 1321185600
#> 353       1321185600 1321790400
#> 354       1321790400 1322395200
#> 355       1322395200 1323000000
#> 356       1323000000 1323604800
#> 357       1323604800 1324209600
#> 358       1324209600 1324814400
#> 359       1324814400 1325419200
#> 360       1325419200 1326024000
#> 361       1326024000 1326628800
#> 362       1326628800 1327233600
#> 363       1327233600 1327838400
#> 364       1327838400 1328443200
#> 365       1328443200 1329048000
#> 366       1329048000 1329652800
#> 367       1329652800 1330257600
#> 368       1330257600 1330862400
#> 369       1330862400 1331467200
#> 370       1331467200 1332072000
#> 371       1332072000 1332676800
#> 372       1332676800 1333281600
#> 373       1333281600 1333886400
#> 374       1333886400 1334491200
#> 375       1334491200 1335096000
#> 376       1335096000 1335700800
#> 377       1335700800 1336305600
#> 378       1336305600 1336910400
#> 379       1336910400 1337515200
#> 380       1337515200 1338120000
#> 381       1338120000 1338724800
#> 382       1338724800 1339329600
#> 383       1339329600 1339934400
#> 384       1339934400 1340539200
#> 385       1340539200 1341144000
#> 386       1341144000 1341748800
#> 387       1341748800 1342353600
#> 388       1342353600 1342958400
#> 389       1342958400 1343563200
#> 390       1343563200 1344168000
#> 391       1344168000 1344772800
#> 392       1344772800 1345377600
#> 393       1345377600 1345982400
#> 394       1345982400 1346587200
#> 395       1346587200 1347192000
#> 396       1347192000 1347796800
#> 397       1347796800 1348401600
#> 398       1348401600 1349006400
#> 399       1349006400 1349611200
#> 400       1349611200 1350216000
#> 401       1350216000 1350820800
#> 402       1350820800 1351425600
#> 403       1351425600 1352030400
#> 404       1352030400 1352635200
#> 405       1352635200 1353240000
#> 406       1353240000 1353844800
#> 407       1353844800 1354449600
#> 408       1354449600 1355054400
#> 409       1355054400 1355659200
#> 410       1355659200 1356264000
#> 411       1356264000 1356868800
#> 412       1356868800 1357473600
#> 413       1357473600 1358078400
#> 414       1358078400 1358683200
#> 415       1358683200 1359288000
#> 416       1359288000 1359892800
#> 417       1359892800 1360497600
#> 418       1360497600 1361102400
#> 419       1361102400 1361707200
#> 420       1361707200 1362312000
#> 421       1362312000 1362916800
#> 422       1362916800 1363521600
#> 423       1363521600 1364126400
#> 424       1364126400 1364731200
#> 425       1364731200 1365336000
#> 426       1365336000 1365940800
#> 427       1365940800 1366545600
#> 428       1366545600 1367150400
#> 429       1367150400 1367755200
#> 430       1367755200 1368360000
#> 431       1368360000 1368964800
#> 432       1368964800 1369569600
#> 433       1369569600 1370174400
#> 434       1370174400 1370779200
#> 435       1370779200 1371384000
#> 436       1371384000 1371988800
#> 437       1371988800 1372593600
#> 438       1372593600 1373198400
#> 439       1373198400 1373803200
#> 440       1373803200 1374408000
#> 441       1374408000 1375012800
#> 442       1375012800 1375617600
#> 443       1375617600 1376222400
#> 444       1376222400 1376827200
#> 445       1376827200 1377432000
#> 446       1377432000 1378036800
#> 447       1378036800 1378641600
#> 448       1378641600 1379246400
#> 449       1379246400 1379851200
#> 450       1379851200 1380456000
#> 451       1380456000 1381060800
#> 452       1381060800 1381665600
#> 453       1381665600 1382270400
#> 454       1382270400 1382875200
#> 455       1382875200 1383480000
#> 456       1383480000 1384084800
#> 457       1384084800 1384689600
#> 458       1384689600 1385294400
#> 459       1385294400 1385899200
#> 460       1385899200 1386504000
#> 461       1386504000 1387108800
#> 462       1387108800 1387713600
#> 463       1387713600 1388318400
#> 464       1388318400 1388923200
#> 465       1388923200 1389528000
#> 466       1389528000 1390132800
#> 467       1390132800 1390737600
#> 468       1390737600 1391342400
#> 469       1391342400 1391947200
#> 470       1391947200 1392552000
#> 471       1392552000 1393156800
#> 472       1393156800 1393761600
#> 473       1393761600 1394366400
#> 474       1394366400 1394971200
#> 475       1394971200 1395576000
#> 476       1395576000 1396180800
#> 477       1396180800 1396785600
#> 478       1396785600 1397390400
#> 479       1397390400 1397995200
#> 480       1397995200 1398600000
#> 481       1398600000 1399204800
#> 482       1399204800 1399809600
#> 483       1399809600 1400414400
#> 484       1400414400 1401019200
#> 485       1401019200 1401624000
#> 486       1401624000 1402228800
#> 487       1402228800 1402833600
#> 488       1402833600 1403438400
#> 489       1403438400 1404043200
#> 490       1404043200 1404648000
#> 491       1404648000 1405252800
#> 492       1405252800 1405857600
#> 493       1405857600 1406462400
#> 494       1406462400 1407067200
#> 495       1407067200 1407672000
#> 496       1407672000 1408276800
#> 497       1408276800 1408881600
#> 498       1408881600 1409486400
#> 499       1409486400 1410091200
#> 500       1410091200 1410696000
#> 501       1410696000 1411300800
#> 502       1411300800 1411905600
#> 503       1411905600 1412510400
#> 504       1412510400 1413115200
#> 505       1413115200 1413720000
#> 506       1413720000 1414324800
#> 507       1414324800 1414929600
#> 508       1414929600 1415534400
#> 509       1415534400 1416139200
#> 510       1416139200 1416744000
#> 511       1416744000 1417348800
#> 512       1417348800 1417953600
#> 513       1417953600 1418558400
#> 514       1418558400 1419163200
#> 515       1419163200 1419768000
#> 516       1419768000 1420372800
#> 517       1420372800 1420977600
#> 518       1420977600 1421582400
#> 519       1421582400 1422187200
#> 520       1422187200 1422792000
#> 521       1422792000 1423396800
#> 522       1423396800 1424001600
#> 523       1424001600 1424606400
#> 524       1424606400 1425211200
#> 525       1425211200 1425816000
#> 526       1425816000 1426420800
#> 527       1426420800 1427025600
#> 528       1427025600 1427630400
#> 529       1427630400 1428235200
#> 530       1428235200 1428840000
#> 531       1428840000 1429444800
#> 532       1429444800 1430049600
#> 533       1430049600 1430654400
#> 534       1430654400 1431259200
#> 535       1431259200 1431864000
#> 536       1431864000 1432468800
#> 537       1432468800 1433073600
#> 538       1433073600 1433678400
#> 539       1433678400 1434283200
#> 540       1434283200 1434888000
#> 541       1434888000 1435492800
#> 542       1435492800 1436097600
#> 543       1436097600 1436702400
#> 544       1436702400 1437307200
#> 545       1437307200 1437912000
#> 546       1437912000 1438516800
#> 547       1438516800 1439121600
#> 548       1439121600 1439726400
#> 549       1439726400 1440331200
#> 550       1440331200 1440936000
#> 551       1440936000 1441540800
#> 552       1441540800 1442145600
#> 553       1442145600 1442750400
#> 554       1442750400 1443355200
#> 555       1443355200 1443960000
#> 556       1443960000 1444564800
#> 557       1444564800 1445169600
#> 558       1445169600 1445774400
#> 559       1445774400 1446379200
#> 560       1446379200 1446984000
#> 561       1446984000 1447588800
#> 562       1447588800 1448193600
#> 563       1448193600 1448798400
#> 564       1448798400 1449403200
#> 565       1449403200 1450008000
#> 566       1450008000 1450612800
#> 567       1450612800 1451217600
#> 568       1451217600 1451822400
#> 569       1451822400 1452427200
#> 570       1452427200 1453032000
#> 571       1453032000 1453636800
#> 572       1453636800 1454241600
#> 573       1454241600 1454846400
#> 574       1454846400 1455451200
#> 575       1455451200 1456056000
#> 576       1456056000 1456660800
#> 577       1456660800 1457265600
#> 578       1457265600 1457870400
#> 579       1457870400 1458475200
#> 580       1458475200 1459080000
#> 581       1459080000 1459684800
#> 582       1459684800 1460289600
#> 583       1460289600 1460894400
#> 584       1460894400 1461499200
#> 585       1461499200 1462104000
#> 586       1462104000 1462708800
#> 587       1462708800 1463313600
#> 588       1463313600 1463918400
#> 589       1463918400 1464523200
#> 590       1464523200 1465128000
#> 591       1465128000 1465732800
#> 592       1465732800 1466337600
#> 593       1466337600 1466942400
#> 594       1466942400 1467547200
#> 595       1467547200 1468152000
#> 596       1468152000 1468756800
#> 597       1468756800 1469361600
#> 598       1469361600 1469966400
#> 599       1469966400 1470571200
#> 600       1470571200 1471176000
#> 601       1471176000 1471780800
#> 602       1471780800 1472385600
#> 603       1472385600 1472990400
#> 604       1472990400 1473595200
#> 605       1473595200 1474200000
#> 606       1474200000 1474804800
#> 607       1474804800 1475409600
#> 608       1475409600 1476014400
#> 609       1476014400 1476619200
#> 610       1476619200 1477224000
#> 611       1477224000 1477828800
#> 612       1477828800 1478433600
#> 613       1478433600 1479038400
#> 614       1479038400 1479643200
#> 615       1479643200 1480248000
#> 616       1480248000 1480852800
#> 617       1480852800 1481457600
#> 618       1481457600 1482062400
#> 619       1482062400 1482667200
#> 620       1482667200 1483272000
#> 621       1483272000 1483876800
#> 622       1483876800 1484481600
#> 623       1484481600 1485086400
#> 624       1485086400 1485691200
#> 625       1485691200 1486296000
#> 626       1486296000 1486900800
#> 627       1486900800 1487505600
#> 628       1487505600 1488110400
#> 629       1488110400 1488715200
#> 630       1488715200 1489320000
#> 631       1489320000 1489924800
#> 632       1489924800 1490529600
#> 633       1490529600 1491134400
#> 634       1491134400 1491739200
#> 635       1491739200 1492344000
#> 636       1492344000 1492948800
#> 637       1492948800 1493553600
#> 638       1493553600 1494158400
#> 639       1494158400 1494763200
#> 640       1494763200 1495368000
#> 641       1495368000 1495972800
#> 642       1495972800 1496577600
#> 643       1496577600 1497182400
#> 644       1497182400 1497787200
#> 645       1497787200 1498392000
#> 646       1498392000 1498996800
#> 647       1498996800 1499601600
#> 648       1499601600 1500206400
#> 649       1500206400 1500811200
#> 650       1500811200 1501416000
#> 651       1501416000 1502020800
#> 652       1502020800 1502625600
#> 653       1502625600 1503230400
#> 654       1503230400 1503835200
#> 655       1503835200 1504440000
#> 656       1504440000 1505044800
#> 657       1505044800 1505649600
#> 658       1505649600 1506254400
#> 659       1506254400 1506859200
#> 660       1506859200 1507464000
#> 661       1507464000 1508068800
#> 662       1508068800 1508673600
#> 663       1508673600 1509278400
#> 664       1509278400 1509883200
#> 665       1509883200 1510488000
#> 666       1510488000 1511092800
#> 667       1511092800 1511697600
#> 668       1511697600 1512302400
#> 669       1512302400 1512907200
#> 670       1512907200 1513512000
#> 671       1513512000 1514116800
#> 672       1514116800 1514721600
#> 673       1514721600 1515326400
#> 674       1515326400 1515931200
#> 675       1515931200 1516536000
#> 676       1516536000 1517140800
#> 677       1517140800 1517745600
#> 678       1517745600 1518350400
#> 679       1518350400 1518955200
#> 680       1518955200 1519560000
#> 681       1519560000 1520164800
#> 682       1520164800 1520769600
#> 683       1520769600 1521374400
#> 684       1521374400 1521979200
#> 685       1521979200 1522584000
#> 686       1522584000 1523188800
#> 687       1523188800 1523793600
#> 688       1523793600 1524398400
#> 689       1524398400 1525003200
#> 690       1525003200 1525608000
#> 691       1525608000 1526212800
#> 692       1526212800 1526817600
#> 693       1526817600 1527422400
#> 694       1527422400 1528027200
#> 695       1528027200 1528632000
#> 696       1528632000 1529236800
#> 697       1529236800 1529841600
#> 698       1529841600 1530446400
#> 699       1530446400 1531051200